### 2024-07-15 #6989 @editorialbot
[PRE REVIEW]: solids4foam-v2.1: A toolbox for performing solid mechanics and fluid-solid interaction simulations in OpenFOAM
**Submitting author:** @philipcardiff (<a href="http://orcid.org/0000-0002-4824-427X">Philip Cardiff</a>)
**Repository:** https://github.com/solids4foam/solids4foam
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1
