### 2024-07-22 73ecdeb
Solution for #122
