+++ src/solids4FoamModels/solidModels/linGeomTotalDispSnesSolid/linGeomTotalDispSnesSolid.C
@@ -1,0 +1,803 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+\*---------------------------------------------------------------------------*/
+
+#include "linGeomTotalDispSnesSolid.H"
+#include "fvm.H"
+#include "fvc.H"
+#include "fvMatrices.H"
+#include "addToRunTimeSelectionTable.H"
+#include "momentumStabilisation.H"
+#include "backwardDdtScheme.H"
+#include "sparseMatrixTools.H"
+#ifdef USE_PETSC
+    #include <petscksp.h>
+    #include <petscsnes.h>
+#endif
+
+// TESTING
+#include "linearElastic.H"
+#include "solidTractionFvPatchVectorField.H"
+
+// * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
+
+#ifdef USE_PETSC
+
+// User data "context" for PETSc functions
+typedef struct appCtx
+{
+    // Reference to the solid model object
+    Foam::solidModels::linGeomTotalDispSnesSolid& solMod_;
+
+    // Constructor
+    appCtx
+    (
+        Foam::solidModels::linGeomTotalDispSnesSolid& solMod
+    )
+    :
+        solMod_(solMod)
+    {}
+} appCtx;
+
+
+PetscErrorCode formResidualLinGeomTotalDispSnesSolid
+(
+    SNES snes,    // snes object
+    Vec x,        // current solution
+    Vec f,        // residual
+    void *ctx     // user context
+)
+{
+    const PetscScalar *xx;
+    PetscScalar       *ff;
+    appCtx *user = (appCtx *)ctx;
+
+    // Access x and f data
+    // VecRestoreArrayRead() and VecRestoreArray() must be called when access is
+    // no longer needed
+    CHKERRQ(VecGetArrayRead(x,&xx));
+    CHKERRQ(VecGetArray(f,&ff));
+
+    // Map the solution to an OpenFOAM field
+    const bool twoD = user->solMod_.twoD();
+    // Foam::volVectorField D("petsc_D", user->solMod_.D());
+    // Foam::volVectorField D("D", user->solMod_.D());
+    Foam::volVectorField& D = user->solMod_.D();
+    Foam::vectorField& DI = D;
+    // const Foam::boolList& ownedByThisProc =
+    //     user->solMod_.gPointIndices().ownedByThisProc();
+    if (Foam::Pstream::parRun())
+    {
+        // globalPoint should do what I need
+        Foam::FatalError
+            << "Fix parallel" << Foam::abort(Foam::FatalError);
+    }
+    {
+        int index = 0;
+        forAll(DI, i)
+        {
+            //if (ownedByThisProc[i])
+            {
+                DI[i].x() = xx[index++];
+                DI[i].y() = xx[index++];
+
+                if (!twoD)
+                {
+                    DI[i].z() = xx[index++];
+                }
+            }
+        }
+    }
+
+
[…]
