### 2024-07-26 0e65853
linGeomTotalDispSnesSolid: working in parallel
