+++ src/solids4FoamModels/solidModels/linGeomTotalDispSnesSolid/linGeomTotalDispSnesSolid.C
@@ -25,14 +25,13 @@ License
 #include "momentumStabilisation.H"
 #include "backwardDdtScheme.H"
 #include "sparseMatrixTools.H"
+#include "processorFvPatchField.H"
+#include "solidTractionFvPatchVectorField.H"
 #ifdef USE_PETSC
     #include <petscksp.h>
     #include <petscsnes.h>
 #endif
 
-// TESTING
-#include "linearElastic.H"
-#include "solidTractionFvPatchVectorField.H"
 
 // * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
 
@@ -63,60 +62,49 @@ PetscErrorCode formResidualLinGeomTotalDispSnesSolid
     void *ctx     // user context
 )
 {
+    // Foam::Pout<< __LINE__ << Foam::endl;
     const PetscScalar *xx;
     PetscScalar       *ff;
     appCtx *user = (appCtx *)ctx;
 
     // Access x and f data
-    // VecRestoreArrayRead() and VecRestoreArray() must be called when access is
-    // no longer needed
-    CHKERRQ(VecGetArrayRead(x,&xx));
-    CHKERRQ(VecGetArray(f,&ff));
+    CHKERRQ(VecGetArrayRead(x, &xx));
+    CHKERRQ(VecGetArray(f, &ff));
 
     // Map the solution to an OpenFOAM field
     const bool twoD = user->solMod_.twoD();
-    // Foam::volVectorField D("petsc_D", user->solMod_.D());
-    // Foam::volVectorField D("D", user->solMod_.D());
+    const int blockSize = twoD ? 2 : 3;
     Foam::volVectorField& D = user->solMod_.D();
     Foam::vectorField& DI = D;
-    // const Foam::boolList& ownedByThisProc =
-    //     user->solMod_.gPointIndices().ownedByThisProc();
-    if (Foam::Pstream::parRun())
-    {
-        // globalPoint should do what I need
-        Foam::FatalError
-            << "Fix parallel" << Foam::abort(Foam::FatalError);
-    }
+
     {
         int index = 0;
-        forAll(DI, i)
+        forAll(DI, localCellI)
         {
-            //if (ownedByThisProc[i])
-            {
-                DI[i].x() = xx[index++];
-                DI[i].y() = xx[index++];
+            DI[localCellI][Foam::vector::X] = xx[index++];
+            DI[localCellI][Foam::vector::Y] = xx[index++];
 
-                if (!twoD)
-                {
-                    DI[i].z() = xx[index++];
-                }
+            if (!twoD)
+            {
+                DI[localCellI][Foam::vector::Z] = xx[index++];
             }
         }
     }
 
+    // Restore the solution vector
+    CHKERRQ(VecRestoreArrayRead(x, &xx));
+
+    // Enforce the D boundary conditions and sync processor boundaries
+    D.correctBoundaryConditions();
 
     // Compute the residual
     const Foam::vectorField res(user->solMod_.residualMomentum(D));
 
     // Map the data to f
-    // const Foam::labelList& localToGlobalPointMap =
-    //     user->solMod_.gPointIndices().localToGlobalPointMap();
-    const int blockSize = twoD ? 2 : 3;
     forAll(res, localBlockRowI)
     {
         const Foam::vector& resI = res[localBlockRowI];
-        //const int blockRowI = localToGlobalPointMap[localBlockRowI];
-        const int blockRowI = localBlockRowI; // serial only
+        const int blockRowI = localBlockRowI;
 
         ff[blockRowI*blockSize] = resI.x();
         ff[blockRowI*blockSize + 1] = resI.y();
@@ -127,9 +115,8 @@ PetscErrorCode formResidualLinGeomTotalDispSnesSolid
         }
     }
 
-    // Restore vectors
[…]
