### 2024-07-27 61db85a
Merged linGeomTotalDispSnesSolid into linGeomTotalDispSolid
