+++ src/solids4FoamModels/solidModels/linGeomTotalDispSnesSolid/linGeomTotalDispSnesSolid.C
@@ -1,880 +1,0 @@
-/*---------------------------------------------------------------------------*\
-License
-    This file is part of solids4foam.
-
-    solids4foam is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your
-    option) any later version.
-
-    solids4foam is distributed in the hope that it will be useful, but
-    WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-
-\*---------------------------------------------------------------------------*/
-
-#include "linGeomTotalDispSnesSolid.H"
-#include "fvm.H"
-#include "fvc.H"
-#include "fvMatrices.H"
-#include "addToRunTimeSelectionTable.H"
-#include "momentumStabilisation.H"
-#include "backwardDdtScheme.H"
-#include "sparseMatrixTools.H"
-#include "processorFvPatchField.H"
-#include "solidTractionFvPatchVectorField.H"
-#ifdef USE_PETSC
-    #include <petscksp.h>
-    #include <petscsnes.h>
-#endif
-
-
-// * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
-
-#ifdef USE_PETSC
-
-// User data "context" for PETSc functions
-typedef struct appCtx
-{
-    // Reference to the solid model object
-    Foam::solidModels::linGeomTotalDispSnesSolid& solMod_;
-
-    // Constructor
-    appCtx
-    (
-        Foam::solidModels::linGeomTotalDispSnesSolid& solMod
-    )
-    :
-        solMod_(solMod)
-    {}
-} appCtx;
-
-
-PetscErrorCode formResidualLinGeomTotalDispSnesSolid
-(
-    SNES snes,    // snes object
-    Vec x,        // current solution
-    Vec f,        // residual
-    void *ctx     // user context
-)
-{
-    // Foam::Pout<< __LINE__ << Foam::endl;
-    const PetscScalar *xx;
-    PetscScalar       *ff;
-    appCtx *user = (appCtx *)ctx;
-
-    // Access x and f data
-    CHKERRQ(VecGetArrayRead(x, &xx));
-    CHKERRQ(VecGetArray(f, &ff));
-
-    // Map the solution to an OpenFOAM field
-    const bool twoD = user->solMod_.twoD();
-    const int blockSize = twoD ? 2 : 3;
-    Foam::volVectorField& D = user->solMod_.D();
-    Foam::vectorField& DI = D;
-
-    {
-        int index = 0;
-        forAll(DI, localCellI)
-        {
-            DI[localCellI][Foam::vector::X] = xx[index++];
-            DI[localCellI][Foam::vector::Y] = xx[index++];
-
-            if (!twoD)
-            {
-                DI[localCellI][Foam::vector::Z] = xx[index++];
-            }
-        }
-    }
-
-    // Restore the solution vector
-    CHKERRQ(VecRestoreArrayRead(x, &xx));
-
-    // Enforce the D boundary conditions and sync processor boundaries
-    D.correctBoundaryConditions();
-
-    // Compute the residual
-    const Foam::vectorField res(user->solMod_.residualMomentum(D));
-
-    // Map the data to f
-    forAll(res, localBlockRowI)
-    {
-        const Foam::vector& resI = res[localBlockRowI];
-        const int blockRowI = localBlockRowI;
-
[…]
