### 2024-07-28 aca8d0d
linGeomTotalDispSolid: add explicit solution approach (merged explicitLinGeomTotalDispSolid)
