+++ src/solids4FoamModels/solidModels/explicitLinGeomTotalDispSolid/explicitLinGeomTotalDispSolid.H
@@ -1,202 +1,0 @@
-/*---------------------------------------------------------------------------*\
-License
-    This file is part of solids4foam.
-
-    solids4foam is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your
-    option) any later version.
-
-    solids4foam is distributed in the hope that it will be useful, but
-    WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-
-Class
-    explicitLinGeomTotalDispSolid
-
-Description
-    Version of the linGeomTotalDispSolid solid model, where a fully explicit
-    solution algorithm is used.
-
-    The explicit implementation uses central differencing, where a linear bulk
-    viscosity pressure term dissipates high frequencies.
-
-    Details of the central differencing solution algorithm can be found in, e.g.
-    - Belytschko, Liu, Moran, 2000, Box 6.1
-    - Ding and Ye, Laser shock peening, Performance and process simulation,
-      2006, Section 3.5.2
-
-    The linear bulk viscosity term is described in the Abaqus documentation.
-
-    A Jameson-Schmidt-Turkel (JST) 4th order diffusion term is used for
-    stabilisation.
-
-Author
-    Philip Cardiff, UCD.  All rights reserved.
-
-SourceFiles
-    explicitLinGeomTotalDispSolid.C
-
-\*---------------------------------------------------------------------------*/
-
-#ifndef explicitLinGeomTotalDispSolid_H
-#define explicitLinGeomTotalDispSolid_H
-
-#include "solidModel.H"
-#include "volFields.H"
-#include "surfaceFields.H"
-#include "pointFields.H"
-#include "uniformDimensionedFields.H"
-#include "mechanicalEnergies.H"
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace solidModels
-{
-
-/*---------------------------------------------------------------------------*\
-                        Class explicitLinGeomTotalDispSolid Declaration
-\*---------------------------------------------------------------------------*/
-
-class explicitLinGeomTotalDispSolid
-:
-    public solidModel
-{
-    // Private data
-
-        //- Implicit stiffness; coefficient of the Laplacian term
-        //  The value of this term only affects convergence and not the answer
-        //  In the case of Hookean elasticity, (2*mu + lambda) is the optimum
-        const volScalarField impK_;
-
-        //- Implicit stiffness surface field
-        surfaceScalarField impKf_;
-
-        //- Reciprocal of impK
-        //  We divide by impK every time tractionBoundarySnGrad is called so we
-        //  will store the reciprocal for effficiency
-        const volScalarField rImpK_;
-
-        //- Scale factor for Rhie-Chow smoothing term
-        //const scalar RhieChowScaleFactor_;
-
-        // Scale factor for Jameson-Schmidt-Turkel smoothing term
-        const scalar JSTScaleFactor_;
-
-        //- Wave speed field
-        surfaceScalarField waveSpeed_;
-
-        //- Energy calculation
-        mechanicalEnergies energies_;
-
-        //- Acceleration
-        volVectorField a_;
-
-    // Private Member Functions
-
-        //- Update the stress field
-        void updateStress();
-
-        //- Smooth the hydrostatic pressure field
[…]
