+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -24,10 +24,8 @@ License
 #include "addToRunTimeSelectionTable.H"
 #include "momentumStabilisation.H"
 #include "solidTractionFvPatchVectorField.H"
-// #ifdef USE_PETSC
-//     #include <petscksp.h>
-//     #include <petscsnes.h>
-// #endif
+#include "fixedDisplacementZeroShearFvPatchVectorField.H"
+#include "symmetryFvPatchFields.H"
 
 
 // * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
@@ -177,46 +175,27 @@ PetscErrorCode formJacobianLinGeomTotalDispSolid
     }
     else
     {
-        Foam::Info<< "Initialising the matrix" << Foam::endl;
-        // Foam::labelList nonZerosPerBlockRow(user->solMod_.mesh().nCells(), 1);
-        // forAll(user->solMod_.mesh().owner(), faceI)
-        // {
-        //     nonZerosPerBlockRow[user->solMod_.mesh().owner()[faceI]]++;
-        //     nonZerosPerBlockRow[user->solMod_.mesh().neighbour()[faceI]]++;
-        // }
-        // const int d_nz = blockSize*Foam::max(nonZerosPerBlockRow);
-        // Foam::Info<< "    d_nz = " << d_nz << Foam::endl;
-
-        // MatSetOption(B, MAT_NEW_NONZERO_ALLOCATION_ERR, PETSC_FALSE);
-
-        // // Allocate parallel matrix
-        // // To-do: count exact number of non-zeros rather than conservatively
-        // // estimating
-        // //CHKERRQ(MatMPIAIJSetPreallocation(A, 0, d_nnz, 0, o_nnz));
-        // // Allocate parallel matrix with the same conservative stencil per node
-        // CHKERRQ(MatMPIAIJSetPreallocation(B, d_nz, NULL, 0, NULL));
+        Foam::Info<< "    Initialising the matrix" << Foam::endl;
 
         // Set the block size
         CHKERRQ(MatSetBlockSize(B, blockSize));
 
+        // Number of vector unknowns on this processor
         const int blockn = user->solMod_.globalCells().localSize();
-        //const int n = blockSize*blockn;
-        // const int blockN = user->solMod_.globalCells().size();
-        // const int N = blockSize*blockN;
+
+        // Count the number of non-zeros in the matrix
 
         // Number of on-processor non-zeros per row
-        // int* d_nnz = (int*)malloc(n*sizeof(int));
-        int* D_nnz = (int*)malloc(blockn*sizeof(int));
+        int* d_nnz = (int*)malloc(blockn*sizeof(int));
 
         // Number of off-processor non-zeros per row
-        // int* o_nnz = (int*)malloc(n*sizeof(int));
-        int* O_nnz = (int*)malloc(blockn*sizeof(int));
+        int* o_nnz = (int*)malloc(blockn*sizeof(int));
 
-        // Initialise D_nnz and O_nnz to zero
+        // Initialise d_nnz and o_nnz to zero
         for (int i = 0; i < blockn; ++i)
         {
-            D_nnz[i] = 1; // count diagonal cell
-            O_nnz[i] = 0;
+            d_nnz[i] = 1; // count diagonal cell
+            o_nnz[i] = 0;
         }
 
         // Count neighbours sharing an internal face
@@ -226,8 +205,8 @@ PetscErrorCode formJacobianLinGeomTotalDispSolid
         {
             const Foam::label ownCellID = own[faceI];
             const Foam::label neiCellID = nei[faceI];
-            D_nnz[ownCellID]++;
-            D_nnz[neiCellID]++;
+            d_nnz[ownCellID]++;
[…]
