### 2024-07-28 5a28f77
linGeomTotalDispSolid PETSc SNES: fix for parallel
