### 2024-07-29 b8ed9c0
Add runSolidModel script to solids4FoamScripts.sh: wip
