+++ applications/scripts/solids4FoamScripts.sh
@@ -645,3 +645,109 @@ function solids4Foam::runParallel()
         fi
     fi
 }
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
+# runSolidModel
+#     Run the case using the specified solid model. This script modifies the
+#     case to work with the specified solid model and then runs the case.
+#     Modification are made to the solidProperties, fvSchemes, fvSolution,
+#     boundary conditions, etc., as needed. 
+# Arguments:
+#     1: Address of the case directory
+#     2: Name of the solid model to use
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
+function solids4Foam::runSolidModel()
+{
+    echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
+    echo "| solids4Foam::runSolidModel start                                   |"
+    echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
+    echo
+
+    # Check number of input parameters is correct
+    if [[ $# -ne 2 ]]
+    then
+        solids4Foam::err "runSolidModel: incorrect number of parameters"
+    fi
+
+    # Give sensible names to the argument
+    CASE_DIR=$1
+    SOLID_MODEL=$2
+
+    # Directory containing the template solid model dicts
+    # Note: the Allwmake script replaces /Users/philipc/OpenFOAM/philipc-v2012/solids4foam/applications/scripts with $(pwd)
+    DICTS_PARENT_DIR=/Users/philipc/OpenFOAM/philipc-v2012/solids4foam/applications/scripts
+    DICTS_DIR=${DICTS_PARENT_DIR}/solidModelDicts/${SOLID_MODEL}
+
+    # Check default dictionaries exist for the specified solid model
+    if [ ! -d "${DICTS_DIR}" ]; then
+        solids4Foam::err "Cannot find ${DICTS_DIR}. Do default dicts exist for this solid model?"
+    fi
+    
+    # Copy default dictionaries for the specified solid model
+
+    echo "Replacing solidProperties"
+    \cp ${DICTS_DIR}/solidProperties ${CASE_DIR}/constant/
+
+    echo "Replacing fvSolution"
+    \cp ${DICTS_DIR}/fvSolution ${CASE_DIR}/system/
+    
+    echo "Replacing fvSchemes"
+    \cp ${DICTS_DIR}/fvSchemes ${CASE_DIR}/system/
+
+    # Add files to case, if the addFilesToCase directory is present
+    if [ -d "${DICTS_DIR}/addFilesToCase" ]; then
+        for f in "${DICTS_DIR}/addFilesToCase/"*; do
+            if [ -f "$f" ]; then  # Ensure it's a file
+                echo "Copying ${f} to ${CASE_DIR}"
+                \cp "${f}" "${CASE_DIR}"
+            fi
+        done
+    fi
+
+    # Check the displacement file is correct
+    if [ ! -f "${DICTS_DIR}/displacementName" ]; then
[…]
