### 2024-07-30 83291e1
Create foamPetscSnesHelper to encapsulate PETSc code and objects and make reuse easier
