+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelper.H
@@ -1,0 +1,229 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+Class
+    foamPetscSnesHelper
+
+Description
+    Helper class and functions for interacting with the PETSc nonlinear
+    solver (SNES).
+
+    Solid models can use this SNES interface by deriving from this class and
+    defining two virtual functions:
+
+    1. residualMomentum
+       This should return the momentum equation imbalance (residual) for
+       the given displacement field
+    2. JacobianMomentum
+       This should return the Jacobian (or an approximation of it) of the
+       residual with respect to the displacement field.
+       A Newton-Krylov method can work well with a compact stencil
+       approximation of the Jacobian, e.g. fvm::laplacian.
+
+Author
+    Philip Cardiff, UCD.  All rights reserved.
+
+SourceFiles
+    foamPetscSnesHelper.C
+
+\*---------------------------------------------------------------------------*/
+
+#ifndef foamPetscSnesHelper_H
+#define foamPetscSnesHelper_H
+
+#ifdef USE_PETSC
+
+#include <petscsnes.h>
+#include <petscksp.h>
+
+#include "globalIndex.H"
+#include "volFields.H"
+#include "sparseMatrix.H"
+
+
+// * * * * * * * * * * * * * * External Structs  * * * * * * * * * * * * * * //
+
+// Forward declaration of the Foam::foamPetscSnesHelper class
+namespace Foam
+{
+    class foamPetscSnesHelper;
+}
+
+// User data "context" for PETSc functions
+typedef struct appCtxfoamPetscSnesHelper
+{
+    // Reference to the solid model object
+    Foam::foamPetscSnesHelper& solMod_;
+
+    // Constructor
+    appCtxfoamPetscSnesHelper
+    (
+        Foam::foamPetscSnesHelper& solMod
+    )
+    :
+        solMod_(solMod)
+    {}
+} appCtxfoamPetscSnesHelper;
+
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+namespace Foam
+{
+
+/*---------------------------------------------------------------------------*\
+                     Class foamPetscSnesHelper Declaration
+\*---------------------------------------------------------------------------*/
+
+class foamPetscSnesHelper
+{
+    // Private data
+
+        //- Reference to the solution field
+        volVectorField& solution_;
+
+        //- Helper object for converting between local (on processor) and global
+        //  cell indices
+        globalIndex globalCells_;
+
+        //- Is this a 2-D case? If so, it is assumed that Z is the empty
+        //  direction
+        const Switch twoD_;
+
+        //- Flag to set whether the solver stops when PETSs SNES returns an
+        //  error
+        const Switch stopOnPetscError_;
+
+        //- Pointer to the PETSc SNES object
+        autoPtr<SNES> snesPtr_;
+
+        //- Pointer to the PETSc solution vector
+        autoPtr<Vec> xPtr_;
+
+        //- Pointer to the PETSc matrix
[…]
