+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.H
@@ -43,42 +43,7 @@ SourceFiles
 #include "uniformDimensionedFields.H"
 #include "sparseMatrix.H"
 #include "globalIndex.H"
-#ifdef USE_PETSC
-    #include <petscksp.h>
-    #include <petscsnes.h>
-#endif
-
-
-// * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
-
-#ifdef USE_PETSC
-
-// Forward declaration of the Foam::solidModels::linGeomTotalDispSolid class
-namespace Foam
-{
-    namespace solidModels
-    {
-        class linGeomTotalDispSolid;
-    }
-}
-
-// User data "context" for PETSc functions
-typedef struct appCtx
-{
-    // Reference to the solid model object
-    Foam::solidModels::linGeomTotalDispSolid& solMod_;
-
-    // Constructor
-    appCtx
-    (
-        Foam::solidModels::linGeomTotalDispSolid& solMod
-    )
-    :
-        solMod_(solMod)
-    {}
-} appCtx;
-
-#endif // ifdef USE_PETSC
+#include "foamPetscSnesHelper.H"
 
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
@@ -97,7 +62,8 @@ namespace solidModels
 
 class linGeomTotalDispSolid
 :
-    public solidModel
+    public solidModel,
+    public foamPetscSnesHelper
 {
     // Private data
 
@@ -120,26 +86,6 @@ class linGeomTotalDispSolid
 
         //- Predict new time-step fields using the velocity field
         const Switch predictor_;
-
-        //- Helper object for converting between local (on processor) and global
-        //  cell indices
-        globalIndex globalCells_;
-
-        //- Flag to set whether the solver stops when PETSs SNES returns an
-        //  error
-        const Switch stopOnPetscError_;
-
-        //- Pointer to the PETSc SNES object
-        autoPtr<SNES> snesPtr_;
-
-        //- Pointer to the PETSc solution vector
-        autoPtr<Vec> xPtr_;
-
-        //- Pointer to the PETSc matrix
-        autoPtr<Mat> APtr_;
-
-        //- PETSc user context
-        autoPtr<appCtx> snesUserPtr_;
 
 
     // Private Member Functions
@@ -218,12 +164,6 @@ public:
                 return D();
             }
 
-            //- Const access to the global index object
-            const globalIndex& globalCells() const
-            {
-                return globalCells_;
-            }
-
 
         // Edit
 
@@ -236,7 +176,7 @@ public:
             //- Calculate the momentum equation residual
             //  This should be const but is not because of the mechanical law
             //  The can be fixed when the mechanical models are refactored
-            tmp<vectorField> residualMomentum
+            virtual tmp<vectorField> residualMomentum
             (
                 const volVectorField& D
             );
@@ -249,7 +189,7 @@ public:
             //  This can also be an approximation of the true Jacobian, e.g. a
             //  compact molecule Laplacian
             //  We initially adopt this approach for testin
-            tmp<sparseMatrix> JacobianMomentum
+            virtual tmp<sparseMatrix> JacobianMomentum
             (
                 const volVectorField& D
             );
@@ -261,9 +201,6 @@ public:
                 const scalarField& pressure,
                 const fvPatch& patch
             ) const;
-
-            //- This function is called at the end of the simulation
-            virtual void end();
 };
 
 

[…]
