### 2024-07-30 cfbb875
nonLinGeomTotalLagTotalDispSolid: work in progress on SNES implementation
