+++ src/solids4FoamModels/solidModels/nonLinGeomTotalLagTotalDispSolid/nonLinGeomTotalLagTotalDispSolid.C
@@ -22,6 +22,9 @@ License
 #include "fvc.H"
 #include "fvMatrices.H"
 #include "addToRunTimeSelectionTable.H"
+#include "solidTractionFvPatchVectorField.H"
+#include "fixedDisplacementZeroShearFvPatchVectorField.H"
+#include "symmetryFvPatchFields.H"
 
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
@@ -72,107 +75,62 @@ void nonLinGeomTotalLagTotalDispSolid::predict()
 }
 
 
-// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
-
-nonLinGeomTotalLagTotalDispSolid::nonLinGeomTotalLagTotalDispSolid
+void nonLinGeomTotalLagTotalDispSolid::enforceTractionBoundaries
 (
-    Time& runTime,
-    const word& region
-)
-:
-    solidModel(typeName, runTime, region),
-    F_
-    (
-        IOobject
-        (
-            "F",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::AUTO_WRITE
-        ),
-        mesh(),
-        dimensionedTensor("I", dimless, I)
-    ),
-    Finv_
-    (
-        IOobject
-        (
-            "Finv",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::NO_WRITE
-        ),
-        inv(F_)
-    ),
-    J_
-    (
-        IOobject
-        (
-            "J",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::NO_WRITE
-        ),
-        det(F_)
-    ),
-    impK_(mechanical().impK()),
-    impKf_(mechanical().impKf()),
-    rImpK_(1.0/impK_),
-    predictor_(solidModelDict().lookupOrDefault<Switch>("predictor", false))
+    surfaceVectorField& traction,
+    const volVectorField& D,
+    const surfaceVectorField& n
+) const
 {
-    DisRequired();
-
-    // Force all required old-time fields to be created
-    fvm::d2dt2(D());
-
-    if (predictor_)
+    // Enforce traction conditions
+    forAll(D.boundaryField(), patchI)
     {
-        // Check ddt scheme for D is not steadyState
-        const word ddtDScheme
+        if
         (
-#ifdef OPENFOAM_NOT_EXTEND
-            mesh().ddtScheme("ddt(" + D().name() +')')
-#else
-            mesh().schemesDict().ddtScheme("ddt(" + D().name() +')')
-#endif
-        );
-
-        if (ddtDScheme == "steadyState")
+            isA<solidTractionFvPatchVectorField>
+            (
+                D.boundaryField()[patchI]
+            )
+        )
         {
-            FatalErrorIn(type() + "::" + type())
-                << "If predictor is turned on, then the ddt(" << D().name()
-                << ") scheme should not be 'steadyState'!" << abort(FatalError);
+            const solidTractionFvPatchVectorField& tracPatch =
+                refCast<const solidTractionFvPatchVectorField>
+                (
+                    D.boundaryField()[patchI]
+                );
+
+            const vectorField& nPatch = n.boundaryField()[patchI];
+
+            traction.boundaryFieldRef()[patchI] =
+                tracPatch.traction() - nPatch*tracPatch.pressure();
         }
-    }
+        else if
+        (
+            isA<fixedDisplacementZeroShearFvPatchVectorField>
+            (
+                D.boundaryField()[patchI]
+            )
+         || isA<symmetryFvPatchVectorField>
+            (
+                D.boundaryField()[patchI]
[…]
