### 2024-07-31 862817c
Fix for when not using SNES in linGeom and nonLinGeomTL
