+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -373,10 +373,17 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
     solidModel(typeName, runTime, region),
     foamPetscSnesHelper
     (
-        fileName(solidModelDict().lookup("optionsFile")),
+        fileName
+        (
+            solidModelDict().lookupOrDefault<fileName>
+            (
+                "optionsFile", "petscOptions"
+            )
+        ),
         D(),
         solidModel::twoD(),
-        solidModelDict().lookupOrDefault<Switch>("stopOnPetscError", true)
+        solidModelDict().lookupOrDefault<Switch>("stopOnPetscError", true),
+        bool(solutionAlg() == solutionAlgorithm::PETSC_SNES)
     ),
     impK_(mechanical().impK()),
     impKf_(mechanical().impKf()),
