+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelper.C
@@ -410,7 +410,7 @@ foamPetscSnesHelper::~foamPetscSnesHelper()
 
         PetscFinalize();
     }
-}    
+}
 
 // * * * * * * * * * * * * * * * Member Functions  * * * * * * * * * * * * * //
 
