### 2024-08-03 f12e428
Update for setFibreField
