+++ applications/utilities/setFibreField/setFibreField.C
@@ -30,15 +30,17 @@ Author
 \*---------------------------------------------------------------------------*/
 
 #include "fvCFD.H"
-
+#include "transform.H"
+#include "unitConversion.H"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 int main(int argc, char *argv[])
 {
+    #include "addRegionOption.H"
     #include "setRootCase.H"
     #include "createTime.H"
-    #include "createMesh.H"
+    #include "createNamedMesh.H"
 
     // Read the transmural distance field
     // This will be calculated using the given boundary conditions
@@ -65,167 +67,124 @@ int main(int argc, char *argv[])
     Info<< "Writing t" << endl;
     t.write();
 
-    const dimensionedScalar dimL("1", dimLength, 1.0);
+    // Base direction as per Rossi-Lassila et al. approach
+    const vector k(0, 0, 1);
 
-    const volScalarField rs("rs", 0.001*(7.0*dimL + 3.0*t));
-    Info<< "Writing rs" << endl;
-    rs.write();
+    // Transmural direction
+    volVectorField et("et", fvc::grad(t));
+    et /= mag(et);
 
-    const volScalarField rl("rl", 0.001*(17.0*dimL + 3.0*t));
-    Info<< "Writing rl" << endl;
-    rl.write();
+    // Normal direction
+    volVectorField en("en", k - (k & et)*et);
+    en /= mag(en);
 
-    const volScalarField alpha("alpha", 90.0*dimL - 180.0*t);
-    Info<< "Writing alpha" << endl;
-    alpha.write();
+    // Longitudinal direction
+    volVectorField el("el", en ^ et);
+    el /= mag(el);
 
-    // Convert alpha to radians
+    Info<< "Writing et, en, el" << endl;
+    et.write();
+    en.write();
+    el.write();
+
+
+    // alpha at the endocardium: SHOULD BE INPUT PARAMETER
+    const scalar alphaEndo = degToRad(60.0);
+
+    // alpha at the epicardium: SHOULD BE INPUT PARAMETER
+    const scalar alphaEpi = degToRad(-60.0);
+
+    // Calculate alphaRadians
     const volScalarField alphaRadians
     (
-        "alphaRadians", alpha*constant::mathematical::pi/180.0
+        "alphaRadians",
+        alphaEndo*(1.0 - t) + alphaEpi*t
     );
     Info<< "Writing alphaRadians" << endl;
     alphaRadians.write();
 
-    // Unit Cartesian vectors
-    const vector iHat(1, 0, 0);
-    const vector jHat(0, 1, 0);
-    const vector kHat(0, 0, 1);
+    // Rotate el about et by alphaRadians to get f
+    volVectorField f0("f0", el);
+    forAll(et, cellI)
+    {
+        const tensor rotT = Ra(et[cellI], alphaRadians[cellI]);
+        f0[cellI] = transform(-rotT, f0[cellI]);
+    }
+    forAll(et.boundaryField(), patchI)
+    {
+        forAll(et.boundaryField()[patchI], faceI)
+        {
+            const tensor rotT = Ra
+            (
+                et.boundaryField()[patchI][faceI],
+                alphaRadians.boundaryField()[patchI][faceI]
+            );
+            f0.boundaryFieldRef()[patchI][faceI] = transform
+            (
+                -rotT,
+                f0.boundaryField()[patchI][faceI]
+            );
+        }
+    }
 
-    // Coordinate fields
[…]
