### 2024-08-21 9444da6
Add sphericalCavity analytical solution functions and case
