+++ src/solids4FoamModels/functionObjects/sphericalCavityAnalyticalSolution/sphericalCavityAnalyticalSolution.C
@@ -1,0 +1,427 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+\*----------------------------------------------------------------------------*/
+
+#include "sphericalCavityAnalyticalSolution.H"
+#include "addToRunTimeSelectionTable.H"
+#include "volFields.H"
+#include "pointFields.H"
+#include "sphericalCavityStressDisplacement.H"
+
+// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
+
+namespace Foam
+{
+    defineTypeNameAndDebug(sphericalCavityAnalyticalSolution, 0);
+
+    addToRunTimeSelectionTable
+    (
+        functionObject,
+        sphericalCavityAnalyticalSolution,
+        dictionary
+    );
+}
+
+
+// * * * * * * * * * * * * * Private Member Functions  * * * * * * * * * * * //
+
+
+bool Foam::sphericalCavityAnalyticalSolution::writeData()
+{
+    // Lookup the solid mesh
+    const fvMesh* meshPtr = NULL;
+    if (time_.foundObject<fvMesh>("solid"))
+    {
+        meshPtr = &(time_.lookupObject<fvMesh>("solid"));
+    }
+    else
+    {
+        meshPtr = &(time_.lookupObject<fvMesh>("region0"));
+    }
+    const fvMesh& mesh = *meshPtr;
+
+    // Lookup the point mesh
+    const pointMesh& pMesh = mesh.lookupObject<pointMesh>("pointMesh");
+
+    // Cell-centres coordinates
+    const volVectorField& C = mesh.C();
+    const vectorField& CI = C;
+
+    // Point coordinates
+    const pointField& points = mesh.points();
+
+    if (gMin(mag(points)) < -SMALL)
+    {
+        FatalErrorIn
+        (
+            "bool Foam::sphericalCavityAnalyticalSolution::writeData()"
+        )   << "The cavity should be centred on the origin!" << endl
+            << "gMin(mag(points)) = " << gMin(mag(points))
+            << abort(FatalError);
+    }
+
+    // Cell analytical fields
+    if (cellDisplacement_ || cellStress_)
+    {
+        // Analytical stress field
+        volSymmTensorField analyticalStress
+        (
+            IOobject
+            (
+                "analyticalCellStress",
+                time_.timeName(),
+                mesh,
+                IOobject::NO_READ,
+                IOobject::AUTO_WRITE
+            ),
+            mesh,
+            dimensionedSymmTensor("zero", dimPressure, symmTensor::zero),
+            "calculated"
+        );
+
+        // Analytical displacement field
+        volVectorField analyticalD
+        (
+            IOobject
+            (
+                "analyticalD",
+                time_.timeName(),
+                mesh,
+                IOobject::NO_READ,
+                IOobject::AUTO_WRITE
+            ),
+            mesh,
+            dimensionedVector("zero", dimLength, vector::zero),
+            "calculated"
+        );
+
+        symmTensorField& sI = analyticalStress;
+        vectorField& aDI = analyticalD;
+
+        forAll(sI, cellI)
+        {
+            if (cellStress_)
+            {
[…]
