### 2024-08-22 e912ded
sphericalCavityAnalyticalSolution: write stress errors to history file
