### 2024-08-26 a4b312e
leastSquaresS4fGrad: add useNeumannBoundaryFaceValues option and consistent symmetry plane treatment
