### 2024-08-27 4ddd8fa
Remove sphericalCavityAnalyticalSolution backup files
