+++ src/solids4FoamModels/functionObjects/sphericalCavityAnalyticalSolution/sphericalCavityAnalyticalSolution.working.C
@@ -1,576 +1,0 @@
-/*---------------------------------------------------------------------------*\
-License
-    This file is part of solids4foam.
-
-    solids4foam is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your
-    option) any later version.
-
-    solids4foam is distributed in the hope that it will be useful, but
-    WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-
-\*----------------------------------------------------------------------------*/
-
-#include "sphericalCavityAnalyticalSolution.H"
-#include "addToRunTimeSelectionTable.H"
-#include "volFields.H"
-#include "pointFields.H"
-#include "coordinateSystem.H"
-#include "cylindricalCS.H"
-
-// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
-
-namespace Foam
-{
-    defineTypeNameAndDebug(sphericalCavityAnalyticalSolution, 0);
-
-    addToRunTimeSelectionTable
-    (
-        functionObject,
-        sphericalCavityAnalyticalSolution,
-        dictionary
-    );
-}
-
-
-// * * * * * * * * * * * * * Private Member Functions  * * * * * * * * * * * //
-
-Foam::scalar Foam::sphericalCavityAnalyticalSolution::KroneckerDelta
-(
-    const int i, const int j
-) const
-{
-    return (i == j) ? 1.0 : 0.0;
-}
-
-Foam::scalar Foam::sphericalCavityAnalyticalSolution::calculateDispComponent
-(
-    const int i,
-    const scalar nu,
-    const scalar sigma_0,
-    const scalar E,
-    const scalar a,
-    const vector& x
-) const
-{
-    const scalar R = mag(x);
-    const scalar x_3 = x[vector::Z];
-
-    const scalar term1 =
-        2
-      + (5*(5 - 4*nu)*pow(a, 3))/((7 - 5*nu)*pow(R, 3))
-      + (6*pow(a, 5))/((7 - 5*nu)*pow(R, 5));
-    const scalar term2 =
-        (-2*nu)/(1 + nu)
-      + (5*nu - 6)*pow(a, 3)/((7 - 5*nu)*pow(R, 3))
-      + 3*pow(a, 5)/((7 - 5*nu)*pow(R, 5));
-    const scalar common_factor = (1 + nu)*sigma_0/(2*E);
-
-    return
-        common_factor
-       *(
-           term1*x_3*KroneckerDelta(i, 2)
-         + term2*x[i]*(1 - 5*pow(x_3, 2)/pow(R, 2))
-       );
-}
-
-
-Foam::symmTensor
-Foam::sphericalCavityAnalyticalSolution::calculateSigma
-(
-    const scalar T,
-    const scalar nu,
-    const scalar a,
-    const vector& x
-) const
-{
-    // Stress formulae from Southwell et al, 1926, On the concentration of
-    // stress in the neighbourhood of a small spherical flaw; and on the
-    // propagation of fatigue fractures in Statistically Isotropic materials
-
-    // Polar radial coordinate
-    // Take care: in general, R != r
-    const scalar R = mag(x);
-
-    // Cylindrical radial coordinate
-    const scalar r = sqrt(sqr(x.x()) + sqr(x.y()));
-
-    // Cylindrical axial coordinate
-    const scalar z = x.z();
-
-    // Initialise stress tensor
[…]
