### 2024-08-27 5a244fe
sphericalCavityAnalyticalSolution: displacement solution is correct
