### 2024-08-28 9ed757d
leastSquaresS4fGrad: comment fixesValue as vectors need to be updated
