### 2024-09-03 4ef4184
Add fvOptions to linGeomTotalDispSolid to allow the method of manufactured solutions
