+++ src/solids4FoamModels/materialModels/mechanicalModel/mechanicalLaws/nonLinearGeometryLaws/YeohElastic/YeohElastic.C
@@ -90,13 +90,13 @@ Foam::YeohElastic::YeohElastic
 {
     // Compute material properties
     // Shear modulus: based on pure shear stress state for hyperelasticity
-    mu() = 2.0*c1_;
+    this->mu(2.0*c1_);
 
     // Bulk modulus
     // The user can specify K directly or the Poisson's ratio, nu
     if (dict.found("K"))
     {
-        K() = dimensionedScalar(dict.lookup("K"));
+        this->K(dimensionedScalar(dict.lookup("K")));
     }
     else if (dict.found("nu") && !dict.found("K"))
     {
@@ -106,7 +106,7 @@ Foam::YeohElastic::YeohElastic
         const volScalarField E(6.0*c1_);
 
         // Compute K based on linear elasticity
-        K() = E/(3.0*(1.0 - 2.0*nu));
+        this->K(E/(3.0*(1.0 - 2.0*nu)));
     }
     else
     {
@@ -125,8 +125,8 @@ Foam::YeohElastic::YeohElastic
         << "    max(K) = "   << gMax(mag(K())())  << endl;
 
     // Update surface fields
-    muf() = fvc::interpolate(mu());
-    Kf()  = fvc::interpolate(K());
+    this->muf(fvc::interpolate(mu()));
+    this->Kf(fvc::interpolate(K()));
 }
 
 
