### 2024-09-10 a936697
Fixes for leastSquaresS4fGrad
