### 2024-09-10 48aba07
Fix for extendedLeastSquares in OF.com/org
