### 2024-09-10 16cc609
nonLinGeomUpdatedLagSolid: add SNES option
