+++ src/solids4FoamModels/solidModels/nonLinGeomUpdatedLagSolid/nonLinGeomUpdatedLagSolid.C
@@ -23,6 +23,9 @@ License
 #include "fvMatrices.H"
 #include "addToRunTimeSelectionTable.H"
 #include "bound.H"
+#include "solidTractionFvPatchVectorField.H"
+#include "fixedDisplacementZeroShearFvPatchVectorField.H"
+#include "symmetryFvPatchFields.H"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
@@ -43,122 +46,102 @@ addToRunTimeSelectionTable
 );
 
 
-// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
+// * * * * * * * * * * *  Private Member Functions * * * * * * * * * * * * * //
 
-nonLinGeomUpdatedLagSolid::nonLinGeomUpdatedLagSolid
-(
-    Time& runTime,
-    const word& region
-)
-:
-    solidModel(typeName, runTime, region),
-    F_
-    (
-        IOobject
-        (
-            "F",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::AUTO_WRITE
-        ),
-        mesh(),
-        dimensionedTensor("I", dimless, I)
-    ),
-    J_
-    (
-        IOobject
-        (
-            "J",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::NO_WRITE
-        ),
-        det(F_)
-    ),
-    relF_
-    (
-        IOobject
-        (
-            "relF",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::NO_WRITE
-        ),
-        I + gradDD().T()
-    ),
-    relFinv_
-    (
-        IOobject
-        (
-            "relFinv",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::NO_WRITE
-        ),
-        inv(relF_)
-    ),
-    relJ_
-    (
-        IOobject
-        (
-            "relJ",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::NO_WRITE
-        ),
-        det(relF_)
-    ),
-    rho_
-    (
-        IOobject
-        (
-            "rho",
-            runTime.timeName(),
-            mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::AUTO_WRITE
-        ),
-        mechanical().rho()
-    ),
-    impK_(mechanical().impK()),
-    impKf_(mechanical().impKf()),
-    rImpK_(1.0/impK_)
+
+void nonLinGeomUpdatedLagSolid::predict()
 {
-    DDisRequired();
+    Info<< "Linear predictor" << endl;
 
-    // Force all required old-time fields to be created
-    fvm::d2dt2(rho_, DD());
-    fvc::d2dt2(rho_, D().oldTime());
+    // Predict D using the velocity field
+    // Note: the case may be steady-state but U can still be calculated using a
+    // transient method
+    DD() = U()*runTime().deltaT() + 0.5*sqr(runTime().deltaT())*A_;
 
-    // For consistent restarts, we will update the relative kinematic fields
-    DD().correctBoundaryConditions();
-    if (restart())
+    // Update gradient of displacement increment
+    mechanical().grad(DD(), gradDD());
+
+    // Relative deformation gradient
+    relF_ = I + gradDD().T();
+
+    // Inverse relative deformation gradient
+    relFinv_ = inv(relF_);
+
+    // Total deformation gradient
[…]
