+++ src/solids4FoamModels/solidModels/explicitGodunovCC/createFields.H
@@ -1,4 +1,0 @@
-Info << "Creating mesh parameters ..." << endl;
-
-// Operations class
-
