+++ src/solids4FoamModels/solidModels/explicitGodunovCC/boundaryConditions/traction/tractionTractionFvPatchVectorField/tractionTractionFvPatchVectorField.C
@@ -1,230 +1,0 @@
-/*---------------------------------------------------------------------------*\
-  =========                 |
-  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
-   \\    /   O peration     | Website:  https://openfoam.org
-    \\  /    A nd           | Copyright (C) 2011-2018 OpenFOAM Foundation
-     \\/     M anipulation  |
--------------------------------------------------------------------------------
-License
-    This file is part of OpenFOAM.
-
-    OpenFOAM is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    OpenFOAM is distributed in the hope that it will be useful, but WITHOUT
-    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
-    for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
-
-\*---------------------------------------------------------------------------*/
-
-#include "tractionTractionFvPatchVectorField.H"
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-
-// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
-
-tractionTractionFvPatchVectorField::
-tractionTractionFvPatchVectorField
-(
-    const fvPatch& p,
-    const DimensionedField<vector, volMesh>& iF
-)
-:
-    fixedValueFvPatchVectorField(p, iF),
-    loadingType_("none"),
-    t_P_(vector::zero),
-    p_P_(0.0),
-    pressureSeries_()
-{}
-
-
-tractionTractionFvPatchVectorField::
-tractionTractionFvPatchVectorField
-(
-    const fvPatch& p,
-    const DimensionedField<vector, volMesh>& iF,
-    const dictionary& dict
-)
-:
-    fixedValueFvPatchVectorField(p, iF),
-    loadingType_(dict.lookupOrDefault<word>("loadingType", "none")),
-    t_P_(dict.lookupOrDefault("traction", vector::zero)),
-    p_P_(dict.lookupOrDefault("pressure", 0.0)),
-    pressureSeries_(dict.subDict("pressureSeries"))
-{
-    fvPatchVectorField::operator=(vectorField("value", dict, p.size()));
-
-    if (loadingType_ == "traction" && !dict.found("traction") == 1)
-    {
-        FatalErrorIn("tractionTractionFvPatchVectorField")
-            << "Keyword 'traction' is undefined in dictionary '"
-            << iF.name() << ".boundaryField." << p.name()
-            << "' for patch type '" << this->type() << "'." << nl
-            << exit(FatalError);
-    }
-
-    if (loadingType_ == "pressure" && !dict.found("pressure") == 1)
-    {
-        FatalErrorIn("tractionTractionFvPatchVectorField")
-            << "Keyword 'pressure' is undefined in dictionary '"
-            << iF.name() << ".boundaryField." << p.name()
-            << "' for patch type '" << this->type() << "'." << nl
-            << exit(FatalError);
-    }
-
[…]
