### 2024-10-22 #6989 @kyleniemeyer
@TJFord @sthavishtha @mohd-afeef-badri @MakisH would any of you be able to review this submission for JOSS? It is a toolbox for simulating solid mechanics and fluid-solid interactions in OpenFOAM. Thanks!
