### 2024-10-22 #6989 @MakisH
@kyleniemeyer I am available and have the can review this submission.

For full disclosure, I have met @philipcardiff in conferences and he has previously contributed solids4foam-related extensions to the preCICE OpenFOAM adapter and tutorials (which I work on), for which reason we appear together in a dataset publication (2022): https://darus.uni-stuttgart.de/dataset.xhtml?persistentId=doi:10.18419/darus-3576 In the same context, he has been invited speaker to a workshop I and colleagues organized in February 2023: https://precice.org/precice-workshop-2023.html

From my side, I have had no involvement to the solids4foam so far, but I have shortly used it before in the context of preCICE. I also know the project from many references in the literature. I think I can be give an objective review, and I am interested in seeing how the project evolved since the last time I had a look into it.
