### 2024-10-24 0fa8c75
Stencil construction algo
