+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -26,6 +26,7 @@ License
 #include "fixedDisplacementZeroShearFvPatchVectorField.H"
 #include "symmetryFvPatchFields.H"
 
+#include <Eigen/Dense>
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
@@ -178,6 +179,232 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
 
             // Update gradient of displacement
             mechanical().grad(D(), gradD());
+
+            // Testing new grad scheme start
+
+            // Here we will calculate cell-centred gradient using
+            // high-order disretisation approach
+            // STEPS:
+            // 1. Loop over all cells
+            //   1.1 For each cell construct stencil
+            //
+            // 2. Loop over all cells
+            //   2.1 For each cell make interpolation coefficients
+            //
+            // 3. Explicitly calculate gradient at cell centres unsing cell
+            //    stencil and corresponting interpolation coeffs for each cell
+            //    in stencil
+
+            //
+            // 1. Step - make stencils
+            //
+            //          I'm using layer approach as it has the best condition
+            //          number for LS matrix. Some authors use sphere and put
+            //          in stencil all cells inside the sphere but that produces
+            //          higher values of condition number at structured mesh
+
+            const label maxStencilSize = 70;
+            const label nLayers = 2;
+            const labelListList& cellCells = mesh().cellCells();
+
+            List<DynamicList<label>> lsStencil(mesh().nCells());
+            forAll(lsStencil, cellI)
+            {
+                lsStencil[cellI].setCapacity(maxStencilSize);
+
+                DynamicList<label>& curStencil = lsStencil[cellI];
+                const labelList& curCellCells = cellCells[cellI];
+
+                labelHashSet stencilCells;
+                labelHashSet prevLayer;
+
+                //Info<<"First layer for cell: " << cellI << endl;
+                // Add first layer of cells
+                forAll(curCellCells, cI)
+                {
+                    stencilCells.insert(curCellCells[cI]);
+                    prevLayer.insert(curCellCells[cI]);
+                }
+
+                // Remaining layers of cells
+                for (int layerI = 0; layerI < nLayers; layerI++)
+                {
+                    labelList prevLayerCells(prevLayer.toc());
+                    labelHashSet curLayer;
+
+                    // Loop over previous layer and add one level of
+                    // layer neighbours
+                    for (const label cellI : prevLayerCells)
+                    {
+                        const labelList& cellINei= mesh().cellCells()[cellI];
+                        forAll (cellINei, nei)
+                        {
+                            if (!stencilCells.found(cellINei[nei]))
+                            {
+                                curLayer.insert(cellINei[nei]);
+                            }
+                        }
+                    }
+                    // Now we have curent layer which we need to add to stencil
+                    // and current layer will now be previous layer for next
+                    // loop
+                    stencilCells.merge(curLayer);
+                    prevLayer.clear();
+                    prevLayer = curLayer;
+                }
+
+                if (Pstream::parRun())
+                {
+                    notImplemented("not implemented for parallel run");
+                }
+
+                curStencil.append(stencilCells.toc());
+            }
+
+            forAll(lsStencil, cellI)
+            {
+                // Note: lsStencil is sorted but I do not see problem in that
[…]
