### 2024-10-24 #6989 @kyleniemeyer
Whoops, sorry for the erroneous commands. I'm actually surprised `editorialbot` didn't raise an error for the editor assignment...

We have two reviewers, so I'm going to start the review issue now.
