+++ src/solids4FoamModels/fluidSolidInterfaces/coconutCouplingInterface/coupled_solvers/models/analytical_1d.py
@@ -1,0 +1,201 @@
+from coconut import tools
+from coconut.coupling_components.component import Component
+
+import numpy as np
+
+
+def create(parameters):
+    return ModelAnalytical1D(parameters)
+
+
+class ModelAnalytical1D(Component):
+    provides_get_solution = provides_set_solution = False
+
+    def __init__(self, parameters):
+        super().__init__()
+
+        # reading
+        self.parameters = parameters
+        self.settings = parameters['settings']
+
+        self.update_every_iteration = self.settings.get('update_every_iteration', False)  # update Jacobian every iter
+
+        # add timestep_start and delta_t to solver_models settings
+        for solver_model_parameters in self.settings['solver_models']:
+            tools.pass_on_parameters(self.settings, solver_model_parameters['settings'],
+                                     ['timestep_start', 'delta_t', 'save_restart'])
+
+        # initialization
+        self.solver_models = []
+        self.solver_models.append(tools.create_instance(self.settings['solver_models'][0]))  # flow model
+        self.solver_models.append(tools.create_instance(self.settings['solver_models'][1]))  # structure model
+
+        self.interface_input = self.solver_models[0].get_interface_input()  # residual of displacements
+        self.interface_output = self.solver_models[1].get_interface_output()  # displacements
+        if self.interface_input.size != self.interface_output.size:
+            raise Exception('Input and output of surrogate model have to be the same size')
+
+        self.x = None
+        self.m = None
+        self.out = None
+        self.jit = None  # dxt/dr
+
+    def initialize(self):
+        super().initialize()
+
+        for solver_model in self.solver_models:
+            solver_model.initialize()
+
+        self.x = self.solver_models[1].get_interface_output()
+        self.m = int(self.x.get_interface_data().shape[0] / 3)
+        self.out = self.x.copy()
+        self.jit = self.get_inverse_jacobian()
+
+    def finalize(self):
+        super().finalize()
+
+        for solver_model in self.solver_models:
+            solver_model.finalize()
+
+    def initialize_solution_step(self):
+        super().initialize_solution_step()
+
+        for solver_model in self.solver_models:
+            solver_model.initialize_solution_step()
+
+        if not self.update_every_iteration:
+            self.jit = self.get_inverse_jacobian()
+
+    def finalize_solution_step(self):
+        super().finalize_solution_step()
+
+        # update with correct final values
+        self.solver_models[0].solve_solution_step(self.x)
+
+        for solver_model in self.solver_models:
+            solver_model.finalize_solution_step()
+
+    def output_solution_step(self):
+        super().output_solution_step()
+
+        for solver_model in self.solver_models:
+            solver_model.output_solution_step()
+
+    def get_interface_input(self):
+        return self.interface_input.copy()
+
+    def get_interface_output(self):
+        return self.interface_output.copy()
+
+    def print_components_info(self, pre):
+        tools.print_info(pre, 'The surrogate model ', self.__class__.__name__, ' has the following solver models:')
+        tools.print_components_info(pre, self.solver_models)
+
+    def predict(self, dr_in, modes=None):
+        dr = dr_in.get_interface_data()[1::3]  # only y-displacement
+        if modes == 0:
+            tools.print_info("Jacobian of model Analytical1D disabled: zero returned", layout='warning')
+            return dr_in * 0
+        elif modes is not None:
[…]
