+++ src/solids4FoamModels/solidModels/fvPatchFields/explicitSolidTractionTraction/explicitSolidTractionTractionFvPatchVectorField.H
@@ -1,0 +1,235 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+Class
+    explicitSolidTractionTractionFvPatchVectorField
+
+Description
+    Fixed traction boundary condition for solid solvers.
+
+Author
+    Philip Cardiff, UCD.
+    Zeljko Tukovic, FSB Zagreb.
+
+SourceFiles
+    explicitSolidTractionTractionFvPatchVectorField.C
+
+\*---------------------------------------------------------------------------*/
+
+#ifndef explicitSolidTractionTractionFvPatchVectorField_H
+#define explicitSolidTractionTractionFvPatchVectorField_H
+
+#ifdef FOAMEXTEND
+    #include "foamTime.H"
+#endif
+#include "fvPatchFields.H"
+#include "fixedValueFvPatchFields.H"
+#include "Switch.H"
+#include "interpolationTable.H"
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+namespace Foam
+{
+
+/*---------------------------------------------------------------------------*\
+             Class explicitSolidTractionTractionFvPatchVectorField Declaration
+\*---------------------------------------------------------------------------*/
+
+class explicitSolidTractionTractionFvPatchVectorField
+:
+    public fixedValueFvPatchVectorField
+{
+
+    // Private Data
+
+        //- Enable non-orthogonal corrections
+        const Switch nonOrthogonalCorrections_;
+
+        //- Traction
+        vectorField traction_;
+
+        //- Pressure
+        scalarField pressure_;
+
+        //- Traction time series
+        interpolationTable<vector> tractionSeries_;
+
+        //- Pressure time series
+        interpolationTable<scalar> pressureSeries_;
+
+        //- Traction field given as a vol field
+        autoPtr<volVectorField> tractionFieldPtr_;
+
+        //- Pressure field given as a vol field
+        autoPtr<volScalarField> pressureFieldPtr_;
+
+        //- Second order accuracy correction
+        const Switch secondOrder_;
+
+        //- Switch so that the effective traction is enforced rather than the
+        //  total traction. This only applies to porous flow based solids
+        const Switch setEffectiveTraction_;
+
+        //- Optional: under-relaxation factor
+        const scalar relaxFac_;
+
+        //- Current time index
+        label curTimeIndex_;
+
+
+public:
+
+    //- Runtime type information
+    TypeName("explicitSolidTractionTraction");
+
+
+    // Constructors
+
+        //- Construct from patch and internal field
+        explicitSolidTractionTractionFvPatchVectorField
+        (
+            const fvPatch&,
+            const DimensionedField<vector, volMesh>&
+        );
+
+        //- Construct from patch, internal field and dictionary
+        explicitSolidTractionTractionFvPatchVectorField
+        (
+            const fvPatch&,
+            const DimensionedField<vector, volMesh>&,
+            const dictionary&
+        );
+
+        //- Construct by mapping given
+        //  explicitSolidTractionTractionFvPatchVectorField onto a new patch
+        explicitSolidTractionTractionFvPatchVectorField
+        (
[…]
