+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/explicitGodunovCCSolid.H
@@ -338,7 +338,9 @@ public:
             }
 
         // Edit
-
+            //- Update the size of the time-step
+            virtual void setDeltaT(Time& runTime);
+            
             //- Evolve the solid solver and solve the mathematical model
             virtual bool evolve();
 
