+++ src/solids4FoamModels/solidModels/fvPatchFields/explicitSolidTractionLinearMomentum/explicitSolidTractionLinearMomentumFvPatchVectorField.C
@@ -1,0 +1,568 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+\*---------------------------------------------------------------------------*/
+
+#include "explicitSolidTractionLinearMomentumFvPatchVectorField.H"
+#include "addToRunTimeSelectionTable.H"
+#include "volFields.H"
+#include "lookupSolidModel.H"
+#include "patchCorrectionVectors.H"
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+namespace Foam
+{
+
+// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
+
+explicitSolidTractionLinearMomentumFvPatchVectorField::
+explicitSolidTractionLinearMomentumFvPatchVectorField
+(
+    const fvPatch& p,
+    const DimensionedField<vector, volMesh>& iF
+)
+:
+    fixedValueFvPatchVectorField(p, iF),
+    nonOrthogonalCorrections_(true),
+    traction_(p.size(), vector::zero),
+    pressure_(p.size(), 0.0),
+    tractionSeries_(),
+    pressureSeries_(),
+    tractionFieldPtr_(),
+    pressureFieldPtr_(),
+    secondOrder_(false),
+    setEffectiveTraction_(false),
+    relaxFac_(1.0),
+    curTimeIndex_(-1)
+{
+    // fvPatchVectorField::operator=(patchInternalField());
+    // gradient() = vector::zero;
+}
+
+
+explicitSolidTractionLinearMomentumFvPatchVectorField::
+explicitSolidTractionLinearMomentumFvPatchVectorField
+(
+    const fvPatch& p,
+    const DimensionedField<vector, volMesh>& iF,
+    const dictionary& dict
+)
+:
+    fixedValueFvPatchVectorField(p, iF),
+    nonOrthogonalCorrections_
+    (
+        dict.lookupOrDefault<Switch>("nonOrthogonalCorrections", true)
+    ),
+    traction_(p.size(), vector::zero),
+    pressure_(p.size(), 0.0),
+    tractionSeries_(),
+    pressureSeries_(),
+    tractionFieldPtr_(),
+    pressureFieldPtr_(),
+    secondOrder_(dict.lookupOrDefault<Switch>("secondOrder", false)),
+    setEffectiveTraction_
+    (
+        dict.lookupOrDefault<Switch>("setEffectiveTraction", false)
+    ),
+    relaxFac_(dict.lookupOrDefault<scalar>("relaxationFactor", 1.0)),
+    curTimeIndex_(-1)
+{
+    Info<< "Creating " << type() << " boundary condition" << endl;
+
+    // if (dict.found("gradient"))
+    // {
+    //     gradient() = vectorField("gradient", dict, p.size());
+    // }
+    // else
+    // {
+    //     gradient() = vector::zero;
+    // }
+
+    if (dict.found("value"))
+    {
+        Field<vector>::operator=(vectorField("value", dict, p.size()));
+    }
+    else
+    {
[…]
