### 2024-11-07 63463d1
adds condition to work with "empty" BCs. cleans D() calculations and unnecessary calls.
