+++ src/solids4FoamModels/solidModels/solidModel/solidModel.C
@@ -1829,24 +1829,24 @@ void Foam::solidModel::setTraction
 
     if(type_ == "explicitGodunovCC")
     {
-        // Info << "interfaceI: "<< interfaceI << nl
-        // << "patchID: "<< patchID << nl
-        // << "patchTraction: "<< patchTraction << endl;
 
-            volVectorField& lm_b = mesh().lookupObjectRef<volVectorField>("lm_b");
-            volVectorField& t_b = mesh().lookupObjectRef<volVectorField>("t_b");
+        volVectorField& lm_b = mesh().lookupObjectRef<volVectorField>("lm_b");
+        volVectorField& t_b = mesh().lookupObjectRef<volVectorField>("t_b");
 
-            setTraction(lm_b.boundaryFieldRef()[patchID], patchTraction);
-            setTraction(t_b.boundaryFieldRef()[patchID], patchTraction);
+        setTraction(lm_b.boundaryFieldRef()[patchID], patchTraction);
+        setTraction(t_b.boundaryFieldRef()[patchID], patchTraction);
+    }
+    else
+    {
+
+#ifdef OPENFOAM_NOT_EXTEND
+        setTraction(solutionD().boundaryFieldRef()[patchID], patchTraction);
+#else
+        setTraction(solutionD().boundaryField()[patchID], patchTraction);
+#endif
     }
 
 
-
-#ifdef OPENFOAM_NOT_EXTEND
-    setTraction(solutionD().boundaryFieldRef()[patchID], patchTraction);
-#else
-    setTraction(solutionD().boundaryField()[patchID], patchTraction);
-#endif
 }
 
 
