+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/gEqns.H
@@ -10,13 +10,9 @@ rhsLm_ = fvc::surfaceIntegrate(tC_*magSf_);
 if (runTime_.timeIndex() == 0)
 {
     // Update coordinates
-    x_ += pDeltaT_*(lm_/rho_) - pDeltaT_*( (x_.prevIter() - x_.oldTime())/(deltaT_));
-    xF_ += pDeltaT_*(lmC_/rho_) - pDeltaT_*( (xF_.prevIter() - xF_.oldTime())/(deltaT_));
+    // x_ += pDeltaT_*(lm_/rho_) - pDeltaT_*( (x_.prevIter() - x_.oldTime())/(deltaT_));
+    // xF_ += pDeltaT_*(lmC_/rho_) - pDeltaT_*( (xF_.prevIter() - xF_.oldTime())/(deltaT_));
     xN_ += pDeltaT_*(lmN_/rho_)- pDeltaT_*( (xN_.prevIter() - xN_.oldTime())/(deltaT_));
-
-    D() += pDeltaT_ *(lm_/rho_) - pDeltaT_*( (D().prevIter() - D().oldTime())/(deltaT_));
-    D().correctBoundaryConditions();
-
 
     lm_ += pDeltaT_*rhsLm_ - pDeltaT_*dampingCoeff_*lm_ - pDeltaT_*( (lm_.prevIter() - lm_.oldTime())/(deltaT_));
 
@@ -27,13 +23,9 @@ if (runTime_.timeIndex() == 0)
 else
 {
     // Update coordinates
-    x_ += pDeltaT_*(lm_/rho_)   - pDeltaT_ * ( (3*x_.prevIter() - 4* x_.oldTime()   + x_.oldTime().oldTime())/(2*deltaT_));
-    xF_ += pDeltaT_*(lmC_/rho_) - pDeltaT_ * ( (3*xF_.prevIter() - 4* xF_.oldTime() + xF_.oldTime().oldTime())/(2*deltaT_));
+    // x_ += pDeltaT_*(lm_/rho_)   - pDeltaT_ * ( (3*x_.prevIter() - 4* x_.oldTime()   + x_.oldTime().oldTime())/(2*deltaT_));
+    // xF_ += pDeltaT_*(lmC_/rho_) - pDeltaT_ * ( (3*xF_.prevIter() - 4* xF_.oldTime() + xF_.oldTime().oldTime())/(2*deltaT_));
     xN_ += pDeltaT_*(lmN_/rho_) - pDeltaT_ * ( (3*xN_.prevIter() - 4* xN_.oldTime() + xN_.oldTime().oldTime())/(2*deltaT_));
-
-    D() += pDeltaT_ *(lm_/rho_) - pDeltaT_*( (3*D().prevIter() - 4* D().oldTime() + D().oldTime().oldTime())/(2*deltaT_));
-    D().correctBoundaryConditions();
-
 
     lm_ += pDeltaT_*rhsLm_ - pDeltaT_*dampingCoeff_*lm_ - pDeltaT_*( (3*lm_.prevIter() - 4* lm_.oldTime() + lm_.oldTime().oldTime())/(2*deltaT_));
 

[…]
