### 2024-11-08 982654c
Added quadratic interpolation
