### 2024-11-08 b6be9f5
Added treatement of boundary cells
