+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -204,7 +204,7 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
             //          higher values of condition number at structured mesh
 
             const label maxStencilSize = 70;
-            const label nLayers = 4;
+            const label nLayers = 2;
             const labelListList& cellCells = mesh().cellCells();
 
             List<DynamicList<label>> lsStencil(mesh().nCells());
@@ -276,7 +276,7 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
             //
 
             // Order of interpolation
-            const label N = 2;
+            const label N = 1;
 
             // Number of terms in Taylor expression
             // 1 for zero order, 4 for 1 order, 10 for second order
@@ -305,6 +305,32 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
             List<DynamicList<scalar>> cy(mesh().nCells());
             List<DynamicList<scalar>> cz(mesh().nCells());
 
+            // Cell boundary faces global index
+            List<DynamicList<label>> cellBoundaryFaces(mesh().nCells());
+
+            const polyBoundaryMesh& boundaryMesh = mesh().boundaryMesh();
+
+            forAll(boundaryMesh, patchI)
+            {
+                if (boundaryMesh[patchI].type() != emptyPolyPatch::typeName)
+                {
+                    const labelUList& faceCells = boundaryMesh[patchI].faceCells();
+
+                    forAll(faceCells, faceI)
+                    {
+                        const label cellID = faceCells[faceI];
+                        const label gI = faceI+boundaryMesh[patchI].start();
+                        cellBoundaryFaces[cellID].append(gI);
+                    }
+                }
+            }
+
+            forAll(cellBoundaryFaces, cellI)
+            {
+                cellBoundaryFaces[cellI].shrink();
+            }
+
+
             forAll(lsStencil, cellI)
             {
                 DynamicList<label>& curStencil = lsStencil[cellI];
@@ -322,7 +348,7 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
                 }
 
                 // Loop over neighbours and construct matrix Q
-                const label Nn = curStencil.size();
+                const label Nn = curStencil.size() + cellBoundaryFaces[cellI].size();
 
                 // For now I will use matrix format from Eigen/Dense library
                 Eigen::MatrixXd Q = Eigen::MatrixXd::Zero(Np, Nn);
@@ -338,13 +364,28 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
 
                 // Loop over cells in stencil, each cell have its corresponding
                 // row in Q matrix
-                forAll(curStencil, cI)
+                for(label cI=0; cI<Nn; cI++)
+                //forAll(curStencil, cI)
                 {
-                    const label neiCellID = curStencil[cI];
-                    const vector neiC = mesh().C()[neiCellID];
+                    vector neiC = vector::zero;
+
+                    if (cI < curStencil.size())
+                    {
+                        const label neiCellID = curStencil[cI];
+                        neiC = mesh().C()[neiCellID];
+                    }
+                    else
+                    {
+                        // For boundary cells we need to add boundary face as
+                        // neigbour
+                        const label i = cI-curStencil.size();
+                        const label globalFaceID = cellBoundaryFaces[cellI][i];
+                        neiC = mesh().Cf()[globalFaceID];
+                    }
+
                     const vector C = mesh().C()[cellI];
 
-                    // Add linear interpolation part
[…]
