### 2024-11-09 b83c64c
Minor output log modifications
