+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -1,0 +1,564 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+\*---------------------------------------------------------------------------*/
+
+#include "higherOrderGrad.H"
+#include "volFields.H"
+#include "surfaceFields.H"
+#include "emptyPolyPatch.H"
+#include <Eigen/Dense>
+
+
+namespace Foam
+{
+
+// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
+
+defineTypeNameAndDebug(higherOrderGrad, 0);
+
+
+// * * * * * * * * * * *  Private Member Functions * * * * * * * * * * * * * //
+
+
+void higherOrderGrad::makeStencils() const
+{
+    Info<< "higherOrderGrad::makeStencils()" << endl;
+
+    if (stencilsPtr_)
+    {
+        FatalErrorInFunction
+            << "Pointer already set!" << abort(FatalError);
+    }
+
+    const fvMesh& mesh = mesh_;
+    const labelListList& cellCells = mesh.cellCells();
+
+    stencilsPtr_.set(new List<DynamicList<label>>(mesh.nCells()));
+    List<DynamicList<label>>& stencils = *stencilsPtr_;
+
+    forAll(stencils, cellI)
+    {
+       DynamicList<label>& curStencil = stencils[cellI];
+       curStencil.setCapacity(maxStencilSize_);
+       const labelList& curCellCells = cellCells[cellI];
+
+       labelHashSet stencilCells;
+       labelHashSet prevLayer;
+
+       // Add first layer of cells
+       forAll(curCellCells, cI)
+       {
+           stencilCells.insert(curCellCells[cI]);
+           prevLayer.insert(curCellCells[cI]);
+       }
+
+       // Remaining layers of cells
+       for (int layerI = 1; layerI < nLayers_; layerI++)
+       {
+           labelList prevLayerCells(prevLayer.toc());
+           labelHashSet curLayer;
+
+           // Loop over previous layer and add one level of
+           // layer neighbours
+           for (const label cI : prevLayerCells)
+           {
+               const labelList& cellINei= mesh.cellCells()[cI];
+               forAll (cellINei, nei)
+               {
+                   if (!stencilCells.found(cellINei[nei]))
+                   {
+                       curLayer.insert(cellINei[nei]);
+                   }
+               }
+           }
+
+           // Now we have curent layer which we need to add to the stencil
+           // and the current layer will now be the previous layer for next
+           // loop
+           prevLayer.clear();
+           prevLayer = curLayer;
+           stencilCells.merge(curLayer);
+       }
+
+       if (Pstream::parRun())
+       {
+           notImplemented("not implemented for parallel run");
+       }
[…]
