### 2024-11-13 4caf0b0
Bug fix for higher order gradient
Higher order gradient scheme implemented by Ivan Batistic, with re-packaging by Philip Cardiff

Co-authored-by: Philip Cardiff <philipcardiff@gmail.com>
Co-authored-by: Ivan Batistic <ivan.batistic2@gmail.com>
