+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -187,202 +187,202 @@ void higherOrderGrad::calcCoeffs() const
 
     forAll(boundaryMesh, patchI)
     {
-       if (boundaryMesh[patchI].type() != emptyPolyPatch::typeName)
-       {
-           const labelUList& faceCells = boundaryMesh[patchI].faceCells();
+        if (boundaryMesh[patchI].type() != emptyPolyPatch::typeName)
+        {
+            const labelUList& faceCells = boundaryMesh[patchI].faceCells();
 
-           forAll(faceCells, faceI)
-           {
-               const label cellID = faceCells[faceI];
-               const label gI = faceI + boundaryMesh[patchI].start();
-               cellBoundaryFaces[cellID].append(gI);
-           }
-       }
+            forAll(faceCells, faceI)
+            {
+                const label cellID = faceCells[faceI];
+                const label gI = faceI + boundaryMesh[patchI].start();
+                cellBoundaryFaces[cellID].append(gI);
+            }
+        }
     }
 
     forAll(cellBoundaryFaces, cellI)
     {
-       cellBoundaryFaces[cellI].shrink();
+        cellBoundaryFaces[cellI].shrink();
     }
 
     const List<DynamicList<label>> stencils = this->stencils();
 
     forAll(stencils, cellI)
     {
-       const DynamicList<label>& curStencil = stencils[cellI];
+        const DynamicList<label>& curStencil = stencils[cellI];
 
-       // Find max distance in this stencil
-       scalar maxDist = 0.0;
-       forAll(curStencil, cI)
-       {
-           const label neiCellID = curStencil[cI];
-           const scalar d = mag(CI[neiCellID] - CI[cellI]);
-           if ( d > maxDist)
-           {
-               maxDist = d;
-           }
-       }
+        // Find max distance in this stencil
+        scalar maxDist = 0.0;
+        forAll(curStencil, cI)
+        {
+            const label neiCellID = curStencil[cI];
+            const scalar d = mag(CI[neiCellID] - CI[cellI]);
+            if (d > maxDist)
+            {
+                maxDist = d;
+            }
+        }
 
-       // Loop over neighbours and construct matrix Q
-       const label Nn = curStencil.size() + cellBoundaryFaces[cellI].size();
+        // Loop over neighbours and construct matrix Q
+        const label Nn = curStencil.size() + cellBoundaryFaces[cellI].size();
 
-       // For now I will use matrix format from Eigen/Dense library
-       Eigen::MatrixXd Q = Eigen::MatrixXd::Zero(Np, Nn);
+        // For now I will use matrix format from Eigen/Dense library
+        Eigen::MatrixXd Q = Eigen::MatrixXd::Zero(Np, Nn);
 
-       // Check to avoid Eigen error
-       if (Nn < Np)
-       {
-           FatalErrorInFunction
-               << "Interpolation stencil needs to be bigger than the "
-              "number of elements in Taylor order!"
-               << exit(FatalError);
-       }
+        // Check to avoid Eigen error
+        if (Nn < Np)
+        {
+            FatalErrorInFunction
+                << "Interpolation stencil needs to be bigger than the "
+               "number of elements in Taylor order!"
+                << exit(FatalError);
+        }
 
-       // Loop over cells in stencil, each cell have its corresponding
-       // row in Q matrix
-       for (label cI = 0; cI < Nn; cI++)
-       {
-           vector neiC = vector::zero;
+        // Loop over cells in stencil, each cell have its corresponding
[…]
