### 2024-11-14 8eb9987
Add optional boundary patches to least squares stencils
