### 2024-11-14 154374f
Efficiency changes for higherOrderGrad
