+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -23,6 +23,7 @@ License
 #include "emptyPolyPatch.H"
 #include <Eigen/Dense>
 
+#include <ctime>
 
 namespace Foam
 {
@@ -214,6 +215,17 @@ void higherOrderGrad::calcCoeffs() const
     const List<DynamicList<label>>& cellBoundaryFaces = this->cellBoundaryFaces();
     const List<DynamicList<label>> stencils = this->stencils();
 
+    if (N_ < 1)
+    {
+        FatalErrorInFunction
+            << "N must be at least 1!" << exit(FatalError);
+    }
+
+    if (N_ > 2)
+    {
+        notImplemented("Orders higher that quadratic not implemented");
+    }
+
     forAll(stencils, cellI)
     {
         const DynamicList<label>& curStencil = stencils[cellI];
@@ -234,68 +246,91 @@ void higherOrderGrad::calcCoeffs() const
         const label Nn = curStencil.size() + cellBoundaryFaces[cellI].size();
 
         // For now I will use matrix format from Eigen/Dense library
-        Eigen::MatrixXd Q = Eigen::MatrixXd::Zero(Np, Nn);
+        //Eigen::MatrixXd Q = Eigen::MatrixXd::Zero(Np, Nn);
+        // Avoid initialisation to zero as we will set every entry below
+        Eigen::MatrixXd Q(Np, Nn);
 
         // Check to avoid Eigen error
         if (Nn < Np)
         {
             FatalErrorInFunction
                 << "Interpolation stencil needs to be bigger than the "
-               "number of elements in Taylor order!"
+                << "number of elements in Taylor order!"
                 << exit(FatalError);
         }
 
         // Loop over cells in stencil, each cell have its corresponding
         // row in Q matrix
-        for (label cI = 0; cI < Nn; cI++)
-        {
-            vector neiC = vector::zero;
 
-            if (cI < curStencil.size())
+        // Add linear interpolation part N = 1
+        for (label cI = 0; cI < curStencil.size(); cI++)
+        {
+            const label neiCellID = curStencil[cI];
+            const vector& neiC = CI[neiCellID];
+            const vector& C = CI[cellI];
+            const vector dx = neiC - C;
+            Q(0, cI) = 1;
+            Q(1, cI) = dx.x();
+            Q(2, cI) = dx.y();
+            Q(3, cI) = dx.z();
+        }
+
+        // Add quadratic interpolation part N = 2
+        if (N_ > 1)
+        {
+            for (label cI = 0; cI < curStencil.size(); cI++)
             {
                 const label neiCellID = curStencil[cI];
-                neiC = CI[neiCellID];
-            }
-            else
-            {
-                // For boundary cells we need to add boundary face as
-                // neigbour
-                const label i = cI - curStencil.size();
-                const label globalFaceID = cellBoundaryFaces[cellI][i];
-                neiC = CfI[globalFaceID];
-            }
-
-            const vector& C = CI[cellI];
-
-            // Add linear interpolation part N = 1
-            if (N_ > 0)
-            {
-                Q(0, cI) = 1;
-                Q(1, cI) = neiC.x() - C.x();
-                Q(2, cI) = neiC.y() - C.y();
-                Q(3, cI) = neiC.z() - C.z();
-            }
-
-            // Add quadratic interpolation part N = 2
-            if (N_ > 1)
-            {
[…]
