+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -21,9 +21,6 @@ License
 #include "volFields.H"
 #include "surfaceFields.H"
 #include "emptyPolyPatch.H"
-#include <Eigen/Dense>
-
-#include <ctime>
 
 namespace Foam
 {
@@ -177,6 +174,17 @@ void higherOrderGrad::calcCoeffs() const
 
     interpGradCoeffsPtr_.set(new List<DynamicList<vector>>(mesh.nCells()));
     List<DynamicList<vector>>& interpGradCoeffs = *interpGradCoeffsPtr_;
+
+    if (!useQRDecomposition_)
+    {
+        choleskyPtr_.set(new List<Eigen::LLT<Eigen::MatrixXd>>(mesh.nCells()));
+        QhatPtr_.set(new List<Eigen::MatrixXd>(mesh.nCells()));
+    }
+
+    sqrtWPtr_.set
+    (
+        new List<Eigen::DiagonalMatrix<double, Eigen::Dynamic>>(mesh.nCells())
+    );
 
     // Refernces for brevity and efficiency
     const vectorField& CI = mesh.C();
@@ -368,105 +376,164 @@ void higherOrderGrad::calcCoeffs() const
         }
 
         // Now when we have W and Q, next step is QR decomposition
-        //const Eigen::MatrixXd sqrtW = W.cwiseSqrt();
-        const Eigen::DiagonalMatrix<double, Eigen::Dynamic> sqrtW =
-            W.diagonal().cwiseSqrt().asDiagonal();
-        //const Eigen::MatrixXd Qhat = Q*sqrtW;
-        //const Eigen::MatrixXd Qhat = Q*sqrtW.diagonal().asDiagonal();
+        Eigen::DiagonalMatrix<double, Eigen::Dynamic>& sqrtW =
+            sqrtWPtr_()[cellI];
+        sqrtW = W.diagonal().cwiseSqrt().asDiagonal();
         const Eigen::MatrixXd Qhat =
             Q.array().rowwise()*sqrtW.diagonal().transpose().array();
-        Eigen::HouseholderQR<Eigen::MatrixXd> qr(Qhat.transpose());
-        //Eigen::ColPivHouseholderQR<Eigen::MatrixXd> qr(Qhat.transpose());
-
-        // Q and R matrices
-        const Eigen::MatrixXd O = qr.householderQ();
-        const Eigen::MatrixXd& R = qr.matrixQR().triangularView<Eigen::Upper>();
 
         // B hat
-        // const Eigen::MatrixXd Bhat =
-        //     sqrtW*Eigen::MatrixXd::Identity(W.rows(), W.cols());
-        //const Eigen::MatrixXd Bhat = sqrtW.diagonal().asDiagonal();
         const Eigen::DiagonalMatrix<double, Eigen::Dynamic>& Bhat =
             sqrtW.diagonal().asDiagonal();
 
-        // Slice Rbar and Qbar, as we do not need full matrix
-        // const Eigen::MatrixXd Rbar = R.topLeftCorner(Np, Np);
-        // const Eigen::MatrixXd Qbar = O.leftCols(Np);
-        const auto Rbar = R.topLeftCorner(Np, Np);
-        const auto Qbar = O.leftCols(Np);
+        // Declare A outside the if-else scope, but do not initialise
+        Eigen::MatrixXd A;
 
-        // Perform element-wise multiplication and convert to MatrixXd
-        const Eigen::MatrixXd QbarBhat =
-            (
-                Qbar.transpose().array().rowwise()
-               *Bhat.diagonal().transpose().array()
-            ).matrix();
-
-        // Solve to get A
-        // const Eigen::MatrixXd A =
-        //     Rbar.colPivHouseholderQr().solve(Qbar.transpose()*Bhat);
[…]
