### 2024-11-16 4ee11b4
Generalise generation of Taylor series terms
