### 2024-11-17 6f60429
Corrects run in parallel, corrects angular momentum algorithm,  reduces dependcy on old directories, cleans unnecessary code
