+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/models/solidMaterialModel/solidMaterialModel.H
@@ -55,6 +55,8 @@ class solidMaterialModel
         //- Energy of algorithm
         volScalarField energyAlgorithm_;
 
+        // const dictionary& mechanicalPropertiesDict_; 
+
        //- Constitutive model
         word model_;
 
@@ -116,6 +118,10 @@ public:
 
         // Access
 
+           const  word& model() const
+            {
+                return model_;
+            }
             const volTensorField& piola() const
             {
                 return P_;
