+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/explicitGodunovCCSolid.C
@@ -39,35 +39,6 @@ defineTypeNameAndDebug(explicitGodunovCCSolid, 0);
 addToRunTimeSelectionTable(solidModel, explicitGodunovCCSolid, dictionary);
 
 
-// * * * * * * * * * * *  Private Member Functions * * * * * * * * * * * * * //
-
-void explicitGodunovCCSolid::updateStress()
-{
-    // Update increment of displacement
-    DD() = D() - D().oldTime();
-
-    // Update gradient of displacement
-    mechanical().grad(D(), gradD());
-
-    // Update gradient of displacement increment
-    gradDD() = gradD() - gradD().oldTime();
-
-    // Calculate the stress using run-time selectable mechanical law
-    // mechanical().correct(sigma());
-    sigma() =  symm( (1.0 / J_) * (P_ & F_.T()));
-    // // Interpolate cell displacements to vertices
-    // mechanical().interpolate(D(), pointD());
-
-    mechanical().interpolate(D(), gradD(), pointD());
-
-    // Increment of displacement
-    DD() = D() - D().oldTime();
-    
-    // // Increment of point displacement
-    pointDD() = pointD() - pointD().oldTime();
-}
-
-
 // * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
 
 explicitGodunovCCSolid::explicitGodunovCCSolid
@@ -77,10 +48,6 @@ explicitGodunovCCSolid::explicitGodunovCCSolid
 )
 :
     solidModel(typeName, runTime, region),
-    impK_(mechanical().impK()),
-    impKf_(mechanical().impKf()),
-    rImpK_(1.0/impK_),
-    //----------------------------------------------
     runTime_(runTime),
     //reading dicts
     mechanicalProperties_//! to be changed
@@ -106,47 +73,23 @@ explicitGodunovCCSolid::explicitGodunovCCSolid
             IOobject::NO_WRITE
         )
     ),
-
-    fvSolution_
-    (
-        IOobject
-        (
-            "fvSolution",
-            runTime.system(),
-            mesh(),
-            IOobject::MUST_READ_IF_MODIFIED,
-            IOobject::NO_WRITE
-        )
-    ),
     
     beta_
     (
-        readScalar(fvSolution_.lookup("incompressiblilityCoefficient"))
+        solidModelDict().lookupOrAddDefault<scalar>("incompressiblilityCoefficient", 1)
     ),
    
     angularMomentumConservation_
     (
-        fvSolution_.lookup("angularMomentumConservation")
+        solidModelDict().lookupOrAddDefault<word>("angularMomentumConservation", "no")
     ),
     
-    dampingCoeff_
-    (
-        "dampingCoeff",
-        dimensionSet(0,0,-1, 0, 0, 0, 0),
-        fvSolution_
-    ),
-
 //-----------------------------------------------------------------
 
     op_(mesh()),
     magSf_(mesh().magSf()),
     Sf_(mesh().Sf()),
     h_(op_.minimumEdgeLength()),
-    // bm_(mesh().boundaryMesh()),
-    // symmetricPatchID_(bm_.findPatchID("symmetric")),
-    // symmetricXpatchID_(bm_.findPatchID("symmetricX")),
-    // symmetricYpatchID_(bm_.findPatchID("symmetricY")),
-    // symmetricZpatchID_(bm_.findPatchID("symmetricZ")),
 
     // Creating mesh coordinate fields 
     C_(mesh().C()),
@@ -396,13 +339,14 @@ explicitGodunovCCSolid::explicitGodunovCCSolid
         angularMomentumConservation_ != "yes" && angularMomentumConservation_ != "no"
     )
     {
-        FatalErrorIn("readControls.H")
[…]
