+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/riemannSolver.H
@@ -24,98 +24,98 @@ if (Pstream::parRun())
 
 forAll(mesh().boundary(), patchi)
 {
-    // // Riemann solver for inter-processor boundaries
-    // if (mesh().boundary()[patchi].coupled())
-    // {
-    //     const vectorField lm_nei(
-    //       lm.boundaryField()[patchi].patchNeighbourField());
+    // Riemann solver for inter-processor boundaries
+    if (mesh().boundary()[patchi].coupled())
+    {
+        const vectorField lm_nei(
+          lm_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const tensorField P_nei(
-    //       P.boundaryField()[patchi].patchNeighbourField());
+        const tensorField P_nei(
+          P_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const vectorField Px_nei(
-    //       Px.boundaryField()[patchi].patchNeighbourField());
+        const vectorField Px_nei(
+          Px_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const vectorField Py_nei(
-    //       Py.boundaryField()[patchi].patchNeighbourField());
+        const vectorField Py_nei(
+          Py_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const vectorField Pz_nei(
-    //       Pz.boundaryField()[patchi].patchNeighbourField());
+        const vectorField Pz_nei(
+          Pz_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const tensorField lmGrad_nei(
-    //       lmGrad.boundaryField()[patchi].patchNeighbourField());
+        const tensorField lmGrad_nei(
+          lmGrad_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const tensorField PxGrad_nei(
-    //       PxGrad.boundaryField()[patchi].patchNeighbourField());
+        const tensorField PxGrad_nei(
+          PxGrad_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const tensorField PyGrad_nei(
-    //       PyGrad.boundaryField()[patchi].patchNeighbourField());
+        const tensorField PyGrad_nei(
+          PyGrad_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const tensorField PzGrad_nei(
-    //       PzGrad.boundaryField()[patchi].patchNeighbourField());
+        const tensorField PzGrad_nei(
+          PzGrad_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const vectorField C_nei(
-    //       C.boundaryField()[patchi].patchNeighbourField());
+        const vectorField C_nei(
+          C_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const scalarField Up_nei(
-    //       Up.boundaryField()[patchi].patchNeighbourField());
+        const scalarField Up_nei(
+          Up_.boundaryField()[patchi].patchNeighbourField());
 
-    //     const scalarField Us_nei(
-    //       Us.boundaryField()[patchi].patchNeighbourField());
+        const scalarField Us_nei(
+          Us_.boundaryField()[patchi].patchNeighbourField());
 
-    //     forAll(mesh().boundary()[patchi], facei)
-    //     {
-    //         const label& bCell =
-    //             mesh().boundaryMesh()[patchi].faceCells()[facei];
[…]
