+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -163,6 +163,12 @@ void higherOrderGrad::generateExponents
     DynamicList<FixedList<label, 3>>& exponents
 ) const
 {
+    if (N_ < 1)
+    {
+        FatalErrorInFunction
+            << "N must be at least 1!" << exit(FatalError);
+    }
+
     // Estimate the number of terms to set the capacity
     const label estimatedSize = (N + 1)*(N + 2)*(N + 3)/6;
     exponents.setCapacity(estimatedSize);
@@ -193,11 +199,11 @@ void higherOrderGrad::generateExponents
 }
 
 
-void higherOrderGrad::calcCoeffs() const
+void higherOrderGrad::calcQRCoeffs() const
 {
-    Info<< "higherOrderGrad::calcCoeffs()" << endl;
+    Info<< "higherOrderGrad::calcQRCoeffs()" << endl;
 
-    if (interpCoeffsPtr_ || interpGradCoeffsPtr_)
+    if (QRInterpCoeffsPtr_ || QRGradCoeffsPtr_)
     {
         FatalErrorInFunction
             << "Pointers already set!" << abort(FatalError);
@@ -205,22 +211,11 @@ void higherOrderGrad::calcCoeffs() const
 
     const fvMesh& mesh = mesh_;
 
-    interpCoeffsPtr_.set(new List<DynamicList<scalar>>(mesh.nCells()));
-    List<DynamicList<scalar>>& interpCoeffs = *interpCoeffsPtr_;
+    QRInterpCoeffsPtr_.set(new List<DynamicList<scalar>>(mesh.nCells()));
+    List<DynamicList<scalar>>& QRInterpCoeffs = *QRInterpCoeffsPtr_;
 
-    interpGradCoeffsPtr_.set(new List<DynamicList<vector>>(mesh.nCells()));
-    List<DynamicList<vector>>& interpGradCoeffs = *interpGradCoeffsPtr_;
-
-    if (!useQRDecomposition_)
-    {
-        choleskyPtr_.set(new List<Eigen::LLT<Eigen::MatrixXd>>(mesh.nCells()));
-        QhatPtr_.set(new List<Eigen::MatrixXd>(mesh.nCells()));
-    }
-
-    sqrtWPtr_.set
-    (
-        new List<Eigen::DiagonalMatrix<double, Eigen::Dynamic>>(mesh.nCells())
-    );
+    QRGradCoeffsPtr_.set(new List<DynamicList<vector>>(mesh.nCells()));
+    List<DynamicList<vector>>& QRGradCoeffs = *QRGradCoeffsPtr_;
 
     // Refernces for brevity and efficiency
     const vectorField& CI = mesh.C();
@@ -243,27 +238,6 @@ void higherOrderGrad::calcCoeffs() const
         factorials[n] = factorials[n - 1]*n;
     }
 
-    if (calcConditionNumber_)
-    {
-        // Cells condition numbers
-        condNumberPtr_.set
-        (
-            new volScalarField
-            (
-               IOobject
-               (
-                   "condNumber",
-                   mesh.time().timeName(),
-                   mesh,
-                   IOobject::NO_READ,
-                   IOobject::AUTO_WRITE
-               ),
-               mesh,
-               dimensionedScalar("0", dimless, Zero)
-            )
-        );
-    }
-
     List<DynamicList<scalar>> c(mesh.nCells());
     List<DynamicList<scalar>> cx(mesh.nCells());
     List<DynamicList<scalar>> cy(mesh.nCells());
@@ -271,12 +245,6 @@ void higherOrderGrad::calcCoeffs() const
 
     const List<DynamicList<label>>& cellBoundaryFaces = this->cellBoundaryFaces();
[…]
