### 2024-11-20 7de0f88
higherOrderGrad: add new makeGlobalCellStencils which works in parallel; next step is to calculate the coefficients using these global stencils
