+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -21,6 +21,8 @@ License
 #include "volFields.H"
 #include "surfaceFields.H"
 #include "emptyPolyPatch.H"
+#include "processorPolyPatch.H"
+
 
 namespace Foam
 {
@@ -33,11 +35,433 @@ defineTypeNameAndDebug(higherOrderGrad, 0);
 // * * * * * * * * * * *  Private Member Functions * * * * * * * * * * * * * //
 
 
+void higherOrderGrad::makeGlobalCellStencils() const
+{
+    InfoInFunction
+        << "start" << endl;
+
+    if (globalCellStencilsPtr_)
+    {
+        FatalErrorInFunction
+            << "Pointer already set" << exit(FatalError);
+    }
+
+    // References from brevity and efficiency
+    const fvMesh& mesh = mesh_;
+    const label nCells = mesh.nCells();
+    const labelUList& owner = mesh.owner();
+    const labelUList& neighbour = mesh.neighbour();
+    const cellList& cells = mesh.cells();
+
+    // Prepare and store processor neighbour face global cells, i.e. global cell
+    // indices of the cells across the processor patches
+    labelListList procPatchNeiGlobalCellIDs(mesh.boundaryMesh().size());
+    if (Pstream::parRun())
+    {
+        PstreamBuffers pBufs(UPstream::commsTypes::nonBlocking);
+
+        // Send global cell IDs for each processor patch
+        forAll(mesh.boundaryMesh(), patchI)
+        {
+            const polyPatch& pp = mesh.boundaryMesh()[patchI];
+            if (isA<processorPolyPatch>(pp))
+            {
+                const processorPolyPatch& procPatch =
+                    refCast<const processorPolyPatch>(pp);
+
+                UOPstream toNeighbProc(procPatch.neighbProcNo(), pBufs);
+
+                toNeighbProc
+                    << globalCells_.toGlobal(pp.faceCells());
+            }
+        }
+
+        pBufs.finishedSends(); // no-op for blocking
+
+        // Receive data
+        forAll(mesh.boundaryMesh(), patchI)
+        {
+            const polyPatch& pp = mesh.boundaryMesh()[patchI];
+            if (isA<processorPolyPatch>(pp))
+            {
+                const processorPolyPatch& procPatch =
+                    refCast<const processorPolyPatch>(pp);
+
+                UIPstream fromNeighbProc(procPatch.neighbProcNo(), pBufs);
+
+                procPatchNeiGlobalCellIDs[patchI].setSize(pp.size());
+
+                fromNeighbProc
+                    >> procPatchNeiGlobalCellIDs[patchI];
+            }
+        }
+    }
+
+    // Initialize visited sets and frontiers for all cells
+    // For each local cell, we will find and store all cells in the stecil as
+    // global cell indices
+    List<labelHashSet> cellVisitedSets(nCells);
+    List<labelList> cellFrontiers(nCells);
+
+    // Initialize the visited sets and current frontiers
+    for (label localCellI = 0; localCellI < nCells; ++localCellI)
+    {
+        // Get global cell ID
+        const label globalCellI = globalCells_.toGlobal(localCellI);
+
+        // Initialize the visited set with the cell itself
+        cellVisitedSets[localCellI].insert(globalCellI);
+
+        // Initialize current frontier with immediate neighbors
+        DynamicList<label> neighborCellIDs;
+
+        const labelList& cellFaces = cells[localCellI];
+
+        forAll(cellFaces, faceI)
+        {
+            const label curFaceID = cellFaces[faceI];
+
+            if (mesh.isInternalFace(curFaceID))
+            {
+                // Internal face
+                const label own = owner[curFaceID];
+                const label nei = neighbour[curFaceID];
[…]
