### 2024-11-20 83eb1c4
Added higherOrderGrad::requestGlobalStencilData function
