### 2024-11-20 3720dca
higherOrderGrad: parallel version of Cholesky seems to work
