+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -675,7 +675,7 @@ void higherOrderGrad::calcQRCoeffs() const
 
     const List<DynamicList<label>>& stencilsBoundaryFaces =
         this->stencilsBoundaryFaces();
-    const List<DynamicList<label>> stencils = this->stencils();
+    const List<DynamicList<label>>& stencils = this->stencils();
 
     forAll(stencils, cellI)
     {
@@ -938,7 +938,7 @@ void higherOrderGrad::calcCholeskyCoeffs() const
 
     const List<DynamicList<label>>& stencilsBoundaryFaces =
         this->stencilsBoundaryFaces();
-    const List<DynamicList<label>> stencils = this->stencils();
+    const List<DynamicList<label>>& stencils = this->stencils();
 
     forAll(stencils, cellI)
     {
@@ -1115,6 +1115,238 @@ void higherOrderGrad::calcCholeskyCoeffs() const
 }
 
 
+void higherOrderGrad::calcGlobalCholeskyCoeffs() const
+{
+    InfoInFunction
+        << "start" << endl;
+
+    if (choleskyPtr_ || QhatPtr_ || sqrtWPtr_)
+    {
+        FatalErrorInFunction
+            << "Pointers already set!" << abort(FatalError);
+    }
+
+    const fvMesh& mesh = mesh_;
+
+    choleskyPtr_.set(new List<Eigen::LLT<Eigen::MatrixXd>>(mesh.nCells()));
+    List<Eigen::LLT<Eigen::MatrixXd>>& cholesky = choleskyPtr_();
+
+    QhatPtr_.set(new List<Eigen::MatrixXd>(mesh.nCells()));
+    List<Eigen::MatrixXd>& Qhat = QhatPtr_();
+
+    sqrtWPtr_.set
+    (
+        new List<Eigen::DiagonalMatrix<double, Eigen::Dynamic>>(mesh.nCells())
+    );
+    List<Eigen::DiagonalMatrix<double, Eigen::Dynamic>>& sqrtW = sqrtWPtr_();
+
+    // Refernces for brevity and efficiency
+    const vectorField& CI = mesh.C();
+
+    // Collect CI for off-processor cells in the stencils
+    Map<vector> globalCI;
+    requestGlobalStencilData(CI, globalCI);
+
+    // Calculate Taylor series exponents
+    // 1 for zero order, 4 for 1 order, 10 for second order, etc.
+    DynamicList<FixedList<label, 3>> exponents;
+    generateExponents(N_, exponents);
+    const label Np = exponents.size();
+    if (debug)
+    {
+        Info<< "Np = " << Np << endl;
+    }
+
+    // Precompute factorials up to N
+    List<scalar> factorials(N_ + 1, 1.0);
+    for (label n = 1; n <= N_; ++n)
+    {
+        factorials[n] = factorials[n - 1]*n;
+    }
+
+    List<DynamicList<scalar>> c(mesh.nCells());
+    List<DynamicList<scalar>> cx(mesh.nCells());
+    List<DynamicList<scalar>> cy(mesh.nCells());
+    List<DynamicList<scalar>> cz(mesh.nCells());
+
+    const labelListList& stencils = this->globalCellStencils();
+
+    forAll(stencils, localCellI)
+    {
+        const labelList& curStencil = stencils[localCellI];
+
+        // Find max distance in this stencil
+        scalar maxDist = 0.0;
+        forAll(curStencil, cI)
+        {
+            const label neiGlobalCellID = curStencil[cI];
+
+            scalar d;
+            if (globalCells_.isLocal(neiGlobalCellID))
+            {
+                const label neiLocalCellID =
+                    globalCells_.toLocal(neiGlobalCellID);
[…]
