### 2024-11-21 2527f41
gradGlobalQR implemented: works in parallel but seems slower: parallel communication approach to be checked
