+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -452,7 +452,11 @@ void higherOrderGrad::makeGlobalCellStencils() const
 
 void higherOrderGrad::makeStencils() const
 {
-    Info<< "higherOrderGrad::makeStencils()" << endl;
+    //if (debug)
+    {
+        InfoInFunction
+            << "start" << endl;
+    }
 
     if (Pstream::parRun())
     {
@@ -561,7 +565,11 @@ void higherOrderGrad::makeStencils() const
         stencilsBoundaryFaces[cellI].shrink();
     }
 
-    Info<< "higherOrderGrad::makeStencils(): end" << endl;
+    //if (debug)
+    {
+        InfoInFunction
+            << "end" << endl;
+    }
 }
 
 
@@ -631,7 +639,11 @@ void higherOrderGrad::generateExponents
 
 void higherOrderGrad::calcQRCoeffs() const
 {
-    Info<< "higherOrderGrad::calcQRCoeffs()" << endl;
+    //if (debug)
+    {
+        InfoInFunction
+            << "start" << endl;
+    }
 
     if (QRInterpCoeffsPtr_ || QRGradCoeffsPtr_)
     {
@@ -882,7 +894,278 @@ void higherOrderGrad::calcQRCoeffs() const
        QRGradCoeffs[cellI].shrink();
     }
 
-    Info<< "higherOrderGrad::calcQRCoeffs(): end" << endl;
+    //if (debug)
+    {
+        InfoInFunction
+            << "end" << endl;
+    }
+}
+
+
+void higherOrderGrad::calcGlobalQRCoeffs() const
+{
+    //if (debug)
+    {
+        InfoInFunction
+            << "start" << endl;
+    }
+
+    if (QRInterpCoeffsPtr_ || QRGradCoeffsPtr_)
+    {
+        FatalErrorInFunction
+            << "Pointers already set!" << abort(FatalError);
+    }
+
+    const fvMesh& mesh = mesh_;
+
+    QRInterpCoeffsPtr_.set(new List<DynamicList<scalar>>(mesh.nCells()));
+    List<DynamicList<scalar>>& QRInterpCoeffs = *QRInterpCoeffsPtr_;
+
+    QRGradCoeffsPtr_.set(new List<DynamicList<vector>>(mesh.nCells()));
+    List<DynamicList<vector>>& QRGradCoeffs = *QRGradCoeffsPtr_;
+
+    // Refernces for brevity and efficiency
+    const vectorField& CI = mesh.C();
+
+    // Collect CI for off-processor cells in the stencils
+    Map<vector> globalCI;
+    requestGlobalStencilData(CI, globalCI);
+
+    // Calculate Taylor series exponents
+    // 1 for zero order, 4 for 1 order, 10 for second order, etc.
+    DynamicList<FixedList<label, 3>> exponents;
+    generateExponents(N_, exponents);
+    const label Np = exponents.size();
+    if (debug)
+    {
+        Info<< "Np = " << Np << endl;
+    }
+
+    // Precompute factorials up to N
+    List<scalar> factorials(N_ + 1, 1.0);
+    for (label n = 1; n <= N_; ++n)
+    {
+        factorials[n] = factorials[n - 1]*n;
+    }
+
+    List<DynamicList<scalar>> c(mesh.nCells());
+    List<DynamicList<scalar>> cx(mesh.nCells());
+    List<DynamicList<scalar>> cy(mesh.nCells());
+    List<DynamicList<scalar>> cz(mesh.nCells());
+
+    const List<labelList>& stencils = globalCellStencils();
+
+    forAll(stencils, localCellI)
+    {
+        const labelList& curStencil = stencils[localCellI];
+
+        // Find max distance in this stencil
+        scalar maxDist = 0.0;
[…]
