+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -450,6 +450,121 @@ void higherOrderGrad::makeGlobalCellStencils() const
 }
 
 
+void higherOrderGrad::makeGlobalFaceStencils() const
+{
+    if (globalFaceStencilsPtr_)
+    {
+        FatalErrorInFunction
+            << "Pointer already set" << exit(FatalError);
+    }
+
+    const fvMesh& mesh = mesh_;
+    const labelList& faceOwner = mesh.faceOwner();
+    const labelList& faceNeighbour = mesh.faceNeighbour();
+
+    globalFaceStencilsPtr_.set(new labelListList(mesh.nFaces()));
+    labelListList& faceStencils = globalFaceStencilsPtr_();
+
+    // Create the cell stencils
+    const labelListList& cellStencils = globalCellStencils();
+
+    // The stencil for each face consists of the cell stencils of the cell(s)
+    // containing that face
+    // We will only keep unique entries
+    forAll(faceStencils, faceI)
+    {
+        // We will use a set to check avoid adding duplicates
+        labelHashSet curStencil;
+
+        // Add owner stencil
+        const label ownLocalCellI = faceOwner[faceI];
+        curStencil.insert(cellStencils[ownLocalCellI]);
+
+        // Add neighbour stencil
+        // Note: we deal with processor patch faces afterwards
+        if (mesh.isInternalFace(faceI))
+        {
+            const label neiLocalCellI = faceNeighbour[faceI];
+            curStencil.merge(labelHashSet(cellStencils[neiLocalCellI]));
+        }
+
+        // Convert set to labelList
+        faceStencils[faceI] = curStencil.toc();
+    }
+
+    // If the face is on a processor patch, we need to get the cell
+    // stencils from the face cells on the neighbouring processor
+
+    if (Pstream::parRun())
+    {
+        PstreamBuffers pBufs(UPstream::commsTypes::nonBlocking);
+
+        // Prepare and send stencils
+        forAll(mesh.boundaryMesh(), patchI)
+        {
+            const polyPatch& pp = mesh.boundaryMesh()[patchI];
+            if (isA<processorPolyPatch>(pp))
+            {
+                const processorPolyPatch& procPatch =
+                    refCast<const processorPolyPatch>(pp);
+
+                // Prepare cell stencils for sending
+                const labelUList& faceCells = pp.faceCells();
+                labelListList patchCellStencils(faceCells.size());
+                forAll(patchCellStencils, fI)
+                {
+                    const label localCellI = faceCells[fI];
+                    patchCellStencils[fI] = cellStencils[localCellI];
+                }
+
+                // Send to neighbour processor
+                UOPstream toNeighbProc(procPatch.neighbProcNo(), pBufs);
+                toNeighbProc
+                    << patchCellStencils;
+            }
+        }
+
+        pBufs.finishedSends(); // no-op for blocking
+
+        // Receive data
+        forAll(mesh.boundaryMesh(), patchI)
+        {
+            const polyPatch& pp = mesh.boundaryMesh()[patchI];
+            if (isA<processorPolyPatch>(pp))
+            {
+                const processorPolyPatch& procPatch =
+                    refCast<const processorPolyPatch>(pp);
+
+                UIPstream fromNeighbProc(procPatch.neighbProcNo(), pBufs);
+
+                labelListList receiveData(pp.size());
+
+                fromNeighbProc
+                    >> receiveData;
+
+                // Merge stencils with local face cells
+                const labelUList& faceCells = pp.faceCells();
[…]
