### 2024-11-28 8012962
Added integration rules to triangleQuadrature
