+++ src/solids4FoamModels/numerics/triangleQuadrature/triangleQuadrature.C
@@ -24,27 +24,204 @@ License
 namespace Foam
 {
 
-const List<Tuple2<label, Tuple2<vectorList, scalarList>>> TRI_QUADRATURE
-(
-);
+// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
 
+Map<triangleQuadrature::quadratureRule> triangleQuadrature::rules_;
+
+// Initialize and return quadrature rules
+void triangleQuadrature::constructRules()
+{
+    if (!rules_.empty())
+    {
+        FatalErrorInFunction
+            << "attempt to re-construct rules when they already exist"
+            << exit(FatalError);
+    }
+
+    // 1 point quadrature, exact for polynomials up to 1 order
+    rules_.insert(1, quadratureRule{
+        List<point>{vector(1.0/3.0, 1.0/3.0, 1.0/3.0)},
+        List<scalar>{1.0}
+    });
+
+    // 3 point quadrature, exact for polynomials up to 2 order
+    rules_.insert(3, quadratureRule{
+        List<point>
+        {
+            point(2.0/3.0, 1.0/6.0, 1.0/6.0),
+            point(1.0/6.0, 2.0/3.0, 1.0/6.0),
+            point(1.0/6.0, 1.0/6.0, 2.0/3.0)
+        },
+        List<scalar>{1.0/3.0, 1.0/3.0, 1.0/3.0}
+    });
+
+    // 4 point quadrature, exact for polynomials up to 3 order
+    rules_.insert(4, quadratureRule{
+        List<point>
+        {
+            point(1.0/3.0, 1.0/3.0, 1.0/3.0),
+            point(0.6, 0.2, 0.2),
+            point(0.2, 0.6, 0.2),
+            point(0.2, 0.2, 0.6)
+        },
+        List<scalar>{-9.0/16.0, 25.0/48.0, 25.0/48.0, 25.0/48.0}
+    });
+
+    // 6 point quadrature, exact for polynomials up to 4 order
+    rules_.insert(6, quadratureRule{
+        List<point>
+        {
+            point(0.108103018168070, 0.445948490915965, 0.445948490915965),
+            point(0.445948490915965, 0.108103018168070, 0.445948490915965),
+            point(0.445948490915965, 0.445948490915965, 0.108103018168070),
+            point(0.816847572980459, 0.091576213509771, 0.091576213509771),
+            point(0.091576213509771, 0.816847572980459, 0.091576213509771),
+            point(0.091576213509771, 0.091576213509771, 0.816847572980459)
+        },
+        List<scalar>
+        {
+            0.223381589678011,
+            0.223381589678011,
+            0.223381589678011,
+            0.109951743655322,
+            0.109951743655322,
+            0.109951743655322
+        }
+    });
+
+    // 7 point quadrature, exact for polynomials up to 5 order
+    rules_.insert(7, quadratureRule{
+        List<point>
+        {
+            point(1.0/3.0, 1.0/3.0, 1.0/3.0),
+            point(0.059715871789770, 0.470142064105115, 0.470142064105115),
+            point(0.470142064105115, 0.059715871789770, 0.470142064105115),
+            point(0.470142064105115, 0.470142064105115, 0.059715871789770),
+            point(0.797426985353087, 0.101286507323456, 0.101286507323456),
+            point(0.101286507323456, 0.797426985353087, 0.101286507323456),
+            point(0.101286507323456, 0.101286507323456, 0.797426985353087),
+        },
+        List<scalar>
+        {
+             0.225000000000000,
+             0.132394152788506,
+             0.132394152788506,
+             0.132394152788506,
+             0.125939180544827,
+             0.125939180544827,
+             0.125939180544827
+        }
+    });
+
[…]
