+++ src/solids4FoamModels/solidModels/solidModel/solidModel.H
@@ -316,27 +316,6 @@ protected:
             const bool writeResiduals = true
         );
 
-        //- Check if a vol field  has converged
-        template<class Type>
-        bool fieldConverged
-        (
-            const int iCorr,
-            const dimensionedScalar pDeltaT,
-            const GeometricField<Type, fvPatchField, volMesh>& vf,
-            const bool writeResiduals = true
-        );
-
-        //- Check if a vol field  has converged
-        template<class Type>
-        bool fieldConverged
-        (
-            const int iCorr,
-            const dimensionedScalar pDeltaT,
-            const GeometricField<Type, pointPatchField, pointMesh>& vf,
-            const bool writeResiduals = true
-        );
-
-
         //- Return solution standard tolerance
         scalar solutionTol() const
         {
