### 2024-11-28 02fef5a
Added barycentric to global function
