### 2024-12-03 d53e1fe
Added calcGlobalQRFaceGPCoeffs, calcGaussPointsAndWeights and fGradGaussPoints
