+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -1566,6 +1566,302 @@ void higherOrderGrad::calcGlobalQRFaceCoeffs() const
 }
 
 
+void higherOrderGrad::calcGlobalQRFaceGPCoeffs() const
+{
+    //if (debug)
+    {
+        InfoInFunction
+            << "start" << endl;
+    }
+
+    if (QRGradFaceGPCoeffsPtr_)
+    {
+        FatalErrorInFunction
+            << "Pointer already set!" << abort(FatalError);
+    }
+
+    const fvMesh& mesh = mesh_;
+
+    QRGradFaceGPCoeffsPtr_.set(new List<List<DynamicList<vector>>>(mesh.nFaces()));
+    List<List<DynamicList<vector>>>& QRGradCoeffs = *QRGradFaceGPCoeffsPtr_;
+
+    forAll(QRGradCoeffs, faceI)
+    {
+        List<DynamicList<vector>>& facePointsGC = QRGradCoeffs[faceI];
+        facePointsGC.setSize(mesh.faces()[faceI].size() * triQuadraturePtsNb_);
+    }
+
+    // Refernces for brevity and efficiency
+    const vectorField& CI = mesh.C();
+    const vectorField& CfI = mesh.Cf();
+
+    // Collect CI for off-processor cells in the stencils
+    Map<vector> globalCI;
+    requestGlobalStencilData(CI, globalCI);
+
+    // Calculate Taylor series exponents
+    // 1 for zero order, 4 for 1 order, 10 for second order, etc.
+    DynamicList<FixedList<label, 3>> exponents;
+    generateExponents(N_, exponents);
+    const label Np = exponents.size();
+    if (debug)
+    {
+        Info<< "Np = " << Np << endl;
+    }
+
+    // Precompute factorials up to N
+    List<scalar> factorials(N_ + 1, 1.0);
+    for (label n = 1; n <= N_; ++n)
+    {
+        factorials[n] = factorials[n - 1]*n;
+    }
+
+    // Coefficients
+    List<List<DynamicList<scalar>>> c(mesh.nFaces());
+    List<List<DynamicList<scalar>>> cx(mesh.nFaces());
+    List<List<DynamicList<scalar>>> cy(mesh.nFaces());
+    List<List<DynamicList<scalar>>> cz(mesh.nFaces());
+
+    // Set size for second list
+    forAll(c, faceI)
+    {
+        c[faceI].setSize(mesh.faces()[faceI].size() * triQuadraturePtsNb_);
+        cx[faceI].setSize(mesh.faces()[faceI].size() * triQuadraturePtsNb_);
+        cy[faceI].setSize(mesh.faces()[faceI].size() * triQuadraturePtsNb_);
+        cz[faceI].setSize(mesh.faces()[faceI].size() * triQuadraturePtsNb_);
+    }
+
+    const List<labelList>& stencils = globalFaceStencils();
+
+    // Gauss point locations on each face
+    const List<List<point>>& faceGP = faceGaussPoints();
+
+    // Gauss point weights
+    const List<List<scalar>>& faceGPW = faceGaussPointsWeight();
+
+    forAll(stencils, faceI)
+    {
+        const labelList& curStencil = stencils[faceI];
+
+        // Centre of current face
+        const vector& curCf = CfI[faceI];
+
+        // Find max distance in this stencil
+        scalar maxDist = 0.0;
+        forAll(curStencil, cI)
+        {
+            const label neiGlobalCellID = curStencil[cI];
+
+            scalar d;
+            if (globalCells_.isLocal(neiGlobalCellID))
+            {
+                const label neiLocalCellID =
+                    globalCells_.toLocal(neiGlobalCellID);
[…]
