### 2024-12-03 454209a
triangleQuadrature: style changes
