+++ src/solids4FoamModels/numerics/triangleQuadrature/triangleQuadrature.C
@@ -39,113 +39,143 @@ void triangleQuadrature::constructRules()
     }
 
     // 1 point quadrature, exact for polynomials up to 1 order
-    rules_.insert(1, quadratureRule{
-        List<point>{vector(1.0/3.0, 1.0/3.0, 1.0/3.0)},
-        List<scalar>{1.0}
-    });
+    rules_.insert
+    (
+        1,
+        quadratureRule
+        {
+            List<point>{vector(1.0/3.0, 1.0/3.0, 1.0/3.0)},
+            List<scalar>{1.0}
+        }
+    );
 
     // 3 point quadrature, exact for polynomials up to 2 order
-    rules_.insert(3, quadratureRule{
-        List<point>
+    rules_.insert
+    (
+        3,
+        quadratureRule
         {
-            point(2.0/3.0, 1.0/6.0, 1.0/6.0),
-            point(1.0/6.0, 2.0/3.0, 1.0/6.0),
-            point(1.0/6.0, 1.0/6.0, 2.0/3.0)
-        },
-        List<scalar>{1.0/3.0, 1.0/3.0, 1.0/3.0}
-    });
+            List<point>
+            {
+                point(2.0/3.0, 1.0/6.0, 1.0/6.0),
+                point(1.0/6.0, 2.0/3.0, 1.0/6.0),
+                point(1.0/6.0, 1.0/6.0, 2.0/3.0)
+            },
+            List<scalar>{1.0/3.0, 1.0/3.0, 1.0/3.0}
+        }
+    );
 
     // 4 point quadrature, exact for polynomials up to 3 order
-    rules_.insert(4, quadratureRule{
-        List<point>
+    rules_.insert
+    (
+        4,
+        quadratureRule
         {
-            point(1.0/3.0, 1.0/3.0, 1.0/3.0),
-            point(0.6, 0.2, 0.2),
-            point(0.2, 0.6, 0.2),
-            point(0.2, 0.2, 0.6)
-        },
-        List<scalar>{-9.0/16.0, 25.0/48.0, 25.0/48.0, 25.0/48.0}
-    });
+            List<point>
+            {
+                point(1.0/3.0, 1.0/3.0, 1.0/3.0),
+                point(0.6, 0.2, 0.2),
+                point(0.2, 0.6, 0.2),
+                point(0.2, 0.2, 0.6)
+            },
+            List<scalar>{-9.0/16.0, 25.0/48.0, 25.0/48.0, 25.0/48.0}
+        }
+    );
 
     // 6 point quadrature, exact for polynomials up to 4 order
-    rules_.insert(6, quadratureRule{
-        List<point>
+    rules_.insert
+    (
+        6,
+        quadratureRule
         {
-            point(0.108103018168070, 0.445948490915965, 0.445948490915965),
-            point(0.445948490915965, 0.108103018168070, 0.445948490915965),
-            point(0.445948490915965, 0.445948490915965, 0.108103018168070),
-            point(0.816847572980459, 0.091576213509771, 0.091576213509771),
-            point(0.091576213509771, 0.816847572980459, 0.091576213509771),
-            point(0.091576213509771, 0.091576213509771, 0.816847572980459)
-        },
-        List<scalar>
-        {
-            0.223381589678011,
-            0.223381589678011,
-            0.223381589678011,
-            0.109951743655322,
-            0.109951743655322,
-            0.109951743655322
+            List<point>
+            {
+                point(0.108103018168070, 0.445948490915965, 0.445948490915965),
+                point(0.445948490915965, 0.108103018168070, 0.445948490915965),
+                point(0.445948490915965, 0.445948490915965, 0.108103018168070),
+                point(0.816847572980459, 0.091576213509771, 0.091576213509771),
[…]
