### 2024-12-04 dbc0c87
Added tmp for barycentricToPoint func to avoid data copy
