### 2024-12-04 93f0bd4
Minor style changes and use references where appropriate
