### 2024-12-04 f974c5c
Style update in sphericalCavityAnalyticalSolution
