### 2024-12-04 ccff271
Updates for snes-ho feature (#142)
* fGradGaussPoints func development

* Added integration rules to triangleQuadrature

* Added barycentric to global function

* Added calcGlobalQRFaceGPCoeffs, calcGaussPointsAndWeights and fGradGaussPoints

* triangleQuadrature: style changes

* Added tmp for barycentricToPoint func to avoid data copy

* Minor style changes and use references where appropriate

* Style update in sphericalCavityAnalyticalSolution

---------

Co-authored-by: Philip Cardiff <philipcardiff@gmail.com>
