+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.H
@@ -22,8 +22,6 @@ Description
     Class for calculating higher order cell-centre gradients using the least
     squares method.
 
-    Currently, linear and quadratic polynomials are implemented.
-
     Cells are added to the stencil in layers.
 
 Author
@@ -41,6 +39,7 @@ SourceFiles
 #include "fvMesh.H"
 #include "globalIndex.H"
 #include <Eigen/Dense>
+#include "tensorList.H"
 
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
@@ -70,6 +69,9 @@ class higherOrderGrad
 
         //- Kernel shape parameter
         const scalar k_;
+
+        //- Number of quadrature points per face triangle
+        const label triQuadraturePtsNb_;
 
         //- Maxium stencil size: used for initialising dynamic lists
         const label maxStencilSize_;
@@ -125,6 +127,9 @@ class higherOrderGrad
         //- QR Gradient face coefficients
         mutable autoPtr<List<DynamicList<vector>>> QRGradFaceCoeffsPtr_;
 
+        //- QR Gradient face Gauss points coefficients
+        mutable autoPtr<List<List<DynamicList<vector>>>> QRGradFaceGPCoeffsPtr_;
+
         //- Cholesky decomposition
         mutable autoPtr<List<Eigen::LLT<Eigen::MatrixXd>>> choleskyPtr_;
 
@@ -137,6 +142,11 @@ class higherOrderGrad
             List<Eigen::DiagonalMatrix<double, Eigen::Dynamic>>
         > sqrtWPtr_;
 
+        //- Face Gauss points
+        mutable autoPtr<List<List<point>>> faceGaussPointsPtr_;
+
+        //- Face Gauss points weights
+        mutable autoPtr<List<List<scalar>>> faceGaussPointsWeightPtr_;
 
     // Private Member Functions
 
@@ -173,8 +183,12 @@ class higherOrderGrad
         void calcGlobalQRCoeffs() const;
 
         //- Calculate coefficients for the QR decomposition approach using the
-        //  the global cell stencils for the faces
+        //  global cell stencils for the faces
         void calcGlobalQRFaceCoeffs() const;
+
+        //- Calculate coefficients for the QR decomposition approach at face
+        //  Gauss points using the global cell stencils for the faces
+        void calcGlobalQRFaceGPCoeffs() const;
 
         //- Calculate coefficients for the Cholesky decomposition approach
         void calcCholeskyCoeffs() const;
@@ -182,6 +196,9 @@ class higherOrderGrad
         //- Calculate coefficients for the Cholesky decomposition approach using
         //  the global cell stencils
         void calcGlobalCholeskyCoeffs() const;
+
+        //- Calculate Gauss points and corresponding weights
+        void calcGaussPointsAndWeights() const;
 
         //- Return the interpolation coefficients
         const List<DynamicList<scalar>>& QRInterpCoeffs() const;
@@ -191,6 +208,9 @@ class higherOrderGrad
 
         //- Return the gradient face coefficients
         const List<DynamicList<vector>>& QRGradFaceCoeffs() const;
+
+        //- Return the gradient face Gauss points coefficients
+        const List<List<DynamicList<vector>>>& QRGradFaceGPCoeffs() const;
 
         //- Return the cell boundary face indices
         const List<DynamicList<label>>& stencilsBoundaryFaces() const;
@@ -205,6 +225,12 @@ class higherOrderGrad
         //  approach
         const List<Eigen::DiagonalMatrix<double, Eigen::Dynamic>>&
             sqrtW() const;
+
+        //- Return the face Gauss points
+        const List<List<point>>& faceGaussPoints() const;
+
+        //- Return the face Gauss points weights
[…]
