### 2024-12-12 3275e76
Fix a typo in the preCICE tutorial documentation
