### 2024-12-12 6ccf160
Fix a typo in the documentation page of optional fixes
