### 2024-12-13 solids4foam/solids4foam#149 @MakisH
Undocumented OpenFOAM version compatibility for tutorial cases
Trying to run the [beamInCrossFlow](https://www.solids4foam.com/tutorials/tutorial3.html) tutorial with OpenFOAM v2312, I got the following output:

```
tutorials/fluidSolidInteraction/beamInCrossFlow [(v2.1)]$ ./Allrun
No OpenFOAM tutorials? : /usr/lib/openfoam/openfoam2312/tutorials
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
| solids4Foam::convertCaseFormat start                                |
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Changing symmetryPlane to symmetry in blockMeshDict

Replacing system/functions with system/functions.openfoam
Changing RASModel to RAS in turbulenceProperties
Changing timeVaryingUniformFixedValue to uniformValue in p
OpenFOAM.com specific: replacing 'leastSquares' with 'pointCellsLeastSquares' in system/solid/fvSchemes
Modifying force.gnuplot in consistent with ESI version

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
| solids4Foam::convertCaseFormat end                                  |
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This case currently does not run with this OpenFOAM verion
```

Searching for the error message, I understand that this appears in:

- [ ] `fluidSolidInteraction/beamInCrossFlow/Allrun`
- [ ] `fluidSolidInteraction/flexibleDamBreak/Allrun`

For both, the check seems to be `if (( "${WM_PROJECT_VERSION#v}" > 2012 ))`. This seems to not be documented in the `README.md` of the respective tutorials, and the error message does not give any hints on which versions would be valid.

In the documentation, where this is explained, I would also expect some details on what exactly is incompatible.

This is in the context of my review for the JOSS publication (https://github.com/openjournals/joss-reviews/issues/7407).
