### 2024-12-13 solids4foam/solids4foam#150 @MakisH
Undocumented Allrun and post-processing of 3dTube tutorial
The [3dTube](https://www.solids4foam.com/tutorials/tutorial4.html) tutorial shows very interesting results, with very nicely designed plots. There is also a link to a [data repository](https://github.com/solids4foam/solids4foam-tutorials-benchmark-data). However, after a first read and execution of the tutorial, I have some questions that I could not answer (at least not without diving deeper into settings):

1. In the `tutorials/fluidSolidInteraction/3dTube [(v2.1)]` path, I only see an `Allrun` script. In the documentation, additional scripts are mentioned (`Allrun.pimpleFluid` and `Allrun.sonicLiquidFluid`), which I don't see. I guess there is some discrepancy between the benchmark and the tutorial case, which the documentation does not cover.
2. I ran the tutorial using simple `Allrun`. Which results are then generated? This would be helpful to add in the `Running the case` section.
3. After I run the case, what kind of scripts can I use (and how) to generate these plots, or a subset that will give me an indication whether everything went fine? I copied and modified [this gnuplot script](https://github.com/solids4foam/solids4foam-tutorials-benchmark-data/blob/main/tutorials/fluidSolidInteraction/3dTube/radialDisplacementDeltaT1e-4.gnuplot), but there was a lot of guessing involved.

Current state:

![Screenshot from 2024-12-13 18-09-46](https://github.com/user-attachments/assets/e75c996a-21d4-42cb-a089-76659a36c2ca)

See also a separate issue about the metadata in the benchmarks repository: https://github.com/solids4foam/solids4foam-tutorials-benchmark-data/issues/2

This is in the context of my review for the JOSS publication (https://github.com/openjournals/joss-reviews/issues/7407).
