### 2024-12-13 #7407 @MakisH
@kyleniemeyer I am through with my first iteration, and I have opened some issues and PRs to discuss individual aspects. I am happy to try more aspects, as needed.

@philipcardiff @iBatistic Let me know once you are ready for a second iteration.
