### 2024-12-13 3c68264
Added CONTRIBUTE section
In issue #145, it was pointed out that the project was missing a "Contributing" section. I've drafted an initial version of it. @philipcardiff, feel free to modify it or remove it if you do not want to include it.
