### 2024-12-16 ede412d
Update preCICE configuration files to v3 (#146)
* Update preCICE configuration files to v3

* Fix a typo in the preCICE tutorial documentation
