### 2024-12-16 470b144
Solution for #122 (#138)
* Make denom always displacement increment in solidModel::converged

* Make denom always displacement increment in solidModel::converged

* Minor style edits

* Fixed correction on symmetry planes

* Removed stressName from solidTorque

* Solution for #122

* Fix for #140

* Scaling fix for linear elastic Cooks membrane

---------

Co-authored-by: Philip Cardiff <philip.cardiff@ucd.ie>
