### 2024-12-16 solids4foam/solids4foam#150 @philipcardiff
1. The README.md file is outdated here; to-do: update it to refer to the new procedure for calling `Allrun`, e.g. `./Allrun sonicLiquidFluid`.
2. See point 1. By default, `Allrun` uses the `pimpleFluid` approach; to-do: update the README.md.
3. Add gnuplot scripts to plot: (i) axial displacement vs time, (ii) radial displacement vs time, and optionally (iii) number of FSI iterations per time.
