### 2024-12-16 c66a480
v2406: update Allrun in beamInCrossFlow and flexibleDamBreak
