### 2024-12-16 0f20d31
Added CONTRIBUTING.md file
