### 2024-12-17 d7608d7
CONTRIBUTING.md: address comments at #152
