### 2024-12-17 bd6e54f
CONTRIBUTING.md: update questions section about forums #152
