### 2024-12-17 ee10976
CONTRIBUTING.md: update questions section (fix duplicate heading) #152
