### 2024-12-17 283e58c
Added CONTRIBUTING.md file (#152)
* Make denom always displacement increment in solidModel::converged

* Make denom always displacement increment in solidModel::converged

* Minor style edits

* Fixed correction on symmetry planes

* Removed stressName from solidTorque

* Solution for #122

* Fix for #140

* Scaling fix for linear elastic Cooks membrane

* Added CONTRIBUTING.md file

* CONTRIBUTING.md: address comments at #152

* CONTRIBUTING.md: update questions section about forums #152

* CONTRIBUTING.md: update questions section (fix duplicate heading) #152

---------

Co-authored-by: Phil
[…]
