### 2024-12-17 ee0c04d
buildAndTest.yml: update FE41 image to resolve GLIBC GitHub action error
