+++ .github/pull_request_template.md
@@ -1,0 +1,70 @@
+## Pull Request Description 🚀
+
+Please provide a **concise summary** of your changes and the motivation behind them. For example:
+- What issue or feature does this pull request address?
+- Why are these changes necessary?
+
+---
+
+## Related Issues and Discussions 🧩
+
+- Link to any related **issues** (e.g., #123) or **discussions**.
+- If this PR fixes a bug, include "Closes #issue-number" to automatically close it upon merging.
+
+---
+
+## Changes Made 🛠️
+
+- [ ] **Bug Fix**: Describe the bug and how it was fixed.
+- [ ] **New Feature**: Briefly explain the new functionality.
+- [ ] **Refactoring**: Details of any code improvements.
+- [ ] **Documentation**: Updates to docs or comments.
+- [ ] **Other**: Explain any additional changes.
+
+### Summary of Changes:
+- **Files Modified**: List the key files affected (e.g., `solidsModel.C`, `solverFoam.H`).
+- **Behavioural Changes**: Are there any changes in output or performance?
+
+---
+
+## Checklist ✅
+
+Please ensure the following before submitting your PR:
+
+- [ ] Code compiles successfully with **OpenFOAM** or **foam-extend** versions (e.g., `v2312`, `4.1`).
+- [ ] I tested the changes with the provided tutorial or relevant test cases.
+- [ ] Added relevant **comments** or **documentation** for clarity.
+- [ ] The PR passes existing **GitHub Actions CI checks** (e.g., build and test).
+- [ ] Code follows the project's **coding style** and conventions.
+
+---
+
+## Test Cases and Results 📊
+
+Provide details of any test cases run and their results:
+- **Case Name**: e.g., `plateUnderCompression`, `beamBendingFSI`
+- **Results**: Brief description (e.g., "Results match the reference solution within tolerance").
+
+If applicable, add a **screenshot** or **plot** of the results for visual confirmation.
+
+---
+
+## Additional Notes 💡
+
+- Highlight any potential **backward compatibility** issues.
+- Mention if there are **dependencies** (e.g., specific OpenFOAM versions, external libraries).
+- Any known limitations or remaining tasks for this PR.
+
+---
+
+## Screenshots (if applicable) 📸
+
+Include relevant screenshots or plots for visualisation.
+
+---
+
+### Final Comments 💬
+
+Thank you for contributing to **solids4foam**! 🎉 
+
+If this is your first time contributing, check out the [CONTRIBUTING.md](link-to-contributing-guide) for more details.
