+++ .github/workflows/markdownlint.yml
@@ -3,9 +3,9 @@ name: Markdown Lint
 
 on:
   push:
-    branches: [ "master" "development" ]
+    branches: [ "master", "development" ]
   pull_request:
-    branches: [ "master" "development" ]
+    branches: [ "master", "development" ]
   workflow_dispatch:
 
 jobs:
