+++ .github/workflows/markdownlint.yml
@@ -19,4 +19,5 @@ jobs:
       - name: Run MarkdownLint
         uses: DavidAnson/markdownlint-cli2-action@v18.0.0
         with:
-          globs: '**/*.md'
+          globs: "ThirdParty/**/*.md,optionalFixes/**/*.md,tutorials/**/*.md,applications/**/*.md,CONTRIBUTING.md/**/*.md,CONTRIBUTORS.md/**/*.md,README.md/**/*.md,etc/**/*.md,src/**/*.md"
+          separator: ','
