### 2024-12-17 16f0235
Update markdown files: lint compliance
