+++ tutorials/solids/linearElasticity/contactPatchTest/README.md
@@ -12,14 +12,27 @@ Prepared by Ivan Batistić
 
 ## Tutorial Aims
 
-- Demonstrate that a uniform pressure field can be transmitted across a non-conformal contact interface;
-- Examine the solver performance when solving contact problems in small deformation settings.
+- Demonstrate that a uniform pressure field can be transmitted across a
+  non-conformal contact interface;
+- Examine the solver performance when solving contact problems in small
+  deformation settings.
 
 ---
 
 ## Case Overview
 
-The contact patch test is designed to check can a uniform pressure field can be transmitted across a non-conformal contact interface. There are different versions of the patch test in the literature [[1](https://link.springer.com/book/10.1007/978-3-662-04864-1), [2](https://www.sciencedirect.com/science/article/abs/pii/S0045794901000487)]; here patch test proposed in [[3]](https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291097-0207%2820000530%2948%3A3%3C435%3A%3AAID-NME891%3E3.0.CO%3B2-V) is considered. The lower block is fixed at the bottom surface, whereas the upper block has a prescribed displacement $$\Delta = 0.01$$ m at its top surface (see Figure 1). Both blocks have the same dimensions, $$1$$ m width and $$0.5$$ m height, and the same material properties, Young’s modulus $$E = 1 × 10^6$$ Pa and zero Poisson’s ratio $$\nu = 0$$.  Gravitation effects are neglected, and there are no body forces. The problem is solved as static, using one loading increment. At the contact interface, frictionless contact is considered, i.e. $$\mu=0$$.
+The contact patch test is designed to check can a uniform pressure field can be
+transmitted across a non-conformal contact interface. There are different
+versions of the patch test in the literature
+[[1](https://link.springer.com/book/10.1007/978-3-662-04864-1), [2](https://www.sciencedirect.com/science/article/abs/pii/S0045794901000487)]; here patch test proposed in [[3]](https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291097-0207%2820000530%2948%3A3%3C435%3A%3AAID-NME891%3E3.0.CO%3B2-V)
+is considered. The lower block is fixed at the bottom surface, whereas the upper
+block has a prescribed displacement $$\Delta = 0.01$$ m at its top surface (see
+Figure 1). Both blocks have the same dimensions, $$1$$ m width and $$0.5$$ m
+height, and the same material properties, Young’s modulus $$E = 1 × 10^6$$ Pa
+and zero Poisson’s ratio $$\nu = 0$$. Gravitation effects are neglected, and
+there are no body forces. The problem is solved as static, using one loading
+increment. At the contact interface, frictionless contact is considered, i.e.
+$$\mu=0$$.
 
 <div style="text-align: center;">
   <img src="./images/contactPatchTest-geometry.png" alt="Image" width="400">
@@ -31,14 +44,22 @@ The contact patch test is designed to check can a uniform pressure field can be 
 
 ## Expected Results
 
-* Under plane strain conditions, the analytical solution to the problem is  [[1]](https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291097-0207%2820000530%2948%3A3%3C435%3A%3AAID-NME891%3E3.0.CO%3B2-V):
-$$
-  \sigma_{x} = \tau_{xy}=0\qquad \sigma_{y} = \dfrac{E}{1-\nu^2}\Delta \qquad \sigma_z = \nu \sigma_y.
-$$
+- Under plane strain conditions, the analytical solution to the problem is
[…]
