+++ tutorials/solids/beamsPlatesShells/squarePlate/README.md
@@ -12,7 +12,8 @@ Prepared by Philip Cardiff and Ivan Batistić
 
 ## Tutorial Aims
 
-- Demonstrate analysing a simple plate-bending problem using a finite area implementation of a Kirchoff-Love plate model.
+- Demonstrate analysing a simple plate-bending problem using a finite area
+  implementation of a Kirchoff-Love plate model.
 
 ---
 
@@ -24,15 +25,23 @@ This case currently only runs with foam extend as it uses the foam extend finite
 
 Two types of beam/plate/shell theory are widely used:
 
-- **Euler–Bernoulli** beam theory, corresponding to **Kirchhoff–Love** shell theory
-  - The planes normal to the midline are assumed to remain plane and normal (no shear stress); this is also called engineering
-    beam theory.
+- **Euler–Bernoulli** beam theory, corresponding to **Kirchhoff–Love** shell
+  theory
+
+  - The planes normal to the midline are assumed to remain plane and normal (no
+    shear stress); this is also called engineering beam theory.
 
 - **Timoshenko beam** theory, corresponding to **Mindlin–Reissner** shell theory
-  - The planes normal to the midline are assumed to remain plane but not necessarily normal (shear stress may be non-zero); this
-    is also called shear beam theory.
+  - The planes normal to the midline are assumed to remain plane but not
+    necessarily normal (shear stress may be non-zero); this is also called shear
+    beam theory.
 
-Kirchhoff–Love shell theory is a subset of Mindlin–Reissner shell theory, i.e. Mindlin–Reissner is applicable in every case that Kirchhoff–Love shell theory is applicable, but Kirchhoff–Love theory is not applicable in every case that Mindlin–Reissner theory is applicable. The following table demonstrates the valid length $$L$$ to thickness $$h$$ ranges for Kirchhoff and Mindlin approaches relative to a 3-D continuum approach.
+Kirchhoff–Love shell theory is a subset of Mindlin–Reissner shell theory, i.e.
+Mindlin–Reissner is applicable in every case that Kirchhoff–Love shell theory is
+applicable, but Kirchhoff–Love theory is not applicable in every case that
+Mindlin–Reissner theory is applicable. The following table demonstrates the
+valid length $$L$$ to thickness $$h$$ ranges for Kirchhoff and Mindlin
+approaches relative to a 3-D continuum approach.
 
 |                                   | Kirchhoff | Mindlin | 3-D continuum |
 | --------------------------------- | :-------: | :-----: | :-----------: |
@@ -40,25 +49,47 @@ Kirchhoff–Love shell theory is a subset of Mindlin–Reissner shell theory, i.
 | **Moderately thick**: $$L/h > 5$$ |     x     |    ✓    |       ✓       |
 | **Thick**: $$L/h < 5$$            |     x     |    x    |       ✓       |
 
-**Kirchhoff-Love Plate Formulation**   
+**Kirchhoff-Love Plate Formulation**
 
-For thin plates $$(L/h > 10)$$, Kirchhoff–Love shell theory allows the conservation of momentum to be reformulated into a fourth-order partial differential equation known as the [biharmonic equation](https://en.wikipedia.org/wiki/Biharmonic_equation), where the unknown scalar, $$w$$, is the transverse displacement (displacement normal to the plate):
+For thin plates $$(L/h > 10)$$, Kirchhoff–Love shell theory allows the
+conservation of momentum to be reformulated into a fourth-order partial
+differential equation known as the
+[biharmonic equation](https://en.wikipedia.org/wiki/Biharmonic_equation), where
+the unknown scalar, $$w$$, is the transverse displacement (displacement normal
+to the plate):
+
 $$
 \rho h \frac{\partial^2 w}{\partial t^2} = -D \nabla^2 \nabla^2 w + p,
 $$
[…]
