+++ CONTRIBUTORS.md
@@ -2,33 +2,33 @@
 sort: 2
 ---
 
-This list has been generated using `git shortlog -s --all`.
-If your name is missing, please let philip.cardiff@ucd.ie know.
+This list has been generated using `git shortlog -s --all`. If your name is
+missing, please let <philip.cardiff@ucd.ie> know.
 
-* Philip Cardiff
-* Zeljko Tukovic
-* Danial Khazaei
-* Denis Maier
-* Iago Lessa de Oliveira
-* Jozsef Nagy
-* Jeff Heylmun
-* Saber Mohammadi
-* Bruno Santos
-* Cyrille Bonamy
-* Jeffrey Heylmun
-* Scott Levie
-* Sindeev Sergey
-* Andrew Whelan
-* Emad Tandis
-* Karen FitzGerald
-* Philipp Milovic
-* Seevani Bali
-* Ivan Batistic
-* Simona Moretti
-* Andrea Luigi Facci
-* Federico Mazzanti
-* Xiaohu Guo
-* Amirhossein Taran
-* Mark Olesen
-* Sandro Manservisi
-* Greg Burgreen
+- Philip Cardiff
+- Zeljko Tukovic
+- Danial Khazaei
+- Denis Maier
+- Iago Lessa de Oliveira
+- Jozsef Nagy
+- Jeff Heylmun
+- Saber Mohammadi
+- Bruno Santos
+- Cyrille Bonamy
+- Jeffrey Heylmun
+- Scott Levie
+- Sindeev Sergey
+- Andrew Whelan
+- Emad Tandis
+- Karen FitzGerald
+- Philipp Milovic
+- Seevani Bali
+- Ivan Batistic
+- Simona Moretti
+- Andrea Luigi Facci
+- Federico Mazzanti
+- Xiaohu Guo
+- Amirhossein Taran
+- Mark Olesen
+- Sandro Manservisi
+- Greg Burgreen

[…]
