### 2024-12-17 6f70e2b
Update markdown files: lint compliance
