+++ src/solids4FoamModels/functionObjects/README.md
@@ -41,13 +41,16 @@ Prepared by Ivan Batistić
 
 ## `cantileverAnalyticalSolution`
 
-- **Function object purpose**  
+- **Function object purpose**
   To generate the analytical solution fields for the bending slender cantilever
   problem. The analytical solution is taken from
-  [[C.E. Augarde, A.J. Deeks, The use of Timoshenko’s exact solution for a  cantilever beam in adaptive analysis. Finite Elements in Analysis and Design, 44, 2008]](https://www.researchgate.net/publication/30053820_The_use_of_Timoshenko's_exact_solution_for_a_cantilever_beam_in_adaptive_analysis):
+  [[C.E. Augarde, A.J. Deeks, The use of Timoshenko’s exact solution for a
+  cantilever beam in adaptive analysis. Finite Elements in Analysis and Design,
+  44, 2008]](https://www.researchgate.net/publication/30053820_The_use_of_Timoshenko's_exact_solution_for_a_cantilever_beam_in_adaptive_analysis):
 
 $$
-\sigma_{xx} = \frac{P(L-x)y}{I}, \qquad \sigma_{yy}=0, \qquad \sigma_{xy}=-\frac{P}{2I}\left( \frac{D^2}{4}-y^2\right),
+\sigma_{xx} = \frac{P(L-x)y}{I}, \qquad \sigma_{yy}=0,
+\qquad \sigma_{xy}=-\frac{P}{2I}\left( \frac{D^2}{4}-y^2\right),
 $$
 
 $$
@@ -62,11 +65,15 @@ $$
 of are of the cross-section, $P$ is applied load and $L$ is length of the beam.
 
 ```note
-Above analytical solution can be used only if the shearing forces on the ends are distributed according to the same parabolic law as the shearing stress $\tau_{xy}$ and the intensity of the normal forces at the built-in end is proportional to $y$.
+Above analytical solution can be used only if the shearing forces on the ends
+are distributed according to the same parabolic law as the shearing stress
+$\tau_{xy}$ and the intensity of the normal forces at the built-in end is
+proportional to $y$.
 ```
 
 ```warning
-The current version of the code assumes a rectangular cross-section with unit width and automatically calculates the second moment of inertia!
+The current version of the code assumes a rectangular cross-section with unit
+width and automatically calculates the second moment of inertia!
 ```
 
 - **Example of usage**
@@ -122,31 +129,34 @@ The current version of the code assumes a rectangular cross-section with unit wi
     calculated one: `analyticalStress-sigma`;
   - `DDiference` field; difference between analytical displacement and
     calculated one: `analyticalD-D`.
-  - Log at the end of each time-step: `Component: 1`  
-    `Norms: mean L1, mean L2, LInfL: 0.12 0.2 0.5`  
+  - Log at the end of each time-step: `Component: 1`
+    `Norms: mean L1, mean L2, LInfL: 0.12 0.2 0.5`
     `...`
 
-- **Tutorial case in which it is used:** :  
+- **Tutorial case in which it is used:** :
[…]

[…]
