### 2024-12-17 8cff9bb
Add CITATION.cff file
