+++ CITATION.cff
@@ -1,0 +1,26 @@
+cff-version: 1.2.0
+message: "If you use solids4foam in your research, please cite it as below."
+authors:
+  - family-names: "Cardiff"
+    given-names: "Philip"
+    affiliation: "University College Dublin"
+    orcid: "https://orcid.org/0000-0002-4824-427X"
+  - family-names: "Tuković"
+    given-names: "Željko"
+    affiliation: "University of Zagreb"
+    orcid: "https://orcid.org/0000-0001-8719-0983"
+  - family-names: "Batistić"
+    given-names: "Ivan"
+    affiliation: "University of Zagreb"
+    orcid: "https://orcid.org/0000-0002-6104-0415"
+title: "solids4foam: An open-source toolbox for solid and fluid-solid interaction simulations"
+version: "2.1"
+doi: "10.5281/zenodo.12626724"
+date-released: "2024-07-02"
+license: "GPL-3.0-or-later"
+repository: "https://github.com/solids4foam/solids4foam"
+keywords:
+  - solid mechanics
+  - fluid-solid interaction
+  - OpenFOAM
+  - finite volume method
