+++ tutorials/fluidSolidInteraction/3dTube/README.md
@@ -23,8 +23,7 @@ You can find the files for this tutorial under
 This case consists of a pressure pulse applied in a thick-walled elastic tube
 (Figure 1).
 
-![](images/3dTube.png)
-
+![Figure 1: Wave propagation in an elastic pipe](images/3dTube.png)
 **Figure 1: Wave propagation in an elastic pipe**
 
 The fluid is assumed incompressible, Newtonian and isothermal, with a density of
@@ -49,14 +48,19 @@ FSI coupling:
   fluid velocity and Neumann conditions to the fluid pressure and solid
   displacement. This approach **does not** require modification of the
   underlying fluid solver; see
-  [Tuković Ž, Karač A, Cardiff P, Jasak H, Ivanković A, 2018, OpenFOAM finite volume solver for fluid–solid interaction. Trans FAMENA, 42(3):1–31.10.21278/TOF.42301](https://hrcak.srce.hr/206941);
+  [Tuković Ž, Karač A, Cardiff P, Jasak H, Ivanković A, 2018, OpenFOAM finite
+  volume solver for fluid–solid interaction. Trans FAMENA,
+  42(3):1–31.10.21278/TOF.42301](https://hrcak.srce.hr/206941);
 
 - **Robin-Neumann coupling**, where a Dirichlet condition is applied to the
   fluid velocity, a Robin condition to the fluid pressure, and a Neumann
   condition to the solid displacement. This approach **does** require
   modification of the underlying fluid solver, and hence it cannot be considered
   a _black-box_ coupling approach; see
-  [Tuković Ž, Bukač M, Cardiff P, Jasak H, Ivanković A, 2018, Added mass partitioned fluid–structure interaction solver based on a robin boundary condition for pressure. In: OpenFOAM selected papers of the 11th workshop. Springer, Berlin, pp 1–23](https://doi.org/10.1007/978-3-319-60846-4_1).
+  [Tuković Ž, Bukač M, Cardiff P, Jasak H, Ivanković A, 2018, Added mass
+  partitioned fluid–structure interaction solver based on a robin boundary
+  condition for pressure. In: OpenFOAM selected papers of the 11th workshop.
+  Springer, Berlin, pp 1–23](https://doi.org/10.1007/978-3-319-60846-4_1).
 
 For each of these two classes of approach, we can employ different methods to
 accelerate the FSI iteration loop convergence; within solids4foam, we can use:
@@ -64,13 +68,18 @@ accelerate the FSI iteration loop convergence; within solids4foam, we can use:
 - Aitken's dynamic relaxation;
 
 - the IQN-ILS algorithm of
-  [Joris Degroote, Robby Haelterman, Sebastiaan Annerel, Peter Bruggeman, Jan Vierendeels, Performance of partitioned procedures in fluid–structure interaction, Computers & Structures, 88, 7–8, 2010, 10.1016/j.compstruc.2009.12.006](https://www.sciencedirect.com/science/article/abs/pii/S0045794909003022).
+  [Joris Degroote, Robby Haelterman, Sebastiaan Annerel, Peter Bruggeman, Jan
+  Vierendeels, Performance of partitioned procedures in fluid–structure
+  interaction, Computers & Structures, 88, 7–8, 2010,
+  10.1016/j.compstruc.2009.12.006](https://www.sciencedirect.com/science/article/abs/pii/S0045794909003022).
 
 As well as using these acceleration algorithms, we can also use a **weakly
 compressible** fluid model rather than the standard **fully incompressible**
 model; for FSI, weakly compressible fluid models how been shown to improve
 convergence, for example, see
[…]
