### 2024-12-17 e2299b6
thermalCavity: readme lint updates
