### 2024-12-17 688346c
thermalCavity/README.md: lint
