+++ tutorials/thermoFluidSolidInteraction/hotTJunction/README.md
@@ -29,37 +29,53 @@ enforces temperature and heat flux continuity at the fluid-solid interface.
 Within the fluid, the temperature differences generate forces that drive fluid
 flow, while in the solid, increases in temperature cause volumetric expansion.
 
-<div style="text-align: center;">
-  <img src="./images/hotTJunction-geometry.png" alt="Image" width="600">
-    <figcaption>
-     <strong>Figure 1: Problem geometry and mesh. The fluid domain is shown in yellow, and the solid domain in blue.</strong>
-    </figcaption>
-</div>
-At time $$t = 0$$, the fluid and solid temperatures are $$280$$ K. The larger left fluid inlet has a prescribed constant velocity of $$(0.1\,0\,0)$$ m/s at $$280$$ K, whereas the smaller upper fluid inlet has a time-varying velocity which linearly ramps from $$(0\;0\;0)$$ at time = 0 s to $$(0\;-1\;0)$$ m/s at time = 0.1 s before linearly ramping back to $$(0\;0\;0)$$ at time = 0.2 s. The temperature at the smaller upper fluid inlet is temperature is $$320$$ K. A constant uniform pressure of 0 Pa is assumed at the outlet on the right. The smaller ends of the solid domain at the outlet and two inlets are fixed (zero displacement), held at a fixed temperature of $$280$$ K. The outer surface has a zero-traction condition and is assumed to be adiabatic (zero heat flux). The solid internal and remaining fluid external boundaries represent the fluid-solid interface, where kinematic (velocity continuity), kinetic (force continuity) and thermal constraints (temperature and heat flux continuity) are enforced. No slip is enforced on the fluid interface.
-In both fluid and solid, gravity is assumed to act in the negative vertical direction $$(0\;$$-$$9.81\; 0)$$ m/s$$^{2}$$ and inertial effects are included, albeit mechanical inertia is small. Small deformations are assumed in the solid. The end time is $$0.2$$ s and time-step size $$\Delta t = 1$$ ms. A uniform quadrilateral cell size of 0.1 mm is used in solid and fluid domains.
+![Figure 1: Problem geometry and mesh. The fluid domain is shown in yellow, and the solid domain in blue.](./images/hotTJunction-geometry.png)
+**Figure 1: Problem geometry and mesh. The fluid domain is shown in yellow, and
+the solid domain in blue.**
+
+At time $$t = 0$$, the fluid and solid temperatures are $$280$$ K. The larger
+left fluid inlet has a prescribed constant velocity of $$(0.1\,0\,0)$$ m/s at
+$$280$$ K, whereas the smaller upper fluid inlet has a time-varying velocity
+which linearly ramps from $$(0\;0\;0)$$ at time = 0 s to $$(0\;-1\;0)$$ m/s at
+time = 0.1 s before linearly ramping back to $$(0\;0\;0)$$ at time = 0.2 s. The
+temperature at the smaller upper fluid inlet is temperature is $$320$$ K. A
+constant uniform pressure of 0 Pa is assumed at the outlet on the right. The
+smaller ends of the solid domain at the outlet and two inlets are fixed (zero
+displacement), held at a fixed temperature of $$280$$ K. The outer surface has a
+zero-traction condition and is assumed to be adiabatic (zero heat flux). The
+solid internal and remaining fluid external boundaries represent the fluid-solid
+interface, where kinematic (velocity continuity), kinetic (force continuity) and
+thermal constraints (temperature and heat flux continuity) are enforced. No slip
+is enforced on the fluid interface.
+
[…]
