### 2024-12-17 89a4c28
Update markdown files: lint compliance
