+++ tutorials/solids/thermoelasticity/hotSphere/README.md
@@ -17,7 +17,8 @@ You can find the files for this tutorial under
 
 ## Case Overview
 
-![](images/sol_overview_1.PNG)
+![Case geometry and mesh](images/sol_overview_1.PNG)
+**Case geometry and mesh*
 
 This case analyses the stresses and displacements generated in a spherical
 pressure vessel subjected to an increasing internal pressure and temperature.
@@ -39,7 +40,17 @@ thermal gradient. The deformations (strains/rotations) are expected to be
 “small”: this means we can use a small strain (linear geometry) approach, where
 the displacements are assumed not to affect the material geometry. The governing
 equations are given by the conservation equations are linear momentum (linear
-geometry form) and energy (heat equation form): ![](images/sol_theory_1.PNG)
+geometry form) and energy (heat equation form):
+
+$$
+\rho \frac{\partial^2 \boldsymbol{u}}{\partial t^2} = \nabla \cdot
+\boldsymbol{\sigma} + \rho \boldsymbol{g}
+$$
+
+$$
+\rho C \frac{\partial T}{\partial t} = \nabla \cdot (k \nabla T) +
+\boldsymbol{\sigma} : \nabla \frac{\partial \boldsymbol{u}}{\partial t}
+$$
 
 In this case, the solver employs a segregated solution methodology, where a loop
 is performed over the momentum equation (solved for displacement `D`) and the
@@ -63,7 +74,9 @@ end
 At 5 s, the expected temperature distribution across the wall thickness is
 expected to be close to linear, where, the von Mises stress distribution is
 quite nonlinear with a minimum 4 mm from the inner wall.
-![](images/sol_expres_1.PNG)
+
+![Expected results](images/sol_expres_1.PNG)
+**Expected results**
 
 ---
 
@@ -97,7 +110,8 @@ seen, the mesh in the fluent format is converted to the OpenFOAM format before
 running the `solids4Foam` solver.
 
 ```tip
-A tutorial case can be cleaned and reset using the included `Allrun` script, i.e. `> ./Allclean`.
+A tutorial case can be cleaned and reset using the included `Allrun` script,
+i.e. `> ./Allclean`.
 ```
 
 ---
@@ -105,15 +119,18 @@ A tutorial case can be cleaned and reset using the included `Allrun` script, i.e
 ## Analysing the Results
 
 When viewing the results in ParaView, it can be insightful to warp the geometry
-by a scaled displacement field. This can be achieved using the _Warp By Vector_
-filter, where the `D` displacement field is selected as the _Vector_ and a
-_Scale Factor_ of 1 shows the true deformation. In this case, using a _Scale
-Factor_ of 100 allows the deformation to be seen. ![](images/sol_runcase_1.PNG)
-![](images/sol_runcase_2.PNG)
+by a scaled displacement field. This can be achieved using the *Warp By Vector*
+filter, where the `D` displacement field is selected as the *Vector* and a
+*Scale Factor* of 1 shows the true deformation. In this case, using a *Scale
+Factor* of 100 allows the deformation to be seen.
+
+![The case in ParaView 1](images/sol_runcase_1.PNG)
+![The case in ParaView 2](images/sol_runcase_2.PNG)
+**Visualising the case in ParaView**
 
 ---
 
-# Delving Deeper
+## Delving Deeper
 
 ---
 
[…]
