### 2024-12-17 a1c167b
cylinderInChannel/README.md: fix rendering
