### 2024-12-17 6c5032c
hotSphere: readme fix
