### 2024-12-17 eba0233
beamInCrossFlow: readme update
