### 2024-12-18 eef0212
Solution for #150 point 1.
