### 2024-12-18 69effeb
Solution for #150 point 2.
