### 2024-12-18 1b53983
Solution for #150 point 3.
