### 2024-12-18 solids4foam/solids4foam#149 @iBatistic
The check in beamInCrossFlow and flexibleDamBreak was removed in commit c66a480. It had already been removed in the feature branch we worked on recently, but we did not remove it from the development branch. I believe there was an issue with version v2312, but Philip resolved it. The valid versions are listed in the Allwmake script.
@philipcardiff, could you please confirm if I’m right? If so, we can close the issue.
