### 2024-12-18 5f90bba
fix lint issues in tutorials/solids
