+++ tutorials/solids/hyperelasticity/cylinderCrush/README.md
@@ -21,9 +21,9 @@ This tutorial currently only works with foam-extend-4.1.
 
 ## Case Overview
 
-<img src="images/cylinder-crushed.png" width="400" />
+![Cylinder crush geomtry](images/cylinder-crushed.png)
 
-**Figure 1: Solid cylinder compressed between two rigid plates**
+Figure 1 - Solid cylinder compressed between two rigid plates
 
 The third benchmark is a solid cylinder made of a rubbery material crushed
 between two rigid plates, as schematically shown in Figure 1. The cylinder has a
@@ -33,14 +33,20 @@ and plate surfaces was modelled as frictionless (see the `solidContact` boundary
 condition setup in the `0/DD` file).
 
 This problem was investigated by
-[[1] T. Sussman and K.-J. Bathe, “A finite element formulation for nonlinear incompresible elastic and inelastic analysis,” Computers and Structures, vol. 26, pp. 357–409, 1987.](https://www.sciencedirect.com/science/article/pii/0045794987902653),
+[[1] T. Sussman and K.-J. Bathe, “A finite element formulation for nonlinear
+incompresible elastic and inelastic analysis,” Computers and Structures, vol.
+26, pp. 357–409,
+1987.](https://www.sciencedirect.com/science/article/pii/0045794987902653),
 who assumed a nearly-incompressible behaviour of the material, and also by
-[I. Bijelonja, I. Demirdžić, and S. Muzaferija, “A finite volume method for large strain analysis of incompressible hyperelastic materials,” International Journal for Numerical Methods in Engineering, vol. 64, pp. 1594–1609, Nov. 2005](https://hrcak.srce.hr/206941),
+[I. Bijelonja, I. Demirdžić, and S. Muzaferija, “A finite volume method for
+large strain analysis of incompressible hyperelastic materials,” International
+Journal for Numerical Methods in Engineering, vol. 64, pp. 1594–1609, Nov.
+2005](https://hrcak.srce.hr/206941),
 who modelled the cylinder material as purely incompressible. In this tutorial,
 the same material law employed by both studies was used: the Mooney-Rivlin law
-with _c<sub>10</sub> = 0.293 MPa_, _c<sub>01</sub> = 0.177 MPa_, and the bulk
-modulus _1410 MPa_. Note that this material is much less stiff than the material
-assumed in the previous two tutorials.
+with $$c_{10} = 0.293$$ MPa, $$c_{01} = 0.177$$ MPa, and the bulk
+modulus $$1410$$ MPa. Note that this material is much less stiff than the
+material assumed in the previous two tutorials.
 
 The domain used was only a quarter of the cylinder due to symmetry (hatched area
 in Figure 1). A mesh with 500 quadrilateral control volumes arranged in a
@@ -59,29 +65,29 @@ Figure 3 shows the Von Mises stress field for the intermediate and final
 configuration of the cylinder, where we can see that the maximum stress occurs
 at the cylinder centre.
 
-<img src="images/mesh-1.png" width="300" />
-<img src="images/mesh-2.png" width="300" />
-<img src="images/mesh-3.png" width="300" />
+![Mesh](images/mesh-1.png)
+![Mesh](images/mesh-2.png)
+![Mesh](images/mesh-3.png)
 
-**Figure 2: Mesh of the undeformed configuration and the deformed mesh at an
-intermediate and final instant.**
+Figure 2 - Mesh of the undeformed configuration and the deformed mesh at an
+intermediate and final instant
 
-<img src="images/stress-2.png" width="400" />
-<img src="images/stress-3.png" width="400" />
+![Stress](images/stress-2.png)
+![Stress](images/stress-3.png)
 
[…]
