+++ .github/pull_request_template.md
@@ -1,3 +1,5 @@
+# `solids4foam` Pull Request
+
 ## Pull Request Description 🚀
 
 Please provide a **concise summary** of your changes and the motivation behind
