+++ README.md
@@ -1,24 +1,26 @@
-<img style="float: left;" src="solids4foamLogoName.png">
+# <!-- solids4foam -->
+
+![solids4foam logo](solids4foamLogoName.png)
 
 ---
 
-### What is solids4foam?
+## What is solids4foam?
 
 solids4foam is a toolbox for performing solid mechanics and fluid-solid
 interactions in OpenFOAM.
 
-### Where do I start?
+## Where do I start?
 
 Please visit the [solids4foam website](https://solids4foam.github.io).
 
-### Disclaimer
+## Disclaimer
 
 This offering is not approved or endorsed by OpenCFD Limited, producer and
 distributor of the OpenFOAM software via
 [www.openfoam.com](https://www.openfoam.com), and owner of the OPENFOAM® and
 OpenCFD® trade marks.
 
-### Acknowledgement
+## Acknowledgements
 
 OPENFOAM® is a registered trademark of OpenCFD Limited, producer and
 distributor of the OpenFOAM software via
