### 2024-12-18 54c9cde
utilities/README.md: fix formatting
