### 2024-12-18 57664d8
Fix for getNumberOfProcessors() & removing compression specifier warning (#153)
* Additional fix for #138

* Write compression from uncrompressed to off

* Solution for #150 point 1.

* Solution for #150 point 2.

* Solution for #150 point 3.

---------

Co-authored-by: Philip Cardiff <philip.cardiff@ucd.ie>
