### 2024-12-19 d6df0d8
Modifies license for all files
