+++ src/solids4FoamModels/solidModels/fvPatchFields/explicitSolidTractionLinearMomentum/explicitSolidTractionLinearMomentumFvPatchVectorField.H
@@ -19,12 +19,16 @@ Class
     explicitSolidTractionLinearMomentumFvPatchVectorField
 
 Description
-    Fixed traction boundary condition for solid solvers.
+    Fixed traction boundary condition for explicitGodunovCC solid solver
+    for linear momentum field.
 
-Author
+Original Authors
     Philip Cardiff, UCD.
     Zeljko Tukovic, FSB Zagreb.
-    khoder Alhamwi Alshaar, IITB.
+
+Modifications by
+    Khoder Alhamwi Alshaar, IITB - Modified to implement the explicit Godunov 
+    scheme for solid mechanics as part of the solids4foam project.
 
 SourceFiles
     explicitSolidTractionLinearMomentumFvPatchVectorField.C

[…]
