### 2024-12-19 solids4foam/solids4foam#149 @philipcardiff
Yep, this seems to be resolved.
