### 2024-12-19 solids4foam/solids4foam#150 @iBatistic
@philipcardiff, please check my reply in #153. I believe the requested changes for this issue have now been resolved.
