### 2024-12-20 b3c297e
3dTube/README.md: update Allrun instructions
