+++ applications/utilities/README.md
@@ -137,10 +137,11 @@ When using `addTinyPatch` the original mesh is overwritten!
   factor is a vector to allow different scalings in different directions; for
   example, for 2-D, the Z component should be set to 0.0.
 
-  If checkMesh fails after moving the points, the local motion is reduced by
-  the factor beta (defaults to 0.8) and the motion is performed again. The
-  maximum number of corrections iterations is set with maxIter (defaults to
-  1000).
+  If there are negative or small cell volumes after moving the points, the
+  local motion is reduced by the factor beta (defaults to 0.8) and the motion
+  is performed again. The maximum number of corrections iterations is set with
+  maxIter (defaults to 1000). A small volume is defined as the factor
+  minCellVol (defaults to 0.1) times the original cel volume.
 
   This utility is useful for creating distorted grids for testing
   discretisations.
@@ -183,6 +184,10 @@ When using `addTinyPatch` the original mesh is overwritten!
   // Optional: local motion rescale factor for applying corrections
   // Defaults to 0.8
   //beta 0.8;
+
+  // Optional: local minimum allow cell volume relative to the initial volume
+  // Defaults to 0.1
+  //minCellVol 0.1;
   ```
 
   The `Gaussian` distribution can only be used in combination with
