+++ applications/utilities/perturbMeshPoints/perturbMeshPoints.C
@@ -53,6 +53,8 @@ Author
 #include "argList.H"
 #include "Random.H"
 #include "twoDPointCorrector.H"
+#include "primitiveMeshTools.H"
+#include "unitConversion.H"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
@@ -199,6 +201,38 @@ void calcFeatures
 
 #endif // OPENFOAM_COM
 
+
+// Modified form OpenFOAM-v2312 primitiveMeshCheck.C
+label numSevereNonOrthoFaces(const fvMesh& mesh)
+{
+    // Calculate the mesh orthogonality
+    tmp<scalarField> tortho = primitiveMeshTools::faceOrthogonality
+    (
+        mesh,
+        mesh.faceAreas(),
+        mesh.cellCentres()
+    );
+    const scalarField& ortho = tortho();
+
+    // Severe nonorthogonality threshold
+    const scalar nonOrthThreshold = 70;
+    const scalar severeNonorthogonalityThreshold =
+        ::cos(degToRad(nonOrthThreshold));
+
+    label severeNonOrth = 0;
+
+    forAll(ortho, facei)
+    {
+        if (ortho[facei] < severeNonorthogonalityThreshold)
+        {
+            severeNonOrth++;
+        }
+    }
+
+    return severeNonOrth;
+}
+
+
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 int main(int argc, char *argv[])
@@ -224,7 +258,7 @@ int main(int argc, char *argv[])
     );
 
     // Read inputs
-    const scalar seed(readScalar(perturbDict.lookup("seed")));
+    const int seed(readInt(perturbDict.lookup("seed")));
     const vector scaleFactor(perturbDict.lookup("scaleFactor"));
     const wordList fixedPatchesList(perturbDict.lookup("fixedPatches"));
     const scalar beta(perturbDict.lookupOrDefault("beta", 0.8));
@@ -252,6 +286,12 @@ int main(int argc, char *argv[])
     // We will not allow the volume of each cell to become less than this
     const scalarField minV(minCellVol*mesh.V());
 
+    // Calculate and store a copy of the mesh indexing as this will not change
+    // and we don't want to recalculate it
+    const labelListList cellCells(mesh.cellCells());
+    const labelListList pointCells(mesh.pointCells());
+    const labelListList pointPoints(mesh.pointPoints());
+
     // Calculate the minimum edge length connected to each point
     const labelListList& pointEdges = mesh.pointEdges();
     const edgeList& edges = mesh.edges();
@@ -270,6 +310,9 @@ int main(int argc, char *argv[])
 
         minEdgeLength[pI] = minLen;
     }
+
+    // Store the original minEdgeLength
+    const scalarField oldMinEdgeLength(minEdgeLength);
 
     // Calculate a mask to identify fixed points
     boolList fixedPoint(oldPoints.size(), false);
@@ -320,6 +363,7 @@ int main(int argc, char *argv[])
     //        is reached
     bool validMesh = false;
     int iter = 0;
+    int nRndReset = 0;
     do
     {
         Info<< "Iteration = " << iter << endl;
@@ -371,7 +415,8 @@ int main(int argc, char *argv[])
         {
             const pointField& pointNormals =
                 mesh.boundaryMesh()[patchI].pointNormals();
-            const labelList& meshPoints = mesh.boundaryMesh()[patchI].meshPoints();
+            const labelList& meshPoints =
+                mesh.boundaryMesh()[patchI].meshPoints();
 
             forAll(pointNormals, pI)
             {
@@ -411,22 +456,60 @@ int main(int argc, char *argv[])
             }
         }
 
[…]
