+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/boundaryConditions/symmetric/symmetricLinearMomentumFvPatchVectorField/symmetricLinearMomentumFvPatchVectorField.C
@@ -9,11 +9,11 @@ License
 
     solids4foam is distributed in the hope that it will be useful, but
     WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General 
-    Public License for more details.
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
 
     You should have received a copy of the GNU General Public License
-    along with solids4foam. If not, see <http://www.gnu.org/licenses/>.
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
 
 \*---------------------------------------------------------------------------*/
 
