### 2024-12-21 a9026a2
Uses solids4FoamChangeCopyright.sh  to correct License
