+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/updateVariables.H
@@ -50,7 +50,7 @@ S_t_ = mech_.Smatrix_t();
 lmR_ = interpolate_.surfaceToVol(lmC_);
 lmRgrad_ = grad_.localGradient(lmR_, lmC_);
 interpolate_.volToPoint(lmR_, lmRgrad_, lmN_);
-#include "strongBCs.H" 
+#include "strongBCs.H"
 lmN_.correctBoundaryConditions();
 
 // Constrained fluxes
