+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/boundaryConditions/symmetric/symmetricTractionFvPatchVectorField/symmetricTractionFvPatchVectorField.C
@@ -14,7 +14,7 @@ License
 
     You should have received a copy of the GNU General Public License
     along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-    
+
 \*---------------------------------------------------------------------------*/
 
 #include "symmetricTractionFvPatchVectorField.H"
