### 2024-12-23 solids4foam/solids4foam#150 @MakisH
On `development`, the `Allrun` part now refers to the different options, but the website still shows the old version with the multiple scripts.

Still, there is no clear guidance on what to do after running the simulation to check the results. There is a link to the reference results repository at the bottom, but the documentation and the scripts there mainly helps someone that wants to reproduce all the results of the detailed study.

What would be nice here would be a short section `Analyzing the Results`, similarly to the `beamInCrossFlow` tutorial, pointing to specific scripts and files for one case as an example, which then the user could modify. In any case, a tutorial has different goals than a study, and one would likely run one or two selected cases individually, not all combinations presented in the results.
