### 2024-12-23 solids4foam/solids4foam#150 @iBatistic
> On development, the Allrun part now refers to the different options, but the website still shows the old version with the multiple scripts.

This issue is related to the branch being used. We made changes on the development branch, but the page is rendered using the linked README files from the main branch. I believe the solution might be to update the README on the main branch, but @philipcardiff would know best.

> Still, there is no clear guidance on what to do after running the simulation to check the results.

We have added Gnuplot scripts that automatically plot the number of FSI iterations per time step, as well as the time evolution of radial and axial displacements of point A:
```
# Run solver
solids4Foam::runApplication solids4Foam

# Create plots
if command -v gnuplot &> /dev/null
then
    echo "Generating plots"
    gnuplot -p displacement.gnuplot &> /dev/null
    gnuplot -p fsiConvergence.gnuplot &> /dev/null
else
    echo "Please install gnuplot if you would like to generate the plots"
fi
```

Since these diagrams are generated automatically for the current simulation and are the same as those on the page, I didn’t think it was necessary to alter the page.
