### 2024-12-23 solids4foam/solids4foam#150 @MakisH
> We have added Gnuplot scripts that automatically plot the number of FSI iterations per time step, as well as the time evolution of radial and axial displacements of point A:

Ah, that's very nice indeed! You could also at the end hint where are the plots written (which files).

> Since these diagrams are generated automatically for the current simulation and are the same as those on the page, I didn’t think it was necessary to alter the page.

From the perspective of a new user, such guidance definitely helps. And I would say that documentation is generally expected to have some duplication, because as humans we only perceive a part of what we encounter at once.

> Sure, we can add the "Analyzing the Results" section. Since other tutorials include it, adding it ensures consistency. That said, I’d also vote to keep the current "Expected Results" section, as it is very useful.

Yes, definitely keep both. Different goals, and the "Expected Results" here is very thorough and interesting.
