+++ applications/utilities/projectPatchToSphere/projectPatchToSphere.C
@@ -1,0 +1,201 @@
+/*---------------------------------------------------------------------------*\
+  =========                 |
+  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
+   \\    /   O peration     |
+    \\  /    A nd           | Copyright held by original author
+     \\/     M anipulation  |
+-------------------------------------------------------------------------------
+License
+    This file is part of OpenFOAM.
+
+    OpenFOAM is free software; you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation; either version 2 of the License, or (at your
+    option) any later version.
+
+    OpenFOAM is distributed in the hope that it will be useful, but WITHOUT
+    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+    for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with OpenFOAM; if not, write to the Free Software Foundation,
+    Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
+
+Description
+    Project points of the specified patch onto the surface of a sphere.
+
+    The sphere is specified by origin and radius, e.g.
+
+        projectPatchToSphere leftPatch (0 0 0) 0.5
+
+    Take care as this does not move internal points so it may cause invalid
+    cells.
+
+Author
+    Philip Cardiff, UCD. All rights reserved.
+
+\*---------------------------------------------------------------------------*/
+
+#include "fvCFD.H"
+#include "argList.H"
+#include "pointMesh.H"
+#include "pointFields.H"
+
+using namespace Foam;
+
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+// Main program:
+
+int main(int argc, char *argv[])
+{
+    argList::noParallel();
+    argList::validArgs.append("patch name");
+    argList::validArgs.append("origin");
+    argList::validArgs.append("radius");
+
+#   include "setRootCase.H"
+
+#ifdef OPENFOAM_NOT_EXTEND
+    const word patchName(args[1]);
+    #ifdef OPENFOAM_COM
+    const vector origin(args.get<vector>(2));
+    const scalar radius(args.get<scalar>(3));
+    #else
+    const vector origin(args.argRead<vector>(2));
+    const scalar radius(args.argRead<scalar>(3));
+    #endif
+#else
+    const word patchName(args.additionalArgs()[0]);
+    const vector origin(IStringStream(args.additionalArgs()[1])());
+    const scalar radius(readScalar(IStringStream(args.additionalArgs()[2])()));
+#endif
+
+    if (radius < SMALL)
+    {
+        FatalError
+            << "radius must be greater than zero"
+            << abort(FatalError);
+    }
+
+    Info<< "Patch:" << patchName << nl
+        << "origin:" << origin << nl
+        << "Radius:" << radius << endl;
+
+#   include "createTime.H"
+#   include "createMesh.H"
+
+    const word oldInstance = mesh.pointsInstance();
+
+
+    // Get patch ID
+
+    const label patchID = mesh.boundaryMesh().findPatchID(patchName);
+
+    if (patchID == -1)
+    {
+        FatalError
+            << "Cannot find patch " << patchName
+            << abort(FatalError);
+    }
+
+    // Read the points field
+    pointIOField points
+    (
+        IOobject
+        (
+            "points",
+            runTime.findInstance(polyMesh::meshSubDir, "points"),
+            polyMesh::meshSubDir,
+            runTime,
+            IOobject::MUST_READ,
+            IOobject::NO_WRITE,
+            false
+        )
+    );
+
[…]
