### 2024-12-29 e19f472
Fix syntax typo in beamInCrossFlow tutorial (#154)
* Fix syntax typo in beamInCrossFlow tutorial

* Fix markdownlint issues

---------

Co-authored-by: Philip Cardiff <philip.cardiff@ucd.ie>
