### 2024-12-31 f0fbb28
3dTube/README.md: minor English updates
