+++ tutorials/fluidSolidInteraction/3dTube/README.md
@@ -31,10 +31,10 @@ The fluid is assumed incompressible, Newtonian and isothermal, with a density of
 of 1333.3 Pa, is applied at the tube inlet for a duration of 3e-3 s. The outlet
 pressure is held at 0 Pa throughout.
 
-The tube wall is assumed an isotropic elastic body under the small-strain
+The tube wall is assumed to be an isotropic elastic body under the small-strain
 regime, modelled with Hooke's law, for consistency with the original publication
 that proposed this benchmark. The density is 1200.0 kg/m3, Young's modulus is
-300 kPa, and Poisson's ratio is 0.3.
+300 kPa and Poisson's ratio is 0.3.
 
 The case demonstrates a strong coupling between the fluid and the solid due to
 the high fluid-to-solid density ratio. When using a partitioned solution
@@ -55,14 +55,14 @@ FSI coupling:
 - **Robin-Neumann coupling**, where a Dirichlet condition is applied to the
   fluid velocity, a Robin condition to the fluid pressure, and a Neumann
   condition to the solid displacement. This approach **does** require
-  modification of the underlying fluid solver, and hence it cannot be considered
+  modification of the underlying fluid solver, and hence, it cannot be considered
   a _black-box_ coupling approach; see
   [Tuković Ž, Bukač M, Cardiff P, Jasak H, Ivanković A, 2018, Added mass
   partitioned fluid–structure interaction solver based on a robin boundary
   condition for pressure. In: OpenFOAM selected papers of the 11th workshop.
   Springer, Berlin, pp 1–23](https://doi.org/10.1007/978-3-319-60846-4_1).
 
-For each of these two classes of approach, we can employ different methods to
+For each of these two classes of approaches, we can employ different methods to
 accelerate the FSI iteration loop convergence; within solids4foam, we can use:
 
 - Aitken's dynamic relaxation;
@@ -94,7 +94,7 @@ In this tutorial, we will compare six variants of the approaches above:
 5. **Robin-Neumann formulation with an incompressible fluid**: Support for this
    approach is implemented in the `pimpleFluid` model, where we apply special
    interface boundary conditions: `elasticWallPressure` for the fluid pressure
-   field; and `elasticWallVelocity` for the fluid velocity field.
+   field, and `elasticWallVelocity` for the fluid velocity field.
 6. **Dirichlet-Neumann formulation with IQNILS acceleration and an
    incompressible fluid model using preCICE**: This approach is the same as
    approach 1, except the [preCICE](http://precice.org) coupling implementation
@@ -105,10 +105,10 @@ In this tutorial, we will compare six variants of the approaches above:
    this preCICE IQNILS approach may perform differently than the built-in
    solids4foam approach.
 
-In all approaches, the solid domain setup is exactly the same, where an
+In all approaches, the solid domain setup is the same, where an
 incremental small strain formulation is used (the `linearGeometry` solid model).
 One-quarter of the tube's cross-section is considered, although the case could
-actually be modelled as 2-D axisymmetric. The test is run for 0.02 s. A
+be modelled as 2-D axisymmetric. The test is run for 0.02 s. A
 relatively tight FSI loop tolerance of 1e-6 is used for all approaches based on
 the interface motion. For approach 6 (preCICE), the relative displacement
 tolerance was set to 1e-6, and the relative force tolerance was set to 1e-3.
@@ -122,7 +122,7 @@ In all cases, the first-order Euler time scheme is used for the solid and fluid.
 For approaches 1 to 4, the tutorial case is located at `solids4foam/tutorials/fluidSolidInteraction/3dTube`.
 The case can be run using the single `Allrun` script. Use `./Allrun` for
[…]
