### 2024-12-31 solids4foam/solids4foam#150 @philipcardiff
Thanks, both,

> > On development, the Allrun part now refers to the different options, but the website still shows the old version with the multiple scripts.
>
> This issue is related to the branch being used. We made changes on the development branch, but the page is rendered using the linked README files from the main branch. I believe the solution might be to update the README on the main branch, but @philipcardiff would know best.

OK, this point is now resolved. Currently, the solids4foam repository is added as a sub-module of the solid4foam.github.io repository, with links to the MarkDown files; I must manually update the solids4foam commit. I can probably automate this with a GitHub action, let me see what my generative AI friend says ...
