### 2024-12-31 6830d75
Add GitHub action to update the soldis4foam.github.io repository
