+++ .github/workflows/update-submodule.yml
@@ -25,7 +25,13 @@ jobs:
           git config submodule.imported/solids4foam.url https://github.com/solids4foam/solids4foam.git
           git submodule update --init --recursive
 
-      # Step 3: Update the submodule to the latest commit on the "development" branch
+      # Step 3: Set Git identity for the runner
+      - name: Configure Git identity
+        run: |
+          git config --global user.name "GitHub Actions"
+          git config --global user.email "actions@github.com"
+          
+      # Step 4: Update the submodule to the latest commit on the "development" branch
       - name: Update submodule to latest commit
         run: |
           cd solids4foam.github.io
@@ -34,9 +40,9 @@ jobs:
           git checkout origin/development
           cd ..
           git add solids4foam
-          git commit -m "Aotmated update of submodule to the latest development commit"
-
-      # Step 4: Push the changes back to the solids4foam.github.io repository
+          git commit -m "Automated update of submodule to the latest development commit"
+      
+      # Step 5: Push the changes back to the solids4foam.github.io repository
       - name: Push updated submodule
         run: |
           cd solids4foam.github.io
