### 2024-12-31 8bf0348
Update action: update-submodule.yml
