### 2024-12-31 fdd3aee
Update action: update-submodule.yml
