+++ .github/workflows/update-submodule.yml
@@ -46,6 +46,6 @@ jobs:
       - name: Push updated submodule
         run: |
           cd solids4foam.github.io
-          git push origin master
+          git push https://x-access-token:${{ secrets.PAT_GITHUB_IO }}@github.com/solids4foam/solids4foam.github.io.git master
         env:
           GITHUB_TOKEN: ${{ secrets.PAT_GITHUB_IO }}
