### 2025-01-01 4fd2adb
beamInCrossFlow and 3dTube README updates
