+++ tutorials/fluidSolidInteraction/3dTube/README.md
@@ -25,7 +25,7 @@ This case consists of a pressure pulse applied in a thick-walled elastic tube
 
 ![Figure 1: Wave propagation in an elastic pipe](images/3dTube.png)
 
-Figure 1: Wave propagation in an elastic pipe
+### Figure 1: Wave propagation in an elastic pipe
 
 The fluid is assumed incompressible, Newtonian and isothermal, with a density of
 1000.0 kg/m3 and kinematic viscosity of 3e-6 m2/s. A pressure wave, with a peak
@@ -212,10 +212,12 @@ conditions `elasticWallVelocity` and `elasticWallPressure` at the interface.
 The `Allrun` script updates the case with links to the correct files to be used
 by each approach. The Allrun script runs the OpenFOAM `blockMesh` utility to
 generate the meshes in the solid and fluid domains, followed by running the
-`solids4Foam` solver. Subsequently, if `gnuplot` is installed, two figures will
+`solids4Foam` solver. Subsequently, if `gnuplot` is installed, three figures will
 be generated in the case directory:
 
 - `axialDisplacement.pdf`: this plots the axial displacement of point A vs time.
+- `radialDisplacement.pdf`: this plots the radial displacement of point A vs
+  time.
 - `fsiConvergence.pdf`: this plots the number of FSI iterations per time step.
 
 The tutorial case for approach 6, which uses preCICE, is located at
@@ -229,6 +231,42 @@ Remember that a tutorial case can be cleaned and reset using the included
 
 ---
 
+## Analysing the Results
+
+As in the `beamInCrossFlow` tutorial, the case can be opened in ParaView (Figure
+2) by creating an empty `case.foam` file in the case directory (e.g.,
+`touch case.foam`) and opening it in ParaView (e.g., _File -> Open ->
+case.foam_). The _Extract Block_ ParaView filter can be used to extract the solid
+region, and a second instance of the _Extract Block_ filter can extract the fluid
+region. In that way, the fluid region can be coloured by a fluid field (e.g. `p`
+or `U`) and the solid region by a solid field (e.g. `D` or `sigmaEq`). In Figure
+2, multiple instances of the _Reflect_ filter are used to display the entire
+fluid domain (rather than one quarter of it) and three quarters of the solid
+domain.
+
+![Diagram of ParaView interface showing the 3-D tube problem](images/3dTube_paraview.png)
+
+### Figure 2: The solid displacement and fluid pressure in ParaView
+
+As noted above, if `gnuplot` is installed, the files `axialDisplacement.pdf`,
+`radialDisplacement.pdf` and `fsiConvergence.pdf` will be available in the case
+directory. You can open these figures using a PDF viewer, e.g.
+`evince axialDisplacement.pdf` in Linux or `open axialDisplacement.pdf` in macOS.
+The `axialDisplacement.pdf` (Figure 3) figure plots the axial displacement of
+point A (see Figure 1) vs time, while `radialDisplacement.pdf` plots the radial
+displacement of point A. The `fsiConvergence.pdf` (Figure 4) file plots the
+number of FSI iterations per time step.
+
+![Axial displacement](images/axialDisplacement.pdf)
+
+### Figure 3: `axialDisplacement.pdf`
+
+![FSI convergence](images/fsiConvergence.pdf)
+
+### Figure 4: `fsiConvergence.pdf`
+
+---
+
 ## Expected Results
 
 This case has been proposed as a benchmark for FSI problems. The solution for
@@ -238,8 +276,8 @@ point A's (see Figure 1) axial and radial displacement is available in:
 ### Large Time Step
 
 Initially, we compare the solutions using a relatively large time step size of
-1e-4 s. Figure 2 shows the radial displacement at point A vs time, Figure 3
[…]
