### 2025-01-01 solids4foam/solids4foam#150 @philipcardiff
> > Since these diagrams are generated automatically for the current simulation and are the same as those on the page, I didn’t think it was necessary to alter the page.
>
> From the perspective of a new user, such guidance definitely helps. And I would say that documentation is generally expected to have some duplication, because as humans we only perceive a part of what we encounter at once.
>
> > Sure, we can add the "Analyzing the Results" section. Since other tutorials include it, adding it ensures consistency. That said, I’d also vote to keep the current "Expected Results" section, as it is very useful.
>
> Yes, definitely keep both. Different goals, and the "Expected Results" here is very thorough and interesting.

I have now added a `Analysing the Results` section and added more details to `Running the Case`: to https://www.solids4foam.com/tutorials/tutorial3.html. I also updated

I agree that this improves this tutorial; thanks again.

I think this point is now resolved, but feel free to re-open.
