### 2025-01-01 #7407 @philipcardiff
> Update: The beamInCrossFlow now runs, but with OpenFOAM v2312 (without the file patches) I am getting a bit different results ⚠️, looking at the value ranges. Is this because of the version difference, or because I am doing something wrong in ParaView?

@MakisH Regarding this point, I have updated the image at https://www.solids4foam.com/tutorials/tutorial3.html, so it should now be consistent with your results. The reason for the difference is that there are two forms of the case (as described at the start of https://www.solids4foam.com/tutorials/tutorial3.html), and the previous image was for the first form of the case, whereas the case in solids4foam is actually set up to match the second form of the case (we made this change at some point, but I forgot to update this image). Along with updating the image, I have made a few minor edits at https://www.solids4foam.com/tutorials/tutorial3.html to clarify this point. Thanks!
