### 2025-01-04 1391f29
higherOrderGrad: minor efficiency and style changes
