+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -824,10 +824,11 @@ void higherOrderGrad::calcQRCoeffs() const
 
         // Find max distance in this stencil
         scalar maxDist = 0.0;
+        const vector& curC = CI[cellI];
         forAll(curStencil, cI)
         {
             const label neiCellID = curStencil[cI];
-            const scalar d = mag(CI[neiCellID] - CI[cellI]);
+            const scalar d = mag(CI[neiCellID] - curC);
             if (d > maxDist)
             {
                 maxDist = d;
@@ -847,7 +848,9 @@ void higherOrderGrad::calcQRCoeffs() const
         {
             FatalErrorInFunction
                 << "Interpolation stencil needs to be bigger than the "
-                << "number of elements in Taylor order!"
+                << "number of elements in Taylor order!" << nl
+                << "Stencil size = " << Nn << ", Taylor elements = " << Np << nl
+                << "Cell centre = " << curC
                 << exit(FatalError);
         }
 
@@ -1021,6 +1024,13 @@ void higherOrderGrad::calcQRCoeffs() const
 
        QRInterpCoeffs[cellI].shrink();
        QRGradCoeffs[cellI].shrink();
+    }
+
+    if (calcConditionNumber_)
+    {
+        Info<< "Writing " << conditionNumber().name() << " to time = "
+            << mesh.time().value() << endl;
+        conditionNumber().write();
     }
 
     //if (debug)
@@ -1591,7 +1601,7 @@ void higherOrderGrad::calcGlobalQRFaceGPCoeffs() const
     forAll(QRGradCoeffs, faceI)
     {
         List<DynamicList<vector>>& facePointsGC = QRGradCoeffs[faceI];
-        facePointsGC.setSize(mesh.faces()[faceI].size() * triQuadraturePtsNb_);
+        facePointsGC.setSize(mesh.faces()[faceI].size()*triQuadraturePtsNb_);
     }
 
     // Refernces for brevity and efficiency
@@ -1678,7 +1688,9 @@ void higherOrderGrad::calcGlobalQRFaceGPCoeffs() const
         {
             FatalErrorInFunction
                 << "Interpolation stencil needs to be bigger than the "
-                << "number of elements in Taylor order!"
+                << "number of elements in Taylor order!" << nl
+                << "Stencil size = " << Nn << ", Taylor elements = " << Np << nl
+                << "Face centre = " << curCf
                 << exit(FatalError);
         }
 
@@ -2356,8 +2368,8 @@ void higherOrderGrad::calcGaussPointsAndWeights() const
         List<point>& fGP = faceGP[i];
         List<scalar>& fGPW = faceGPW[i];
 
-        fGP.setSize(mesh.faces()[i].size() * triQuadraturePtsNb_);
-        fGPW.setSize(mesh.faces()[i].size() * triQuadraturePtsNb_);
+        fGP.setSize(mesh.faces()[i].size()*triQuadraturePtsNb_);
+        fGPW.setSize(mesh.faces()[i].size()*triQuadraturePtsNb_);
     }
 
     // 1. Stage - decompose faces into triangles. Store triangle points
@@ -2501,7 +2513,7 @@ const List<DynamicList<vector>>& higherOrderGrad::QRGradFaceCoeffs() const
 
 
 const List<List<DynamicList<vector>>>&
-higherOrderGrad::QRGradFaceGPCoeffs()const
+higherOrderGrad::QRGradFaceGPCoeffs() const
 {
     if (!QRGradFaceGPCoeffsPtr_)
     {
[…]
