### 2025-01-07 #7407 @MakisH
Thanks for the several updates in these tutorials, @philipcardiff. Note that the ranges in the image are still a bit different than the ranges in my results. Are these because of the optional fixes in `backwardDdtScheme.C` that I have not applied for v2312, or are the results with a different OpenFOAM version?

Looking at the state, I now only want to have another, deeper look at the 3dTube tutorial, and of course at the paper itself once the rest converges.

Anything else / more specific aspects you would like feedback on?
