### 2025-01-14 af21dbe
Update leastSquaresS4f to use 1/d^q weights
