+++ src/solids4FoamModels/numerics/leastSquaresS4fGrad/leastSquaresS4fVectors.C
@@ -93,23 +93,39 @@ void Foam::leastSquaresS4fVectors::calcLeastSquaresVectors()
     const labelUList& neighbour = mesh_.neighbour();
 
     const volVectorField& C = mesh.C();
-    const surfaceScalarField& w = mesh.weights();
-    const surfaceScalarField& magSf = mesh.magSf();
+    //const surfaceScalarField& w = mesh.weights();
+    //const surfaceScalarField& magSf = mesh.magSf();
 
+    // Exponent in the inverse distance weight for the least squares system
+    // Weights are calculated as 1/d^q, i.e. q = 0: unweighted least squares
+    // Syrakos et al, 2023, doi.org/10.1016/j.matcom.2022.09.008 showed q = 3 to
+    // work best for cell-centre gradients
+    const scalar q = 3;
 
     // Set up temporary storage for the dd tensor (before inversion)
     symmTensorField dd(mesh_.nCells(), Zero);
 
     forAll(owner, facei)
     {
-        label own = owner[facei];
-        label nei = neighbour[facei];
+        const label own = owner[facei];
+        const label nei = neighbour[facei];
 
-        vector d = C[nei] - C[own];
-        symmTensor wdd = (magSf[facei]/magSqr(d))*sqr(d);
+        const vector d = C[nei] - C[own];
 
-        dd[own] += (1 - w[facei])*wdd;
-        dd[nei] += w[facei]*wdd;
+        // Weight
+        const scalar v = 1.0/pow(mag(d), q);
+
+        const symmTensor wdd = v*sqr(d);
+
+        dd[own] += wdd;
+        dd[nei] += wdd;
+
+        // Original OpenFOAM approach where the weights are magSf/magSqr(d)
+        // symmTensor wdd = (magSf[facei]/magSqr(d))*sqr(d);
+
+        // Original OpenFOAM approach using face weights
+        // dd[own] += (1 - w[facei])*wdd;
+        // dd[nei] += w[facei]*wdd;
     }
 
 
@@ -118,19 +134,19 @@ void Foam::leastSquaresS4fVectors::calcLeastSquaresVectors()
 
     forAll(pVectorsBf, patchi)
     {
-        const fvsPatchScalarField& pw = w.boundaryField()[patchi];
-        const fvsPatchScalarField& pMagSf = magSf.boundaryField()[patchi];
+        // const fvsPatchScalarField& pw = w.boundaryField()[patchi];
+        // const fvsPatchScalarField& pMagSf = magSf.boundaryField()[patchi];
 
-        const fvPatch& p = pw.patch();
+        const fvPatch& p = pVectorsBf[patchi].patch();
         const labelUList& faceCells = p.patch().faceCells();
         const vectorField& Cf = p.Cf();
 
         // Build the d-vectors
         // In OF.com/OF.org, p.delta are the orthogonal components of the real d
         // vectors, so we need to build them ourselves
-        //vectorField pd(p.delta());
+        // vectorField pd(p.delta());
 
-        if (pw.coupled())
+        if (p.coupled())
         {
             // Coupled d vectors
             const vectorField pd(p.delta());
@@ -138,9 +154,13 @@ void Foam::leastSquaresS4fVectors::calcLeastSquaresVectors()
             forAll(pd, patchFacei)
             {
                 const vector& d = pd[patchFacei];
+                const scalar v = 1.0/pow(mag(d), q);
 
-                dd[faceCells[patchFacei]] +=
-                    ((1 - pw[patchFacei])*pMagSf[patchFacei]/magSqr(d))*sqr(d);
+                dd[faceCells[patchFacei]] += v*sqr(d);
+
+                // Original OpenFOAM version
[…]
