### 2025-01-14 603b234
foamPetscSnesHelper: allow option to return error code
