### 2025-01-15 082678c
Redefines setTraction function in explicitGodunovCCSolid Class
