+++ src/solids4FoamModels/solidModels/explicitGodunovCCSolid/gEqns.H
@@ -53,10 +53,10 @@ if (runTime_.timeIndex() == 0)
 
 else
 {
-    xN_ += pDeltaT_*(lmN_/rho_) - pDeltaT_ * ( (3*xN_.prevIter() - 4* xN_.oldTime() + xN_.oldTime().oldTime())/(2*deltaT_));
+    xN_ += pDeltaT_*(lmN_/rho_) - pDeltaT_*( (3*xN_.prevIter() - 4* xN_.oldTime() + xN_.oldTime().oldTime())/(2*deltaT_));
 
-    lm_ += pDeltaT_*rhsLm_ - pDeltaT_*( (3*lm_.prevIter() - 4* lm_.oldTime() + lm_.oldTime().oldTime())/(2*deltaT_));
+    lm_ += pDeltaT_*rhsLm_ - pDeltaT_*( (3*lm_.prevIter() - 4*lm_.oldTime() + lm_.oldTime().oldTime())/(2*deltaT_));
     lm_ -= pDeltaT_*dampingCoeff()*lm_;
 
-    F_ += pDeltaT_*fvc::surfaceIntegrate((lmC_/rho_)*Sf_) - pDeltaT_*( (3*F_.prevIter() - 4* F_.oldTime() + F_.oldTime().oldTime())/(2*deltaT_));
+    F_ += pDeltaT_*fvc::surfaceIntegrate((lmC_/rho_)*Sf_) - pDeltaT_*( (3*F_.prevIter() - 4*F_.oldTime() + F_.oldTime().oldTime())/(2*deltaT_));
 }
