### 2025-01-19 8bd739a
linGeomTotalDispSolid: make segregated implicit consistent with SNES
