### 2025-01-19 9e2e12b
leastSquaresS4fVectors: revert to previous OF vectors
