+++ src/solids4FoamModels/numerics/leastSquaresS4fGrad/leastSquaresS4fVectors.C
@@ -93,39 +93,23 @@ void Foam::leastSquaresS4fVectors::calcLeastSquaresVectors()
     const labelUList& neighbour = mesh_.neighbour();
 
     const volVectorField& C = mesh.C();
-    //const surfaceScalarField& w = mesh.weights();
-    //const surfaceScalarField& magSf = mesh.magSf();
+    const surfaceScalarField& w = mesh.weights();
+    const surfaceScalarField& magSf = mesh.magSf();
 
-    // Exponent in the inverse distance weight for the least squares system
-    // Weights are calculated as 1/d^q, i.e. q = 0: unweighted least squares
-    // Syrakos et al, 2023, doi.org/10.1016/j.matcom.2022.09.008 showed q = 3 to
-    // work best for cell-centre gradients
-    const scalar q = 3;
 
     // Set up temporary storage for the dd tensor (before inversion)
     symmTensorField dd(mesh_.nCells(), Zero);
-
+    Info<< "*********HI" << endl;
     forAll(owner, facei)
     {
-        const label own = owner[facei];
-        const label nei = neighbour[facei];
+        label own = owner[facei];
+        label nei = neighbour[facei];
 
-        const vector d = C[nei] - C[own];
+        vector d = C[nei] - C[own];
+        symmTensor wdd = (magSf[facei]/magSqr(d))*sqr(d);
 
-        // Weight
-        const scalar v = 1.0/pow(mag(d), q);
-
-        const symmTensor wdd = v*sqr(d);
-
-        dd[own] += wdd;
-        dd[nei] += wdd;
-
-        // Original OpenFOAM approach where the weights are magSf/magSqr(d)
-        // symmTensor wdd = (magSf[facei]/magSqr(d))*sqr(d);
-
-        // Original OpenFOAM approach using face weights
-        // dd[own] += (1 - w[facei])*wdd;
-        // dd[nei] += w[facei]*wdd;
+        dd[own] += (1 - w[facei])*wdd;
+        dd[nei] += w[facei]*wdd;
     }
 
 
@@ -134,19 +118,19 @@ void Foam::leastSquaresS4fVectors::calcLeastSquaresVectors()
 
     forAll(pVectorsBf, patchi)
     {
-        // const fvsPatchScalarField& pw = w.boundaryField()[patchi];
-        // const fvsPatchScalarField& pMagSf = magSf.boundaryField()[patchi];
+        const fvsPatchScalarField& pw = w.boundaryField()[patchi];
+        const fvsPatchScalarField& pMagSf = magSf.boundaryField()[patchi];
 
-        const fvPatch& p = pVectorsBf[patchi].patch();
+        const fvPatch& p = pw.patch();
         const labelUList& faceCells = p.patch().faceCells();
         const vectorField& Cf = p.Cf();
 
         // Build the d-vectors
         // In OF.com/OF.org, p.delta are the orthogonal components of the real d
         // vectors, so we need to build them ourselves
-        // vectorField pd(p.delta());
+        //vectorField pd(p.delta());
 
-        if (p.coupled())
+        if (pw.coupled())
         {
             // Coupled d vectors
             const vectorField pd(p.delta());
@@ -154,13 +138,9 @@ void Foam::leastSquaresS4fVectors::calcLeastSquaresVectors()
             forAll(pd, patchFacei)
             {
                 const vector& d = pd[patchFacei];
-                const scalar v = 1.0/pow(mag(d), q);
 
-                dd[faceCells[patchFacei]] += v*sqr(d);
-
-                // Original OpenFOAM version
-                // dd[faceCells[patchFacei]] +=
[…]
