### 2025-01-22 30e81b6
Add new checkConvergence, similar to PETSc snes convergence checks
