### 2025-01-22 becfef0
nonLinGeomTotalLagTotalDispSolid: use new convergence check
