+++ src/solids4FoamModels/solidModels/nonLinGeomTotalLagTotalDispSolid/nonLinGeomTotalLagTotalDispSolid.C
@@ -165,11 +165,13 @@ bool nonLinGeomTotalLagTotalDispSolid::evolveImplicitSegregated()
     }
 
     int iCorr = 0;
+    scalar currentResidualNorm = 0;
+    scalar initialResidualNorm = 0;
+    scalar deltaXNorm = 0;
+    scalar xNorm = 0;
 #ifdef OPENFOAM_NOT_EXTEND
-    SolverPerformance<vector> solverPerfD;
     SolverPerformance<vector>::debug = 0;
 #else
-    lduSolverPerformance solverPerfD;
     blockLduMatrix::debug = 0;
 #endif
 
@@ -228,8 +230,21 @@ bool nonLinGeomTotalLagTotalDispSolid::evolveImplicitSegregated()
         // Enforce any cell displacements
         solidModel::setCellDisps(DEqn);
 
-        // Solve the linear system
-        solverPerfD = DEqn.solve();
+        // Solve the linear system and store the residual
+        currentResidualNorm = mag(DEqn.solve().initialResidual());
+
+        // Store the initial residual
+        if (iCorr == 0)
+        {
+            initialResidualNorm = currentResidualNorm;
+        }
+
+        // Norm of the solution correction
+        deltaXNorm =
+            gSum(mag(D().primitiveField() - D().prevIter().primitiveField()));
+
+        // Norm of the solution
+        xNorm = gSum(mag(D().primitiveField()));
 
         // Fixed or adaptive field under-relaxation
         relaxField(D(), iCorr);
@@ -262,18 +277,32 @@ bool nonLinGeomTotalLagTotalDispSolid::evolveImplicitSegregated()
     }
     while
     (
-       !converged
+        !checkConvergence
         (
-            iCorr,
-#ifdef OPENFOAM_NOT_EXTEND
-            mag(solverPerfD.initialResidual()),
-            cmptMax(solverPerfD.nIterations()),
-#else
-            solverPerfD.initialResidual(),
-            solverPerfD.nIterations(),
-#endif
-            D()
-        ) && ++iCorr < nCorr()
+            currentResidualNorm,
+            initialResidualNorm,
+            deltaXNorm,
+            xNorm,
+            ++iCorr,
+            nCorr(),
+            solidModelDict().lookupOrDefault<scalar>
+            (
+                "rTol",
+                solutionTol()
+            ),
+            solidModelDict().lookupOrDefault<scalar>("aTol", 1e-50),
+            solidModelDict().lookupOrDefault<scalar>
+            (
+                "sTol",
+                solutionTol()
+            ),
+            solidModelDict().lookupOrDefault<scalar>("divTol", 1e4),
+            infoFrequency(),
+            solidModelDict().lookupOrDefault<Switch>
+            (
+                "writeConvergedReason", true
+            )
+        )
     );
 
     // Interpolate cell displacements to vertices
