+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -135,10 +135,8 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
         scalar deltaXNorm = 0;
         scalar xNorm = 0;
 #ifdef OPENFOAM_NOT_EXTEND
-        SolverPerformance<vector> solverPerfD;
         SolverPerformance<vector>::debug = 0;
 #else
-        lduSolverPerformance solverPerfD;
         blockLduMatrix::debug = 0;
 #endif
 
@@ -191,8 +189,14 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
             // Enforce any cell displacements
             solidModel::setCellDisps(DEqn);
 
-            // Solve the linear system
-            solverPerfD = DEqn.solve();
+            // Solve the linear system and store the residual
+            currentResidualNorm = mag(DEqn.solve().initialResidual());
+
+            // Store the initial residual
+            if (iCorr == 0)
+            {
+                initialResidualNorm = currentResidualNorm;
+            }
 
             // Norm of the solution correction
             deltaXNorm =
@@ -230,15 +234,6 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
 
             // Calculate the stress using run-time selectable mechanical law
             mechanical().correct(sigma());
-
-            // Current residual
-            currentResidualNorm = mag(solverPerfD.initialResidual());
-
-            // Initial residual
-            if (iCorr == 0)
-            {
-                initialResidualNorm = currentResidualNorm;
-            }
         }
         while
         (
@@ -250,14 +245,19 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
                 xNorm,
                 ++iCorr,
                 nCorr(),
-                solidModelDict().lookupOrDefault<scalar>("rTol", 1e-8),
-                solidModelDict().lookupOrDefault<scalar>("aTol", 1e-50),
-                solidModelDict().lookupOrDefault<scalar>("sTol", 1e-8),
-                solidModelDict().lookupOrDefault<scalar>("divTol", 1e4),
-                solidModelDict().lookupOrDefault<label>
+                solidModelDict().lookupOrDefault<scalar>
                 (
-                    "writeResidualFrequency", 1
+                    "rTol",
+                    solutionTol()
                 ),
+                solidModelDict().lookupOrDefault<scalar>("aTol", 1e-50),
+                solidModelDict().lookupOrDefault<scalar>
+                (
+                    "sTol",
+                    solutionTol()
+                ),
+                solidModelDict().lookupOrDefault<scalar>("divTol", 1e4),
+                infoFrequency(),
                 solidModelDict().lookupOrDefault<Switch>
                 (
                     "writeConvergedReason", true
