### 2025-01-30 f65b6f5
linGeomTotalDispSolid and nonLinGeomTotalLagTotalDispSolid: print initial residuals; and use L2 instead of L1 norm
