### 2025-01-30 247207b
Remove higher order code to allow merging
