+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.C
@@ -1,3339 +1,0 @@
-/*---------------------------------------------------------------------------*\
-License
-    This file is part of solids4foam.
-
-    solids4foam is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your
-    option) any later version.
-
-    solids4foam is distributed in the hope that it will be useful, but
-    WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-
-\*---------------------------------------------------------------------------*/
-
-#include "higherOrderGrad.H"
-#include "volFields.H"
-#include "surfaceFields.H"
-#include "emptyPolyPatch.H"
-#include "processorPolyPatch.H"
-
-#include "triangle.H"
-#include "triFace.H"
-#include "triangleQuadrature.H"
-
-namespace Foam
-{
-
-// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
-
-defineTypeNameAndDebug(higherOrderGrad, 0);
-
-
-// * * * * * * * * * * *  Private Member Functions * * * * * * * * * * * * * //
-
-
-void higherOrderGrad::makeGlobalCellStencils() const
-{
-    InfoInFunction
-        << "start" << endl;
-
-    if (globalCellStencilsPtr_)
-    {
-        FatalErrorInFunction
-            << "Pointer already set" << exit(FatalError);
-    }
-
-    // References from brevity and efficiency
-    const fvMesh& mesh = mesh_;
-    const label nCells = mesh.nCells();
-    const labelUList& owner = mesh.owner();
-    const labelUList& neighbour = mesh.neighbour();
-    const cellList& cells = mesh.cells();
-
-    // Prepare and store processor neighbour face global cells, i.e. global cell
-    // indices of the cells across the processor patches
-    labelListList procPatchNeiGlobalCellIDs(mesh.boundaryMesh().size());
-    if (Pstream::parRun())
-    {
-        PstreamBuffers pBufs(UPstream::commsTypes::nonBlocking);
-
-        // Send global cell IDs for each processor patch
-        forAll(mesh.boundaryMesh(), patchI)
-        {
-            const polyPatch& pp = mesh.boundaryMesh()[patchI];
-            if (isA<processorPolyPatch>(pp))
-            {
-                const processorPolyPatch& procPatch =
-                    refCast<const processorPolyPatch>(pp);
-
-                UOPstream toNeighbProc(procPatch.neighbProcNo(), pBufs);
-
-                toNeighbProc
-                    << globalCells_.toGlobal(pp.faceCells());
-            }
-        }
-
-        pBufs.finishedSends(); // no-op for blocking
-
-        // Receive data
-        forAll(mesh.boundaryMesh(), patchI)
-        {
-            const polyPatch& pp = mesh.boundaryMesh()[patchI];
-            if (isA<processorPolyPatch>(pp))
-            {
-                const processorPolyPatch& procPatch =
-                    refCast<const processorPolyPatch>(pp);
-
-                UIPstream fromNeighbProc(procPatch.neighbProcNo(), pBufs);
-
-                procPatchNeiGlobalCellIDs[patchI].setSize(pp.size());
-
-                fromNeighbProc
-                    >> procPatchNeiGlobalCellIDs[patchI];
-            }
-        }
-    }
-
-    // Initialize visited sets and frontiers for all cells
[…]
