+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.H
@@ -42,7 +42,6 @@ SourceFiles
 #include "pointFields.H"
 #include "uniformDimensionedFields.H"
 #include "foamPetscSnesHelper.H"
-#include "higherOrderGrad.H"
 
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
@@ -88,9 +87,6 @@ class linGeomTotalDispSolid
 
         //- Field to provide fvOptions with the correct dimensions
         volScalarField ds_;
-
-        //- Higher order gradient calculator
-        autoPtr<higherOrderGrad> hoGradPtr_;
 
 
     // Private Member Functions

[…]
