### 2025-01-30 1825430
Remove triangleQuadrature
