+++ src/solids4FoamModels/numerics/triangleQuadrature/triangleQuadrature.C
@@ -1,264 +1,0 @@
-/*---------------------------------------------------------------------------*\
-License
-    This file is part of solids4foam.
-
-    solids4foam is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your
-    option) any later version.
-
-    solids4foam is distributed in the hope that it will be useful, but
-    WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-
-\*---------------------------------------------------------------------------*/
-
-#include "triangleQuadrature.H"
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-
-// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
-
-Map<triangleQuadrature::quadratureRule> triangleQuadrature::rules_;
-
-// Initialize and return quadrature rules
-void triangleQuadrature::constructRules()
-{
-    if (rules_.size() != 0)
-    {
-        FatalErrorInFunction
-            << "attempt to re-construct rules when they already exist"
-            << exit(FatalError);
-    }
-
-    // 1 point quadrature, exact for polynomials up to 1 order
-    rules_.insert
-    (
-        1,
-        quadratureRule
-        {
-            List<point>{vector(1.0/3.0, 1.0/3.0, 1.0/3.0)},
-            List<scalar>{1.0}
-        }
-    );
-
-    // 3 point quadrature, exact for polynomials up to 2 order
-    rules_.insert
-    (
-        3,
-        quadratureRule
-        {
-            List<point>
-            {
-                point(2.0/3.0, 1.0/6.0, 1.0/6.0),
-                point(1.0/6.0, 2.0/3.0, 1.0/6.0),
-                point(1.0/6.0, 1.0/6.0, 2.0/3.0)
-            },
-            List<scalar>{1.0/3.0, 1.0/3.0, 1.0/3.0}
-        }
-    );
-
-    // 4 point quadrature, exact for polynomials up to 3 order
-    rules_.insert
-    (
-        4,
-        quadratureRule
-        {
-            List<point>
-            {
-                point(1.0/3.0, 1.0/3.0, 1.0/3.0),
-                point(0.6, 0.2, 0.2),
-                point(0.2, 0.6, 0.2),
-                point(0.2, 0.2, 0.6)
-            },
-            List<scalar>{-9.0/16.0, 25.0/48.0, 25.0/48.0, 25.0/48.0}
-        }
-    );
-
-    // 6 point quadrature, exact for polynomials up to 4 order
-    rules_.insert
-    (
-        6,
-        quadratureRule
-        {
-            List<point>
-            {
-                point(0.108103018168070, 0.445948490915965, 0.445948490915965),
-                point(0.445948490915965, 0.108103018168070, 0.445948490915965),
-                point(0.445948490915965, 0.445948490915965, 0.108103018168070),
-                point(0.816847572980459, 0.091576213509771, 0.091576213509771),
-                point(0.091576213509771, 0.816847572980459, 0.091576213509771),
-                point(0.091576213509771, 0.091576213509771, 0.816847572980459)
-            },
-            List<scalar>
-            {
-                0.223381589678011,
-                0.223381589678011,
-                0.223381589678011,
-                0.109951743655322,
-                0.109951743655322,
-                0.109951743655322
-            }
-        }
-    );
-
[…]
