### 2025-01-30 9d29bd7
linGeomTotalDispSolid: remove higherOrder reference
