### 2025-01-30 651f52d
linGeomTotalDispSolid: make norms consistent with nonLinTL
