### 2025-01-30 0025495
linGeomTotalDispSolid: remove unused header
