### 2025-01-30 71e75c8
Feature petsc snes higher order philipc (#157)
* Stencil construction algo

* Required code completed - fix needed

* Bugs resolved

* Added quadratic interpolation

* Added treatement of boundary cells

* Minor output log modifications

* Place new higher order grad scheme in a class

* Bug fix for higher order gradient

Higher order gradient scheme implemented by Ivan Batistic, with re-packaging by Philip Cardiff

Co-authored-by: Philip Cardiff <philipcardiff@gmail.com>
Co-authored-by: Ivan Batistic <ivan.batistic2@gmail.com>

* Add optional boundary patches to least squares stencils

* Effi
[…]
