+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -26,7 +26,6 @@ License
 #include "fixedDisplacementZeroShearFvPatchVectorField.H"
 #include "symmetryFvPatchFields.H"
 
-
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 namespace Foam
@@ -118,6 +117,9 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
     Info<< "Evolving solid solver using an implicit segregated approach"
         << endl;
 
+    // Update D boundary conditions
+    D().correctBoundaryConditions();
+
     if (predictor_ && newTimeStep())
     {
         predict();
@@ -127,15 +129,20 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
     do
     {
         int iCorr = 0;
+        scalar currentResidualNorm = 0;
+        scalar initialResidualNorm = 0;
+        scalar deltaXNorm = 0;
+        scalar xNorm = 0;
 #ifdef OPENFOAM_NOT_EXTEND
-        SolverPerformance<vector> solverPerfD;
         SolverPerformance<vector>::debug = 0;
 #else
-        lduSolverPerformance solverPerfD;
         blockLduMatrix::debug = 0;
 #endif
 
         Info<< "Solving the momentum equation for D" << endl;
+
+        // Unit normal vectors at the faces
+        const surfaceVectorField n(mesh().Sf()/mesh().magSf());
 
         // Momentum equation loop
         do
@@ -143,15 +150,29 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
             // Store fields for under-relaxation and residual calculation
             D().storePrevIter();
 
+            // Calculate raction vectors at the faces
+            surfaceVectorField traction(n & fvc::interpolate(sigma()));
+
+            // Add stabilisation to the traction
+            // We add this before enforcing the traction condition as the stabilisation
+            // is set to zero on traction boundaries
+            // To-do: add a stabilisation traction function to momentumStabilisation
+            const scalar scaleFactor =
+                readScalar(stabilisation().dict().lookup("scaleFactor"));
+            const surfaceTensorField gradDf(fvc::interpolate(gradD()));
+            traction += scaleFactor*impKf_*(fvc::snGrad(D()) - (n & gradDf));
+
+            // Enforce traction boundary conditions
+            enforceTractionBoundaries(traction, D(), n);
+
             // Linear momentum equation total displacement form
             fvVectorMatrix DEqn
             (
                 rho()*fvm::d2dt2(D())
              == fvm::laplacian(impKf_, D(), "laplacian(DD,D)")
               - fvc::laplacian(impKf_, D(), "laplacian(DD,D)")
-              + fvc::div(sigma(), "div(sigma)")
+              + fvc::div(mesh().magSf()*traction)
               + rho()*g()
-              + stabilisation().stabilisation(D(), gradD(), impK_)
               + fvOptions()(ds_, D())
             );
 
@@ -167,8 +188,33 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
             // Enforce any cell displacements
             solidModel::setCellDisps(DEqn);
 
-            // Solve the linear system
-            solverPerfD = DEqn.solve();
+            // Solve the linear system and store the residual
+            currentResidualNorm = mag(DEqn.solve().initialResidual());
+
+            // Norm of the solution correction
+            deltaXNorm =
+                sqrt
+                (
+                    gSum
+                    (
+                        magSqr
+                        (
+                            D().primitiveField()
+                          - D().prevIter().primitiveField()
+                        )
+                    )
+                );
+
+            // Norm of the solution
[…]
