### 2025-01-31 54c3969
Add wrapper for checkConvergence
