### 2025-02-01 #7407 @mohd-afeef-badri
Dear @philipcardiff and co-authors,

I have finally been able to review your JOSS paper and toolbox **solids4foam**. Below are my observations and suggestions:

### Installation and Setup ###
- I installed **solids4foam-v2.1** on Ubuntu 24.04 with **OpenFOAM 2312**. The installation proceeded without issues.
- I encountered the "fix file" issue, and the instructions provided:
  ```
  export S4F_NO_FILE_FIXES=1 && ./Allwmake -j
  ```
  allowed me to proceed with the installation smoothly. Thank you for anticipating such issues and providing fixes for them.
- The installation documentation on the website is generally clear, but I noticed a small error in the extraction commands:
  ```
  solids4foam-v2.1.zip: extracted with > unzip v2.1.tar.gz
  solids4foam-v2.1.tgz: extracted with > tar xzf unzip v2.1.tar.gz
  ```
  The authors may want to correct these commands with the appropriate filenames.
- One missing aspect in the installation documentation is guidance on setting up the **development branch** instead of **v2.1**. Additionally, information about **compatible OpenFOAM versions** for the development branch should be included in the **solids4foam-OpenFOAM compatibility table**.

### Testing ###
- All tests were successfully executed:
  ```
  Summary statistics
      Number of cases where a run was attempted: 45
      Number of cases where the solver failed: 0
      Number of cases where another command failed: 0

  All tests passed: enjoy solids4foam.
  ```

### Tutorials and Documentation ###
- I was able to run most of the tutorials successfully and found the documentation excellent. The tutorials provide the necessary mathematics, references, and detailed explanations of the code. I commend the authors for this high-quality work.
- However, I have a query regarding the **underlying mesh for simulations**:
  - Does **solids4foam** work exclusively with **hexahedral/quadrilateral** meshes, or does it support **tetrahedral/triangular** meshes as well?
  - What mesh formats are acceptable?
  - For complex geometries, can external **CAD-based meshing software** (e.g., Gmsh, Salome) be used?
- While the tutorials are well-structured, it would be valuable to include a **tutorial on custom case workflows**. This could involve:
  - Providing a given geometry.
  - Letting the user generate the mesh or walking the user through mesh generation.
  - Setting up groups/labels for boundary conditions.
  - Explaining how to apply boundary and initial conditions in **solids4foam**.
  - If such a tutorial already exists, it would be helpful to highlight it more prominently.
In my opinion, such a tutorial should boost the use of your toolbox further.

I appreciate the effort put into developing  and documenting **solids4foam** and look forward to the authors' responses.
