### 2025-02-02 a9eb8d5
Remove intermediate sparseMatrix class and use PETSc Mat directly
