+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -668,17 +668,19 @@ bool linGeomTotalDispSolid::evolve()
 }
 
 
-tmp<vectorField> linGeomTotalDispSolid::residualMomentum
+label linGeomTotalDispSolid::formResidual
 (
-    const volVectorField& D
+    PetscScalar *f,
+    const PetscScalar *x
 )
 {
-    // Prepare result
-    tmp<vectorField> tresidual(new vectorField(D.size(), vector::zero));
-    vectorField& residual = tresidual.ref();
+    // Copy x into the D field
+    volVectorField& D = const_cast<volVectorField&>(this->D());
+    vectorField& DI = D;
+    extractField(DI, x, solidModel::twoD());
 
     // Enforce the boundary conditions
-    const_cast<volVectorField&>(D).correctBoundaryConditions();
+    D.correctBoundaryConditions();
 
     // Update gradient of displacement
     mechanical().grad(D, gradD());
@@ -708,12 +710,14 @@ tmp<vectorField> linGeomTotalDispSolid::residualMomentum
     // F = div(sigma) + rho*g
     //     - rho*d2dt2(D) - dampingCoeff*rho*ddt(D) + stabilisationTerm
     // where, here, we roll the stabilisationTerm into the div(sigma)
-    residual =
+    vectorField residual
+    (
         fvc::div(mesh().magSf()*traction)
       + rho()
        *(
             g() - fvc::d2dt2(D) - dampingCoeff()*fvc::ddt(D)
-        );
+        )
+    );
 
     // Make residual extensive as fvc operators are intensive (per unit volume)
     residual *= mesh().V();
@@ -722,25 +726,26 @@ tmp<vectorField> linGeomTotalDispSolid::residualMomentum
     // Note that "source()" is already multiplied by the volumes
     residual -= fvOptions()(ds_, const_cast<volVectorField&>(D))().source();
 
-    return tresidual;
+    // Copy the residual into the f field
+    insertField(f, residual, solidModel::twoD());
+
+    return 0;
 }
 
 
-tmp<sparseMatrix> linGeomTotalDispSolid::JacobianMomentum
+label linGeomTotalDispSolid::formJacobian
 (
-    const volVectorField& D
+    Mat jac,
+    const PetscScalar *x
 )
 {
-    // Count the number of non-zeros for a Laplacian discretisation
-    // This equals the sum of one plus the number of internal faces for each,
-    // which can be calculated as nCells + 2*nInternalFaces
-    // Multiply by the blockSize since we will form the block matrix
-    const int blockSize = solidModel::twoD() ? 2 : 3;
-    const label numNonZeros =
-        blockSize*returnReduce
-        (
-            mesh().nCells() + 2.0*mesh().nInternalFaces(), sumOp<label>()
-        );
+    // Copy x into the D field
+    volVectorField& D = const_cast<volVectorField&>(this->D());
+    vectorField& DI = D;
+    extractField(DI, x, solidModel::twoD());
+
+    // Enforce the boundary conditions
+    D.correctBoundaryConditions();
 
     // Calculate a segregated approximation of the Jacobian
     fvVectorMatrix approxJ
@@ -757,157 +762,10 @@ tmp<sparseMatrix> linGeomTotalDispSolid::JacobianMomentum
     // Optional: under-relaxation of the linear system
     approxJ.relax();
 
-    // Convert fvMatrix matrix to sparseMatrix
+    // Convert fvMatrix matrix to PETSc matrix
+    insertFvMatrixIntoPETScMatrix(approxJ, jac);
 
-    // Initialise matrix
-    tmp<sparseMatrix> tmatrix(new sparseMatrix(numNonZeros));
[…]
