+++ src/solids4FoamModels/solidModels/nonLinGeomTotalLagTotalDispSolid/nonLinGeomTotalLagTotalDispSolid.C
@@ -25,6 +25,7 @@ License
 #include "solidTractionFvPatchVectorField.H"
 #include "fixedDisplacementZeroShearFvPatchVectorField.H"
 #include "symmetryFvPatchFields.H"
+#include "minEdgeLength.H"
 
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
@@ -170,156 +171,246 @@ bool nonLinGeomTotalLagTotalDispSolid::evolveImplicitSegregated()
     blockLduMatrix::debug = 0;
 #endif
 
-    int iCorr = 0;
-    scalar currentResidualNorm = 0;
-    scalar initialResidualNorm = 0;
-    scalar deltaXNorm = 0;
-    scalar xNorm = 0;
     const convergenceParameters convParam =
         readConvergenceParameters(solidModelDict());
 
-    Info<< "Solving the total Lagrangian form of the momentum equation for D"
-        << endl;
+    // Set initial pseudo time step based on Courant number
+    const scalar initCoNo = 100;
+    const dimensionedScalar minRho = min(rho());
+    dimensionedScalar pseudoDeltaT
+    (
+        "pseudoDeltaT",
+        dimTime,
+        initCoNo*minRho.value()*sqr(minEdgeLength(mesh()))
+        /(6*gMax(impKf_.primitiveField()))
+    );
+    const dimensionedScalar rT("rT", dimless/dimTime, 1.0);
 
-    // Undeformed unit normal vectors at the faces
-    const surfaceVectorField n(mesh().Sf()/mesh().magSf());
+    // Set the target number of outer iterations per pseudo time step
+    label nTarget = 100;
 
-    // Momentum equation loop
+    // Initialise the number of iterations as nTarget
+    label nIter = nTarget;
+
+    // Set the rate at which the pseudo time step changes
+    const scalar nIterExpon = 0.5;
+
+    // Store the pseudo old time D field
+    volVectorField DOldPseudo("DOldPseudo", D());
+
+    // Pseudo time step counter
+    label iPseudo = 0;
+
+    // Store initial residual at the start of the time step for convergence
+    // checking
+    scalar initialPseudoResidualNorm = 0;
+    scalar initialResidualNorm = 0;
+    scalar pseudoDeltaXNorm = 0;
+    scalar xNorm = 0;
+
+    // Pseudo time step loop
     do
     {
-        // Store fields for under-relaxation and residual calculation
-        D().storePrevIter();
+        DOldPseudo.storePrevIter();
 
-        // Calculate deformed area vectors and normals
-        const surfaceVectorField SfCurrent
-        (
-            fvc::interpolate(J_*Finv_.T()) & mesh().Sf()
-        );
-        const surfaceScalarField magSfCurrent(mag(SfCurrent));
-        const surfaceVectorField nCurrent(SfCurrent/magSfCurrent);
+        int iCorr = 0;
+        scalar currentResidualNorm = 0;
+        initialResidualNorm = 0;
+        scalar deltaXNorm = 0;
+        xNorm = 0;
 
-        // Traction vectors at the faces
-        surfaceVectorField traction(nCurrent & fvc::interpolate(sigma()));
+        Info<< "Solving the total Lagrangian form of the momentum equation for D"
+            << endl;
 
-        // Add stabilisation to the traction
-        // We add this before enforcing the traction condition as the stabilisation
-        // is set to zero on traction boundaries
-        // To-do: add a stabilisation traction function to momentumStabilisation
-        const scalar scaleFactor =
-            readScalar(stabilisation().dict().lookup("scaleFactor"));
-        const surfaceTensorField gradDf(fvc::interpolate(gradD()));
[…]
