### 2025-02-03 a8a7440
Replace insertField and extractField with insertFieldComponents and extractFieldComponents
