+++ src/solids4FoamModels/solidModels/nonLinGeomUpdatedLagSolid/nonLinGeomUpdatedLagSolid.C
@@ -566,7 +566,13 @@ label nonLinGeomUpdatedLagSolid::formResidual
     // Copy x into the DD field
     volVectorField& DD = const_cast<volVectorField&>(this->DD());
     vectorField& DDI = DD;
-    extractField(DDI, x, solidModel::twoD());
+    foamPetscSnesHelper::extractFieldComponents<vector>
+    (
+        x,
+        DDI,
+        solidModel::twoD() ? 2 : 3, // Block size of x
+        0                           // Location of first DI component
+    );
 
     // Enforce the boundary conditions
     DD.correctBoundaryConditions();
@@ -646,7 +652,13 @@ label nonLinGeomUpdatedLagSolid::formResidual
     //residual -= fvOptions()(ds_, const_cast<volVectorField&>(D))().source();
 
     // Copy the residual into the f field
-    insertField(f, residual, solidModel::twoD());
+    foamPetscSnesHelper::insertFieldComponents<vector>
+    (
+        residual,
+        f,
+        solidModel::twoD() ? 2 : 3, // Block size of x
+        0                           // Location of first component
+    );
 
     return 0;
 }
@@ -661,7 +673,13 @@ label nonLinGeomUpdatedLagSolid::formJacobian
     // Copy x into the DD field
     volVectorField& DD = const_cast<volVectorField&>(this->DD());
     vectorField& DDI = DD;
-    extractField(DDI, x, solidModel::twoD());
+    foamPetscSnesHelper::extractFieldComponents<vector>
+    (
+        x,
+        DDI,
+        solidModel::twoD() ? 2 : 3, // Block size of x
+        0                           // Location of first DI component
+    );
 
     // Enforce the boundary conditions
     DD.correctBoundaryConditions();
