+++ src/solids4FoamModels/solidModels/nonLinGeomTotalLagTotalDispSolid/nonLinGeomTotalLagTotalDispSolid.C
@@ -600,7 +600,13 @@ label nonLinGeomTotalLagTotalDispSolid::formResidual
     // Copy x into the D field
     volVectorField& D = const_cast<volVectorField&>(this->D());
     vectorField& DI = D;
-    extractField(DI, x, solidModel::twoD());
+    foamPetscSnesHelper::extractFieldComponents<vector>
+    (
+        x,
+        DI,
+        solidModel::twoD() ? 2 : 3, // Block size of x
+        0                           // Location of first DI component
+    );
 
     // Enforce the boundary conditions
     D.correctBoundaryConditions();
@@ -679,7 +685,13 @@ label nonLinGeomTotalLagTotalDispSolid::formResidual
     //residual -= fvOptions()(ds_, const_cast<volVectorField&>(D))().source();
 
     // Copy the residual into the f field
-    insertField(f, residual, solidModel::twoD());
+    foamPetscSnesHelper::insertFieldComponents<vector>
+    (
+        residual,
+        f,
+        solidModel::twoD() ? 2 : 3, // Block size of x
+        0                           // Location of first component
+    );
 
     return 0;
 }
@@ -694,7 +706,13 @@ label nonLinGeomTotalLagTotalDispSolid::formJacobian
     // Copy x into the D field
     volVectorField& D = const_cast<volVectorField&>(this->D());
     vectorField& DI = D;
-    extractField(DI, x, solidModel::twoD());
+    foamPetscSnesHelper::extractFieldComponents<vector>
+    (
+        x,
+        DI,
+        solidModel::twoD() ? 2 : 3, // Block size of x
+        0                           // Location of first DI component
+    );
 
     // Enforce the boundary conditions
     D.correctBoundaryConditions();
