+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -663,7 +663,7 @@ label linGeomTotalDispSolid::formResidual
     // Copy x into the D field
     volVectorField& D = const_cast<volVectorField&>(this->D());
     vectorField& DI = D;
-    foamPetscSnesHelper::extractFieldComponents<vector>
+    foamPetscSnesHelper::ExtractFieldComponents<vector>
     (
         x,
         DI,
@@ -719,7 +719,7 @@ label linGeomTotalDispSolid::formResidual
     residual -= fvOptions()(ds_, const_cast<volVectorField&>(D))().source();
 
     // Copy the residual into the f field
-    foamPetscSnesHelper::insertFieldComponents<vector>
+    foamPetscSnesHelper::InsertFieldComponents<vector>
     (
         residual,
         f,
@@ -740,7 +740,7 @@ label linGeomTotalDispSolid::formJacobian
     // Copy x into the D field
     volVectorField& D = const_cast<volVectorField&>(this->D());
     vectorField& DI = D;
-    foamPetscSnesHelper::extractFieldComponents<vector>
+    foamPetscSnesHelper::ExtractFieldComponents<vector>
     (
         x,
         DI,
@@ -767,7 +767,7 @@ label linGeomTotalDispSolid::formJacobian
     approxJ.relax();
 
     // Convert fvMatrix matrix to PETSc matrix
-    insertFvMatrixIntoPETScMatrix(approxJ, jac);
+    foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix(approxJ, jac, 0, 0);
 
     return 0;
 }
