+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelperTemplates.C
@@ -20,6 +20,7 @@ License
 #ifdef USE_PETSC
 
 #include "foamPetscSnesHelper.H"
+#include "processorFvPatch.H"
 
 namespace Foam
 {
@@ -28,7 +29,272 @@ namespace Foam
 
 
 template <class Type>
-void foamPetscSnesHelper::extractFieldComponents
+label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
+(
+    const fvMatrix<Type>& fvM,
+    Mat jac,
+    const label rowOffset,
+    const label colOffset
+) const
+{
+    // Initialise block coeff
+    const label nCoeffCmpts = blockSize_*blockSize_;
+    PetscScalar values[nCoeffCmpts];
+    for (int i = 0; i < nCoeffCmpts; ++i)
+    {
+        values[i] = 0;
+    }
+
+    // Get the number of components per field element from the traits (e.g.,
+    // 1 for scalar, 3 for vector, etc.)
+    const label vfBlockSize = pTraits<Type>::nComponents;
+
+    // Insert the diagonal
+    {
+        // Retrieve the matrix diagonal
+        const Field<Type> diag(fvM.DD());
+
+        // Insert the coeffs
+        forAll(diag, blockRowI)
+        {
+            // Obtain a pointer to the underlying scalar data in
+            // diag[blockRowI] (assumes contiguous storage)
+            const scalar* curDiagPtr =
+                reinterpret_cast<const scalar*>(diag[blockRowI].begin());
+
+            // Construct the diag block coefficient
+            for (label cmptI = 0; cmptI < vfBlockSize; ++cmptI)
+            {
+                for (label cmptJ = 0; cmptJ < vfBlockSize; ++cmptJ)
+                {
+                    if (cmptI == cmptJ)
+                    {
+                        values
+                        [
+                            (cmptI + rowOffset)*blockSize_ + cmptJ + colOffset
+                        ] = curDiagPtr[cmptI];
+                    }
+                }
+            }
+
+            // Get the global row index
+            const label globalBlockRowI =
+                foamPetscSnesHelper::globalCells().toGlobal(blockRowI);
+
+            // Insert block coefficient
+            CHKERRQ
+            (
+                MatSetValuesBlocked
+                (
+                    jac, 1, &globalBlockRowI, 1, &globalBlockRowI, values,
+                    ADD_VALUES
+                );
+            );
+        }
+    }
+
+    // Insert the off-diagonal
+    const fvMesh& mesh = fvM.psi().mesh();
+    {
+        const labelUList& own = mesh.owner();
+        const labelUList& nei = mesh.neighbour();
+        const scalarField& upper = fvM.upper();
+
+        forAll(own, faceI)
+        {
+            // Construct the upper block coefficient
+            const scalar upperScalarCoeff = upper[faceI];
+            for (label cmptI = 0; cmptI < vfBlockSize; ++cmptI)
+            {
+                for (label cmptJ = 0; cmptJ < vfBlockSize; ++cmptJ)
+                {
+                    if (cmptI == cmptJ)
+                    {
+                        values
+                        [
+                            (cmptI + rowOffset)*blockSize_ + cmptJ + colOffset
+                        ] = upperScalarCoeff;
+                    }
+                }
+            }
+
+            const label blockRowI = own[faceI];
+            const label blockColI = nei[faceI];
+
+            const label globalBlockRowI =
+                foamPetscSnesHelper::globalCells().toGlobal(blockRowI);
+            const label globalBlockColI =
+                foamPetscSnesHelper::globalCells().toGlobal(blockColI);
+
+            // Insert block coefficients
+            CHKERRQ
+            (
+                MatSetValuesBlocked
[…]
