+++ src/solids4FoamModels/solidModels/nonLinGeomTotalLagTotalDispSolid/nonLinGeomTotalLagTotalDispSolid.C
@@ -600,7 +600,7 @@ label nonLinGeomTotalLagTotalDispSolid::formResidual
     // Copy x into the D field
     volVectorField& D = const_cast<volVectorField&>(this->D());
     vectorField& DI = D;
-    foamPetscSnesHelper::extractFieldComponents<vector>
+    foamPetscSnesHelper::ExtractFieldComponents<vector>
     (
         x,
         DI,
@@ -685,7 +685,7 @@ label nonLinGeomTotalLagTotalDispSolid::formResidual
     //residual -= fvOptions()(ds_, const_cast<volVectorField&>(D))().source();
 
     // Copy the residual into the f field
-    foamPetscSnesHelper::insertFieldComponents<vector>
+    foamPetscSnesHelper::InsertFieldComponents<vector>
     (
         residual,
         f,
@@ -706,7 +706,7 @@ label nonLinGeomTotalLagTotalDispSolid::formJacobian
     // Copy x into the D field
     volVectorField& D = const_cast<volVectorField&>(this->D());
     vectorField& DI = D;
-    foamPetscSnesHelper::extractFieldComponents<vector>
+    foamPetscSnesHelper::ExtractFieldComponents<vector>
     (
         x,
         DI,
@@ -733,7 +733,7 @@ label nonLinGeomTotalLagTotalDispSolid::formJacobian
     approxJ.relax();
 
     // Convert fvMatrix matrix to PETSc matrix
-    insertFvMatrixIntoPETScMatrix(approxJ, jac);
+    foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix(approxJ, jac, 0, 0);
 
     return 0;
 }

[…]
