### 2025-02-03 1f22c8a
foamPetscSnesHelper: defined purely in terms of blockSize and mesh
