### 2025-02-04 d7af72f
Fix for 2-D indexing when using the new general foamPetscSnesHelper
