+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelperTemplates.C
@@ -34,20 +34,18 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
     const fvMatrix<Type>& fvM,
     Mat jac,
     const label rowOffset,
-    const label colOffset
+    const label colOffset,
+    const label nScalarEqns
 ) const
 {
     // Initialise block coeff
     const label nCoeffCmpts = blockSize_*blockSize_;
     PetscScalar values[nCoeffCmpts];
-    for (int i = 0; i < nCoeffCmpts; ++i)
-    {
-        values[i] = 0;
-    }
+    std::memset(values, 0, sizeof(values));
 
     // Get the number of components per field element from the traits (e.g.,
     // 1 for scalar, 3 for vector, etc.)
-    const label vfBlockSize = pTraits<Type>::nComponents;
+    //const label vfBlockSize = pTraits<Type>::nComponents;
 
     // Insert the diagonal
     {
@@ -63,9 +61,9 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
                 reinterpret_cast<const scalar*>(diag[blockRowI].begin());
 
             // Construct the diag block coefficient
-            for (label cmptI = 0; cmptI < vfBlockSize; ++cmptI)
+            for (label cmptI = 0; cmptI < nScalarEqns; ++cmptI)
             {
-                for (label cmptJ = 0; cmptJ < vfBlockSize; ++cmptJ)
+                for (label cmptJ = 0; cmptJ < nScalarEqns; ++cmptJ)
                 {
                     if (cmptI == cmptJ)
                     {
@@ -104,9 +102,9 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
         {
             // Construct the upper block coefficient
             const scalar upperScalarCoeff = upper[faceI];
-            for (label cmptI = 0; cmptI < vfBlockSize; ++cmptI)
+            for (label cmptI = 0; cmptI < nScalarEqns; ++cmptI)
             {
-                for (label cmptJ = 0; cmptJ < vfBlockSize; ++cmptJ)
+                for (label cmptJ = 0; cmptJ < nScalarEqns; ++cmptJ)
                 {
                     if (cmptI == cmptJ)
                     {
@@ -217,9 +215,9 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
                             intCoeffs[faceI].begin()
                         );
 
-                    for (label cmptI = 0; cmptI < vfBlockSize; ++cmptI)
+                    for (label cmptI = 0; cmptI < nScalarEqns; ++cmptI)
                     {
-                        for (label cmptJ = 0; cmptJ < vfBlockSize; ++cmptJ)
+                        for (label cmptJ = 0; cmptJ < nScalarEqns; ++cmptJ)
                         {
                             if (cmptI == cmptJ)
                             {
@@ -251,9 +249,9 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
                             neiCoeffs[faceI].begin()
                         );
 
-                    for (label cmptI = 0; cmptI < vfBlockSize; ++cmptI)
+                    for (label cmptI = 0; cmptI < nScalarEqns; ++cmptI)
                     {
-                        for (label cmptJ = 0; cmptJ < vfBlockSize; ++cmptJ)
+                        for (label cmptJ = 0; cmptJ < nScalarEqns; ++cmptJ)
                         {
                             if (cmptI == cmptJ)
                             {
[…]
