+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -677,8 +677,8 @@ label linGeomTotalDispSolid::formResidual
     (
         x,
         DI,
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first DI component
+        0,                          // Location of first DI component
+        solidModel::twoD() ? 2 : 3  // Number of components to extract
     );
 
     // Enforce the boundary conditions
@@ -733,8 +733,8 @@ label linGeomTotalDispSolid::formResidual
     (
         residual,
         f,
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first component
+        0,                          // Location of first DI component
+        solidModel::twoD() ? 2 : 3  // Number of components to extract
     );
 
     return 0;
@@ -754,8 +754,8 @@ label linGeomTotalDispSolid::formJacobian
     (
         x,
         DI,
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first DI component
+        0,                          // Location of first DI component
+        solidModel::twoD() ? 2 : 3  // Number of components to extract
     );
 
     // Enforce the boundary conditions
@@ -777,7 +777,10 @@ label linGeomTotalDispSolid::formJacobian
     approxJ.relax();
 
     // Convert fvMatrix matrix to PETSc matrix
-    foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix(approxJ, jac, 0, 0);
+    foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
+    (
+        approxJ, jac, 0, 0, solidModel::twoD() ? 2 : 3
+    );
 
     return 0;
 }
