+++ src/solids4FoamModels/solidModels/nonLinGeomUpdatedLagSolid/nonLinGeomUpdatedLagSolid.C
@@ -582,8 +582,8 @@ label nonLinGeomUpdatedLagSolid::formResidual
     (
         x,
         DDI,
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first DI component
+        0,                          // Location of first DDI component
+        solidModel::twoD() ? 2 : 3  // Number of components to extract
     );
 
     // Enforce the boundary conditions
@@ -668,8 +668,8 @@ label nonLinGeomUpdatedLagSolid::formResidual
     (
         residual,
         f,
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first component
+        0,                          // Location of first DI component
+        solidModel::twoD() ? 2 : 3  // Number of components to extract
     );
 
     return 0;
@@ -689,8 +689,8 @@ label nonLinGeomUpdatedLagSolid::formJacobian
     (
         x,
         DDI,
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first DI component
+        0,                          // Location of first DDI component
+        solidModel::twoD() ? 2 : 3  // Number of components to extract
     );
 
     // Enforce the boundary conditions
@@ -712,7 +712,10 @@ label nonLinGeomUpdatedLagSolid::formJacobian
     approxJ.relax();
 
     // Convert fvMatrix matrix to PETSc matrix
-    foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix(approxJ, jac, 0, 0);
+    foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
+    (
+        approxJ, jac, 0, 0, solidModel::twoD() ? 2 : 3
+    );
 
     return 0;
 }

[…]
