+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.H
@@ -85,6 +85,16 @@ class linGeomTotalDispSolid
         //- Predict new time-step fields using the velocity field
         const Switch predictor_;
 
+        //- Solve pressure as an independent unknown
+        const Switch solvePressure_;
+
+        //- Pressure
+        //  Hydrostatic component of the stress tensor: p = -hyd(sigma)
+        autoPtr<volScalarField> pPtr_;
+
+        //- Number of unknowns per cell, used by PETSc SNES solver
+        const label blockSize_;
+
         //- Field to provide fvOptions with the correct dimensions
         volScalarField ds_;
 
@@ -164,6 +174,9 @@ public:
                 // This model solves for D
                 return D();
             }
+
+            //- Non-const access to the pressure
+            volScalarField& p();
 
 
         // Edit
