+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -293,8 +293,8 @@ bool linGeomTotalDispSolid::evolveSnes()
         (
             D().primitiveFieldRef(),
             foamPetscSnesHelper::solution(),
-            solidModel::twoD() ? 2 : 3, // Block size of x
-            0                           // Location of first component
+            blockSize_,
+            0
         );
     }
 
@@ -307,9 +307,22 @@ bool linGeomTotalDispSolid::evolveSnes()
     (
         foamPetscSnesHelper::solution(),
         D().primitiveFieldRef(),
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first component
+        blockSize_,
+        0
     );
+
+    if (solvePressure_)
+    {
+        // Map the PETSc solution to the p field
+        // p is located in the 4th component
+        foamPetscSnesHelper::ExtractFieldComponents<scalar>
+        (
+            foamPetscSnesHelper::solution(),
+            p().primitiveFieldRef(),
+            blockSize_,
+            3
+        );
+    }
 
     // Interpolate cell displacements to vertices
     mechanical().interpolate(D(), gradD(), pointD());
@@ -439,7 +452,16 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
             )
         ),
         mesh(),
-        solidModel::twoD() ? 2 : 3,
+        solidModelDict().lookupOrDefault<Switch>("solvePressure", false) // looked twice?
+      ? label(solidModel::twoD() ? 3 : 4)
+      : label(solidModel::twoD() ? 2 : 3),
+        solidModelDict().lookupOrDefault<Switch>("solvePressure", false)
+      ? (
+            solidModel::twoD()
+          ? labelListList({ {0, 1}, {2} }) // 0-1: momentum. 2: pressure
+          : labelListList({ {0, 1, 2}, {3} }) // 0-2: momentum. 3: pressure
+        )
+      : labelListList({}), // No need to assign fields
         solidModelDict().lookupOrDefault<Switch>("stopOnPetscError", true),
         bool(solutionAlg() == solutionAlgorithm::PETSC_SNES)
     ),
@@ -460,6 +482,17 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
         dimensionedVector("zero", dimLength/pow(dimTime, 2), vector::zero)
     ),
     predictor_(solidModelDict().lookupOrDefault<Switch>("predictor", false)),
+    solvePressure_
+    (
+        solidModelDict().lookupOrDefault<Switch>("solvePressure", false)
+    ),
+    pPtr_(),
+    blockSize_
+    (
+        solvePressure_
+      ? label(solidModel::twoD() ? 3 : 4)
+      : label(solidModel::twoD() ? 2 : 3)
+    ),
     ds_
     (
         IOobject
@@ -483,6 +516,25 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
     D().correctBoundaryConditions();
     D().storePrevIter();
     mechanical().grad(D(), gradD());
+
+    Info<< "solvePressure = " << solvePressure_ << endl;
+
+    if (solvePressure_)
+    {
+        if (solutionAlg() != solutionAlgorithm::PETSC_SNES)
+        {
+            FatalErrorInFunction
+                << "The solution algorithm must be "
+                << solidModel::solutionAlgorithmNames_
+                   [
+                       solidModel::solutionAlgorithm::PETSC_SNES
+                   ]
+                << " when solvePressure is enabled" << abort(FatalError);
+        }
+
+        // Ensure p is created
+        p();
+    }
 
     if (predictor_)
     {
[…]
