### 2025-02-06 8383fa6
Added correct function with Gauss point stress and grad
