### 2025-02-06 3deb864
linGeomTotalDispSolid pressure WIP
