+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -311,6 +311,8 @@ bool linGeomTotalDispSolid::evolveSnes()
         0
     );
 
+    D().correctBoundaryConditions();
+
     if (solvePressure_)
     {
         // Map the PETSc solution to the p field
@@ -320,8 +322,10 @@ bool linGeomTotalDispSolid::evolveSnes()
             foamPetscSnesHelper::solution(),
             p().primitiveFieldRef(),
             blockSize_,
-            3
+            blockSize_ - 1
         );
+
+        p().correctBoundaryConditions();
     }
 
     // Interpolate cell displacements to vertices
@@ -508,6 +512,7 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
     )
 {
     DisRequired();
+    Info<< "blockSize = " << blockSize_ << endl;
 
     // Force all required old-time fields to be created
     fvm::d2dt2(D());
@@ -749,6 +754,8 @@ label linGeomTotalDispSolid::formResidual
     const PetscScalar *x
 )
 {
+    const fvMesh& mesh = this->mesh();
+
     // Copy x into the D field
     volVectorField& D = const_cast<volVectorField&>(this->D());
     vectorField& DI = D;
@@ -757,16 +764,19 @@ label linGeomTotalDispSolid::formResidual
     // Enforce the boundary conditions
     D.correctBoundaryConditions();
 
-    // Copy x into the p field
-    volScalarField& p = const_cast<volScalarField&>(this->p());
-    scalarField& pI = p;
-    foamPetscSnesHelper::ExtractFieldComponents<scalar>
-    (
-        x, pI, blockSize_, blockSize_ - 1
-    );
+    if (solvePressure_)
+    {
+        // Copy x into the p field
+        volScalarField& p = const_cast<volScalarField&>(this->p());
+        scalarField& pI = p;
+        foamPetscSnesHelper::ExtractFieldComponents<scalar>
+        (
+            x, pI, blockSize_, blockSize_ - 1
+        );
 
-    // Enforce the boundary conditions
-    p.correctBoundaryConditions();
+        // Enforce the boundary conditions
+        p.correctBoundaryConditions();
+    }
 
     // Update gradient of displacement
     mechanical().grad(D, gradD());
@@ -775,7 +785,6 @@ label linGeomTotalDispSolid::formResidual
     mechanical().correct(sigma());
 
     // Unit normal vectors at the faces
-    const fvMesh& mesh = this->mesh();
     const surfaceVectorField n(mesh.Sf()/mesh.magSf());
 
     // Traction vectors at the faces
@@ -821,30 +830,32 @@ label linGeomTotalDispSolid::formResidual
 
     if (solvePressure_)
     {
+        volScalarField& p = const_cast<volScalarField&>(this->p());
+
         // Calculate pressure equation residual
         // Res = p/k + div(D) - gamma*laplacian(p) + gamma*div(grad(p))
         // where
         //   - k: bulk modulus
         //   - gamma: controls the amount of smoothing
 
-        const dimensionedScalar val("val", dimPressure, 1e6);
         const dimensionedScalar pressureSmoothingCoeff
         (
             solidModelDict().lookup("pressureSmoothingCoeff")
         );
         scalarField pressureResidual
         (
-          //   p/mechanical().bulkModulus()
-          // + tr(gradD())
-          // // + fvc::div(D)
           - p
           + fvc::laplacian(pressureSmoothingCoeff, p, "laplacian(Dp,p)")
-          //- mechanical().bulkModulus()*tr(gradD())
-          + val
+          - fvc::div(pressureSmoothingCoeff*fvc::grad(p))
[…]
