### 2025-02-10 238d324
resolved bug for gauss points coeffs calculation
