### 2025-02-11 4c916c6
calculation of sigma at Gauss points
