### 2025-02-11 c25c652
Added divGaussPoints skeleton
