### 2025-02-11 9e37cef
update divGaussPoints function
