+++ src/solids4FoamModels/numerics/fvc/fvcSurfaceGaussQuadIntegrate.C
@@ -29,55 +29,18 @@ namespace fvc
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
-template<class Type>
-void surfaceGaussQuadIntegrate
-(
-    Field<Type>& ivf,
-    const GeometricField<Type, fvsPatchField, surfaceMesh>& ssf
-)
-{
-    const fvMesh& mesh = ssf.mesh();
-
-    const labelUList& owner = mesh.owner();
-    const labelUList& neighbour = mesh.neighbour();
-
-    const Field<Type>& issf = ssf;
-
-    // Here I need to perform Gaussian Quadrature integration. For the first
-    // iteration Gauss points and weights can be argument in surfaceGaussQuadIntegrate
-    // function.
-
-
-    forAll(owner, facei)
-    {
-        ivf[owner[facei]] += issf[facei];
-        ivf[neighbour[facei]] -= issf[facei];
-    }
-
-    forAll(mesh.boundary(), patchi)
-    {
-        const labelUList& pFaceCells =
-            mesh.boundary()[patchi].faceCells();
-
-        const fvsPatchField<Type>& pssf = ssf.boundaryField()[patchi];
-
-        forAll(mesh.boundary()[patchi], facei)
-        {
-            ivf[pFaceCells[facei]] += pssf[facei];
-        }
-    }
-
-    ivf /= mesh.Vsc();
-}
-
 
 template<class Type>
 tmp<GeometricField<Type, fvPatchField, volMesh>>
 surfaceGaussQuadIntegrate
 (
-    const GeometricField<Type, fvsPatchField, surfaceMesh>& ssf
-)
+    const List<List<symmTensor>>& sigmaGP,
+    const List<List<scalar>>>& gpW
+);
 {
+    // Here I need to perform Gaussian Quadrature integration. For the first
+    // iteration Gauss points and weights can be argument in surfaceGaussQuadIntegrate
+    // function.
     const fvMesh& mesh = ssf.mesh();
 
     tmp<GeometricField<Type, fvPatchField, volMesh>> tvf
@@ -99,27 +62,100 @@ surfaceGaussQuadIntegrate
     );
     GeometricField<Type, fvPatchField, volMesh>& vf = tvf.ref();
 
-    surfaceIntegrate(vf.primitiveFieldRef(), ssf);
-    vf.correctBoundaryConditions();
+    const labelUList& owner = mesh.owner();
+    const labelUList& neighbour = mesh.neighbour();
+
+    //const Field<Type>& issf = vf;
 
     return tvf;
 }
 
+// template<class Type>
+// void surfaceGaussQuadIntegrate
+// (
+//     Field<Type>& ivf,
+//     const GeometricField<Type, fvsPatchField, surfaceMesh>& ssf
+// )
+// {
+//     const fvMesh& mesh = ssf.mesh();
 
-template<class Type>
-tmp<GeometricField<Type, fvPatchField, volMesh>>
-surfaceGaussQuadIntegrate
-(
-    const tmp<GeometricField<Type, fvsPatchField, surfaceMesh>>& tssf
-)
-{
-    tmp<GeometricField<Type, fvPatchField, volMesh>> tvf
-    (
-        fvc::surfaceIntegrate(tssf())
-    );
-    tssf.clear();
-    return tvf;
-}
+//     const labelUList& owner = mesh.owner();
+//     const labelUList& neighbour = mesh.neighbour();
+
+//     const Field<Type>& issf = ssf;
+
+//     forAll(owner, facei)
+//     {
+//         ivf[owner[facei]] += issf[facei];
+//         ivf[neighbour[facei]] -= issf[facei];
+//     }
+
[…]
