+++ src/solids4FoamModels/numerics/fvc/fvcDivGaussPoints.C
@@ -45,11 +45,13 @@ namespace fvc
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
+
 template<class Type>
 tmp<GeometricField<Type, fvPatchField, volMesh>>
 divGaussPoints
 (
-    const GeometricField<Type, fvsPatchField, surfaceMesh>& ssf
+    const List<List<symmTensor>>& sigmaGP,
+    const List<List<scalar>>>& gpW
 )
 {
     return tmp<GeometricField<Type, fvPatchField, volMesh>>
@@ -57,244 +59,262 @@ divGaussPoints
         new GeometricField<Type, fvPatchField, volMesh>
         (
             "divGaussPoints("+ssf.name()+')',
-            fvc::surfaceIntegrate(ssf)
+            fvc::surfaceGaussQuadIntegrate(sigamGP, gpW)
         )
     );
 }
 
 
-template<class Type>
-tmp<GeometricField<Type, fvPatchField, volMesh>>
-divGaussPoints
-(
-    const tmp<GeometricField<Type, fvsPatchField, surfaceMesh>>& tssf
-)
-{
-    tmp<GeometricField<Type, fvPatchField, volMesh>> DivGaussPoints(fvc::divGaussPoints(tssf()));
-    tssf.clear();
-    return DivGaussPoints;
-}
+// template<class Type>
+// tmp<GeometricField<Type, fvPatchField, volMesh>>
+// divGaussPoints
+// (
+//     const GeometricField<Type, fvsPatchField, surfaceMesh>& ssf
+// )
+// {
+//     return tmp<GeometricField<Type, fvPatchField, volMesh>>
+//     (
+//         new GeometricField<Type, fvPatchField, volMesh>
+//         (
+//             "divGaussPoints("+ssf.name()+')',
+//             fvc::surfaceIntegrate(ssf)
+//         )
+//     );
+// }
 
 
-template<class Type>
-tmp
-<
-    GeometricField
-    <
-        typename innerProduct<vector, Type>::type, fvPatchField, volMesh
-    >
->
-divGaussPoints
-(
-    const GeometricField<Type, fvPatchField, volMesh>& vf,
-    const word& name
-)
-{
-    return fv::divGaussPointsScheme<Type>::New
-    (
-        vf.mesh(), vf.mesh().divScheme(name)
-    ).ref().fvcDivGaussPoints(vf);
-}
+// template<class Type>
+// tmp<GeometricField<Type, fvPatchField, volMesh>>
+// divGaussPoints
+// (
+//     const tmp<GeometricField<Type, fvsPatchField, surfaceMesh>>& tssf
+// )
+// {
+//     tmp<GeometricField<Type, fvPatchField, volMesh>> DivGaussPoints(fvc::divGaussPoints(tssf()));
+//     tssf.clear();
+//     return DivGaussPoints;
+// }
 
 
-template<class Type>
-tmp
-<
-    GeometricField
-    <
-        typename innerProduct<vector, Type>::type, fvPatchField, volMesh
-    >
->
-divGaussPoints
-(
-    const tmp<GeometricField<Type, fvPatchField, volMesh>>& tvvf,
-    const word& name
-)
-{
-    typedef typename innerProduct<vector, Type>::type DivGaussPointsType;
-    tmp<GeometricField<DivGaussPointsType, fvPatchField, volMesh>> DivGaussPoints
-    (
-        fvc::divGaussPoints(tvvf(), name)
-    );
-    tvvf.clear();
-    return DivGaussPoints;
-}
-
-template<class Type>
-tmp
-<
-    GeometricField
-    <
[…]

[…]
