### 2025-02-15 5817549
Use shear modulus for momentum coefficients and divide p equation by kappa
