+++ src/solids4FoamModels/materialModels/mechanicalModel/mechanicalModel.H
@@ -153,6 +153,9 @@ public:
             //- Return bulk modulus
             tmp<volScalarField> bulkModulus() const;
 
+            //- Return shear modulus
+            tmp<volScalarField> shearModulus() const;
+
             //- Return material residual i.e. a measured of how convergence of
             //  the material model
             scalar residual();
