+++ src/solids4FoamModels/materialModels/mechanicalModel/mechanicalLaws/linearGeometryLaws/linearElastic/linearElastic.H
@@ -106,6 +106,9 @@ public:
         //- Return bulk modulus
         virtual tmp<volScalarField> bulkModulus() const;
 
+        //- Return shear modulus
+        virtual tmp<volScalarField> shearModulus() const;
+
         //- Return the implicit stiffness
         //  This is the diffusivity for the Laplacian term
         virtual tmp<volScalarField> impK() const;
