### 2025-02-15 365891d
Parallelisation of pressure approach - still only one set of LS vectors
