+++ src/solids4FoamModels/numerics/leastSquaresS4fGrad/leastSquaresS4fVectors.C
@@ -69,6 +69,8 @@ Foam::leastSquaresS4fVectors::leastSquaresS4fVectors
         dimensionedVector(dimless/dimLength, Zero)
     )
 {
+    InfoInFunction
+        << "Constructing new LS vecs!" << endl;
     calcLeastSquaresVectors();
 }
 
