+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelper.C
@@ -23,6 +23,8 @@ License
 #include "SparseMatrixTemplate.H"
 #include "processorFvPatch.H"
 #include "leastSquaresS4fVectors.H"
+#include "symmetryFvPatchFields.H"
+#include "symmetryPlaneFvPatchFields.H"
 
 // * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
 
@@ -200,120 +202,149 @@ defineTypeNameAndDebug(foamPetscSnesHelper, 0);
 
 // * * * * * * * * * * * * * * * Private Function  * * * * * * * * * * * * * //
 
-label foamPetscSnesHelper::InsertFvmGradIntoPETScMatrix
-(
-    Mat jac,
-    const label rowOffset,
-    const label colOffset,
-    const label nScalarEqns
-) const
+void foamPetscSnesHelper::makeNeiProcFields() const
 {
-    // Get reference to least square vectors
+    if
+    (
+        neiProcGlobalIDs_.size() != 0
+     || neiProcVolumes_.size() != 0
+     || neiProcLs_.size() != 0
+    )
+    {
+        FatalErrorInFunction
+            << "Pointers already set!" << abort(FatalError);
+    }
+
     const fvMesh& mesh = mesh_;
-    const boolList useBoundaryFaceValues(mesh.boundary().size(), false);
-    const leastSquaresS4fVectors& lsv =
-        leastSquaresS4fVectors::New(mesh, useBoundaryFaceValues);
 
-    const surfaceVectorField& ownLs = lsv.pVectors();
-    const surfaceVectorField& neiLs = lsv.nVectors();
+    neiProcGlobalIDs_.setSize(mesh.boundaryMesh().size());
+    neiProcVolumes_.setSize(mesh.boundaryMesh().size());
+    neiProcLs_.setSize(mesh.boundaryMesh().size());
 
-    const labelUList& own = mesh.owner();
-    const labelUList& nei = mesh.neighbour();
+    PtrList<labelList>& neiProcGlobalIDs = neiProcGlobalIDs_;
+    PtrList<scalarField>& neiProcVolumes = neiProcVolumes_;
+    PtrList<vectorField>& neiProcLs = neiProcLs_;
 
     const scalarField& VI = mesh.V();
 
-    // Initialise the block coefficient
-    const label nCoeffCmpts = blockSize_*blockSize_;
-    PetscScalar values[nCoeffCmpts];
-    std::memset(values, 0, sizeof(values));
+    // Get reference to least square vectors
+    const boolList useBoundaryFaceValues(mesh.boundary().size(), false);
+    const leastSquaresS4fVectors& lsv =
+        leastSquaresS4fVectors::New(mesh, useBoundaryFaceValues);
+    const surfaceVectorField& ownLs = lsv.pVectors();
 
-    forAll(own, faceI)
+    // Send the data
+    forAll(mesh.boundary(), patchI)
     {
-        // Local block row ID
-        const label ownFaceI = own[faceI];
+        const fvPatch& fp = mesh.boundary()[patchI];
 
-        // Local block column ID
-        const label neiFaceI = nei[faceI];
-
-        // Global block row ID
-        const label globalBlockRowI =
-            foamPetscSnesHelper::globalCells().toGlobal(ownFaceI);
-
-        // Global block column ID
-        const label globalBlockColI =
-            foamPetscSnesHelper::globalCells().toGlobal(neiFaceI);
-
-        // Explicit gradient
-        // Type deltaVsf = vsf[neiFacei] - vsf[ownFacei];
-        // lsGrad[ownFacei] += ownLs[facei]*deltaVsf;
[…]
