+++ src/solids4FoamModels/numerics/leastSquaresS4fGrad/leastSquaresS4fGrad.C
@@ -77,7 +77,7 @@ Foam::fv::leastSquaresS4fGrad<Type>::calcGrad
     {
         if (isA<solidTractionFvPatchVectorField>(vsf.boundaryField()[patchI]))
         {
-            useBoundaryFaceValues = false;
+            useBoundaryFaceValues[patchI] = false;
         }
     }
 
