+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelper.H
@@ -55,6 +55,7 @@ SourceFiles
 #include "volFields.H"
 #include "sparseMatrix.H"
 #include "fvMatrices.H"
+#include "leastSquaresS4fVectors.H"
 
 
 // * * * * * * * * * * * * * * External Structs  * * * * * * * * * * * * * * //
@@ -137,6 +138,12 @@ class foamPetscSnesHelper
         //- Least squares vectors for the neighbouring processor patch cells
         mutable PtrList<vectorField> neiProcLs_;
 
+        //- Least squares vectors
+        //  Note: these are different than the ones used for the displacement
+        //  field as different boundary patch face values may be included
+        mutable autoPtr<leastSquaresS4fVectors> lsVectorsPtr_;
+
+
     // Private Member Functions
 
         //- Make the neighbour processor patch fields
@@ -150,6 +157,12 @@ class foamPetscSnesHelper
 
         //- Return a reference to neiProcLs_
         const PtrList<vectorField>& neiProcLs() const;
+
+        //- Make least squares vectors
+        void makeLsVectors(const volScalarField& p) const;
+
+        //- Return a reference to the least squares vectors
+        const leastSquaresS4fVectors& lsVectors(const volScalarField& p) const;
 
         //- Disallow default bitwise copy construct
         foamPetscSnesHelper(const foamPetscSnesHelper&);
@@ -255,6 +268,7 @@ public:
             //  Note that the -grad(p) is insert, not grad(p)
             label InsertFvmGradIntoPETScMatrix
             (
+                const volScalarField& p,
                 Mat jac,
                 const label rowOffset,
                 const label colOffset,
