+++ src/solids4FoamModels/materialModels/mechanicalModel/mechanicalLaws/nonLinearGeometryLaws/GuccioneElastic/GuccioneElastic.H
@@ -170,6 +170,9 @@ public:
         //- Return the bulk modulus field
         virtual tmp<volScalarField> bulkModulus() const;
 
+        //- Return the shear modulus field
+        virtual tmp<volScalarField> shearModulus() const;
+
         //- Calculate the stress
         virtual void correct(volSymmTensorField& sigma);
 
