+++ src/solids4FoamModels/solidModels/nonLinGeomTotalLagTotalDispSolid/nonLinGeomTotalLagTotalDispSolid.C
@@ -388,6 +388,62 @@ bool nonLinGeomTotalLagTotalDispSolid::evolveSnes()
 }
 
 
+void nonLinGeomTotalLagTotalDispSolid::makePDiffusivity() const
+{
+    if (pDiffusivityPtr_.valid())
+    {
+        FatalErrorInFunction
+            << "Pointer already set!" << abort(FatalError);
+    }
+
+    const scalar pressureSmoothingCoeff
+    (
+        readScalar(solidModelDict().lookup("pressureSmoothingCoeff"))
+    );
+
+    fvVectorMatrix approxJ
+    (
+        fvm::laplacian(impKf_, D(), "laplacian(DD,D)")
+      - rho()*fvm::d2dt2(D())
+    );
+
+    if (dampingCoeff().value() > SMALL)
+    {
+        approxJ -= dampingCoeff()*rho()*fvm::ddt(D());
+    }
+
+    // Optional: under-relaxation of the linear system
+    approxJ.relax();
+
+    pDiffusivityPtr_.set
+    (
+        new surfaceScalarField
+        (
+            IOobject
+            (
+                "pDiffusivity",
+                mesh().time().timeName(),
+                mesh(),
+                IOobject::NO_READ,
+                IOobject::NO_WRITE
+            ),
+            -pressureSmoothingCoeff*impKf_/fvc::interpolate(approxJ.A())
+        )
+    );
+}
+
+
+const surfaceScalarField& nonLinGeomTotalLagTotalDispSolid::pDiffusivity() const
+{
+    if (pDiffusivityPtr_.empty())
+    {
+        makePDiffusivity();
+    }
+
+    return pDiffusivityPtr_.ref();
+}
+
+
 // * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
 
 nonLinGeomTotalLagTotalDispSolid::nonLinGeomTotalLagTotalDispSolid
@@ -407,7 +463,16 @@ nonLinGeomTotalLagTotalDispSolid::nonLinGeomTotalLagTotalDispSolid
             )
         ),
         mesh(),
-        solidModel::twoD() ? 2 : 3,
+        solvePressure()
+      ? label(solidModel::twoD() ? 3 : 4)
+      : label(solidModel::twoD() ? 2 : 3),
+        solvePressure()
+      ? (
+            solidModel::twoD()
+          ? labelListList({ {0, 1}, {2} })    // 0-1: momentum. 2: pressure
+          : labelListList({ {0, 1, 2}, {3} }) // 0-2: momentum. 3: pressure
+        )
+      : labelListList({}), // No need to assign fields
         solidModelDict().lookupOrDefault<Switch>("stopOnPetscError", true),
         bool(solutionAlg() == solutionAlgorithm::PETSC_SNES)
     ),
@@ -460,15 +525,46 @@ nonLinGeomTotalLagTotalDispSolid::nonLinGeomTotalLagTotalDispSolid
         ),
         fvc::d2dt2(D())
     ),
-    impK_(mechanical().impK()),
-    impKf_(mechanical().impKf()),
+    impK_
+    (
+        solvePressure()
+      ? 2.0*mechanical().shearModulus()
+      : mechanical().impK()
+    ),
+    impKf_(fvc::interpolate(impK_)),
     rImpK_(1.0/impK_),
-    predictor_(solidModelDict().lookupOrDefault<Switch>("predictor", false))
+    pDiffusivityPtr_(),
+    predictor_(solidModelDict().lookupOrDefault<Switch>("predictor", false)),
+    blockSize_
+    (
+        solvePressure()
+      ? label(solidModel::twoD() ? 3 : 4)
[…]
