### 2025-02-16 5710379
Add newtonIcoFluid: working for cylinderInChannel case
