+++ src/solids4FoamModels/fluidModels/newtonIcoFluid/newtonIcoFluid.C
@@ -1,0 +1,574 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+\*---------------------------------------------------------------------------*/
+
+#include "newtonIcoFluid.H"
+#include "addToRunTimeSelectionTable.H"
+#include "CorrectPhi.H"
+#include "fvc.H"
+#include "fvm.H"
+#include "constrainHbyA.H"
+#include "constrainPressure.H"
+#include "findRefCell.H"
+#include "elasticSlipWallVelocityFvPatchVectorField.H"
+#include "elasticWallVelocityFvPatchVectorField.H"
+#include "elasticWallPressureFvPatchScalarField.H"
+#include "movingWallPressureFvPatchScalarField.H"
+#include "EulerDdtScheme.H"
+#include "backwardDdtScheme.H"
+#include "thermalRobinFvPatchScalarField.H"
+
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+namespace Foam
+{
+
+namespace fluidModels
+{
+
+// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
+
+defineTypeNameAndDebug(newtonIcoFluid, 0);
+addToRunTimeSelectionTable(fluidModel, newtonIcoFluid, dictionary);
+
+// * * * * * * * * * * * * * * * Private Members * * * * * * * * * * * * * * //
+
+void newtonIcoFluid::makeRAUf() const
+{
+    if (rAUfPtr_.valid())
+    {
+        FatalErrorInFunction
+            << "Pointer already set!" << abort(FatalError);
+    }
+
+    volVectorField& U = const_cast<volVectorField&>(this->U());
+
+    fvVectorMatrix UEqn
+    (
+        fvm::ddt(U)
+      + fvm::div(phi(), U)
+      - fvc::laplacian(turbulence_->nuEff(), U)
+        //+ turbulence_->divDevReff(U)
+    );
+
+    UEqn.relax();
+
+    const scalar pressureSmoothingCoeff
+    (
+        readScalar(fluidProperties().lookup("pressureSmoothingCoeff"))
+    );
+
+    rAUfPtr_.set
+    (
+        new surfaceScalarField
+        (
+            IOobject
+            (
+                "rAUf",
+                runTime().timeName(),
+                mesh(),
+                IOobject::NO_READ,
+                IOobject::NO_WRITE
+            ),
+            pressureSmoothingCoeff*fvc::interpolate(1.0/UEqn.A())
+        )
+    );
+}
+
+
+const surfaceScalarField& newtonIcoFluid::rAUf() const
+{
+    if (rAUfPtr_.empty())
+    {
+        makeRAUf();
+    }
+
+    return rAUfPtr_.ref();
+}
+
+// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
+
+newtonIcoFluid::newtonIcoFluid
+(
+    Time& runTime,
+    const word& region
+)
+:
[…]
