+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelper.C
@@ -24,6 +24,7 @@ License
 #include "processorFvPatch.H"
 #include "symmetryFvPatchFields.H"
 #include "symmetryPlaneFvPatchFields.H"
+#include "leastSquaresVectors.H"
 
 // * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
 
@@ -178,6 +179,7 @@ PetscErrorCode formJacobianFoamPetscSnesHelper
     // Complete matrix assembly
     CHKERRQ(MatAssemblyBegin(B, MAT_FINAL_ASSEMBLY));
     CHKERRQ(MatAssemblyEnd(B, MAT_FINAL_ASSEMBLY));
+
     if (jac != B)
     {
         CHKERRQ(MatAssemblyBegin(jac, MAT_FINAL_ASSEMBLY));
@@ -207,7 +209,6 @@ void foamPetscSnesHelper::makeNeiProcFields() const
     (
         neiProcGlobalIDs_.size() != 0
      || neiProcVolumes_.size() != 0
-     || neiProcLs_.size() != 0
     )
     {
         FatalErrorInFunction
@@ -218,19 +219,11 @@ void foamPetscSnesHelper::makeNeiProcFields() const
 
     neiProcGlobalIDs_.setSize(mesh.boundaryMesh().size());
     neiProcVolumes_.setSize(mesh.boundaryMesh().size());
-    neiProcLs_.setSize(mesh.boundaryMesh().size());
 
     PtrList<labelList>& neiProcGlobalIDs = neiProcGlobalIDs_;
     PtrList<scalarField>& neiProcVolumes = neiProcVolumes_;
-    PtrList<vectorField>& neiProcLs = neiProcLs_;
 
     const scalarField& VI = mesh.V();
-
-    // Get reference to least square vectors
-    const boolList useBoundaryFaceValues(mesh.boundary().size(), false);
-    const leastSquaresS4fVectors& lsv =
-        leastSquaresS4fVectors::New(mesh, useBoundaryFaceValues);
-    const surfaceVectorField& ownLs = lsv.pVectors();
 
     // Send the data
     forAll(mesh.boundary(), patchI)
@@ -266,13 +259,6 @@ void foamPetscSnesHelper::makeNeiProcFields() const
             (
                 Pstream::commsTypes::blocking, patchVols
             );
-
-            // Send patch least squares vectors to the neighbour proc
-            const vectorField& patchLs = ownLs.boundaryField()[patchI];
-            procPatch.send
-            (
-                Pstream::commsTypes::blocking, patchLs
-            );
         }
     }
 
@@ -301,14 +287,6 @@ void foamPetscSnesHelper::makeNeiProcFields() const
             (
                 Pstream::commsTypes::blocking, patchNeiVols
             );
-
-            // Receive patch least squares vectors from the neighbour proc
-            neiProcLs.set(patchI, new vectorField(fp.size()));
-            vectorField& patchNeiLs = neiProcLs_[patchI];
-            procPatch.receive
-            (
-                Pstream::commsTypes::blocking, patchNeiLs
-            );
         }
     }
 }
@@ -336,25 +314,19 @@ const PtrList<scalarField>& foamPetscSnesHelper::neiProcVolumes() const
 }
 
 
-const PtrList<vectorField>& foamPetscSnesHelper::neiProcLs() const
+const leastSquaresS4fVectors& foamPetscSnesHelper::lsVectors
+(
+    const volScalarField& p
+) const
 {
-    if (neiProcLs_.size() == 0)
+    // Lookup and return vectors if they exist for this field
+    // Otherwise create and return them
[…]
