### 2025-02-19 e2b0870
newtonIcoFluid working with Rhie-Chow smoothing: good convergence with JFNK
