### 2025-02-19 374281e
Move generic coefficient insertion routines to foamPetscSnesHelper
