+++ src/solids4FoamModels/fluidModels/newtonIcoFluid/newtonIcoFluid.C
@@ -435,8 +435,8 @@ label newtonIcoFluid::formResidual
         (
             (rAUf()*mesh.Sf()) & fvc::interpolate(fvc::grad(p))
         )
-      //- tr(fvc::grad(U)) // or - fvc::div(U)
-      - fvc::div(U)
+      - tr(fvc::grad(U))
+        //- fvc::div(U)
     );
 
     // Make residual extensive
@@ -493,11 +493,11 @@ label newtonIcoFluid::formJacobian
     turbulence_->correct();
 
     // Calculate the segregated approximatoion of momentum equation Jacobian
+    // Note: the nonlinear convection term is added separately below
     fvVectorMatrix UEqn
     (
         fvm::laplacian(turbulence_->nuEff(), U)
       - fvm::ddt(U)
-      - fvm::div(phi(), U, "jacobian-div(phi,U)")
     );
 
     UEqn.relax();
@@ -508,240 +508,17 @@ label newtonIcoFluid::formJacobian
         UEqn, jac, 0, 0, fluidModel::twoD() ? 2 : 3
     );
 
-    // Add additional term from div(phi,U) upwind linearisation
-    // (d/dU)(\sum phi*U = phi*I + w*Sf*Up
-    // for phi > 0, where phi = Sf*(w*Up + (1-w)*Un)
-    // The w*Sf*Up term is missing from fvm::div(phi(), U) as it requires a
-    // coupled solver. So we will add the w*Sf*Up term here
-    // Similary, for phi < 0, the neighbour coefficient is (1 - w)*Sf*Un
-    {
-        const scalarField& phiI = phi();
-        const vectorField& UI = U;
-        const vectorField& SfI = mesh.Sf();
-        const scalarField& wI = mesh.weights();
-        const labelList& own = mesh.owner();
-        const labelList& nei = mesh.neighbour();
-        const scalar sign = -1;
-        tensor coeff = tensor::zero;
-
-        // Prepare coeff array
-        const label nCoeffCmpts = blockSize_*blockSize_;
-        PetscScalar values[nCoeffCmpts];
-        std::memset(values, 0, sizeof(values));
-
-        forAll(phiI, faceI)
-        {
-            // Local row ID
-            const label ownCellID = own[faceI];
-
-            // Local column ID
-            const label neiCellID = nei[faceI];
-
-            // Global block row ID
-            const label globalBlockRowI =
-                foamPetscSnesHelper::globalCells().toGlobal(ownCellID);
-
-            // Global block column ID
-            const label globalBlockColI =
-                foamPetscSnesHelper::globalCells().toGlobal(neiCellID);
-
-            if (phiI[faceI] > 0)
-            {
-                // Add w*Sf*Up to owner eqn
-                // coeff = sign*wI[faceI]*SfI[faceI]*UI[ownCellID];
-                coeff = sign*wI[faceI]*UI[ownCellID]*SfI[faceI];
-                for (label i = 0; i < (blockSize_ - 1); ++i)
-                {
-                    for (label j = 0; j < (blockSize_ - 1); ++j)
-                    {
-                        // Copy 3x3 (or 2x2 in 2-D) coeff into the top left of
-                        // the 4x4 (or 3x3 in 2-D) values matrix
-                        values[i*blockSize_ + j] = coeff[i*3 + j];
-                    }
-                }
-                MatSetValuesBlocked
-                (
-                    jac, 1, &globalBlockRowI, 1, &globalBlockRowI, values,
-                    ADD_VALUES
-                );
-
-                // Add (1 - w)*Sf*Up as nei to contribution to own eqn
[…]
