+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelper.C
@@ -25,6 +25,7 @@ License
 #include "symmetryFvPatchFields.H"
 #include "symmetryPlaneFvPatchFields.H"
 #include "leastSquaresVectors.H"
+#include "fvm.H"
 
 // * * * * * * * * * * * * * * External Functions  * * * * * * * * * * * * * //
 
@@ -806,6 +807,395 @@ label foamPetscSnesHelper::InsertFvmGradIntoPETScMatrix
                 }
             }
         }
+    }
+
+    return 0;
+}
+
+
+label foamPetscSnesHelper::InsertFvmDivPhiUIntoPETScMatrix
+(
+    const volVectorField& U,
+    const surfaceScalarField& phi,
+    Mat jac,
+    const label rowOffset,
+    const label colOffset,
+    const label nScalarEqns,
+    const bool flipSign
+) const
+{
+    // Take references for efficiency and brevity
+    const fvMesh& mesh = mesh_;
+    const scalarField& phiI = phi;
+    const vectorField& UI = U;
+    const vectorField& SfI = mesh.Sf();
+    const scalarField& wI = mesh.weights();
+    const labelList& own = mesh.owner();
+    const labelList& nei = mesh.neighbour();
+    const scalar sign = flipSign ? -1 : 1;
+    tensor coeff = tensor::zero;
+
+    // Add segregated component of convection term
+    {
+        fvVectorMatrix UEqn
+        (
+          - fvm::div(phi, U, "jacobian-div(phi,U)")
+        );
+
+        UEqn.relax();
+
+        // Convert fvMatrix matrix to PETSc matrix
+        foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
+        (
+            UEqn, jac, 0, 0, nScalarEqns
+        );
+    }
+
+    // Add additional term from div(phi,U) upwind linearisation
+    // (d/dU)(\sum phi*U = phi*I + w*Sf*Up
+    // for phi > 0, where phi = Sf*(w*Up + (1-w)*Un)
+    // The w*Sf*Up term is missing from fvm::div(phi(), U) as it requires a
+    // coupled solver. So we will add the w*Sf*Up term here
+    // Similary, for phi < 0, the neighbour coefficient is (1 - w)*Sf*Un
+
+    // Prepare coeff array
+    const label nCoeffCmpts = blockSize_*blockSize_;
+    PetscScalar values[nCoeffCmpts];
+    std::memset(values, 0, sizeof(values));
+
+    forAll(phiI, faceI)
+    {
+        // Local row ID
+        const label ownCellID = own[faceI];
+
+        // Local column ID
+        const label neiCellID = nei[faceI];
+
+        // Global block row ID
+        const label globalBlockRowI =
+            foamPetscSnesHelper::globalCells().toGlobal(ownCellID);
+
+        // Global block column ID
+        const label globalBlockColI =
+            foamPetscSnesHelper::globalCells().toGlobal(neiCellID);
+
+        if (phiI[faceI] > 0)
+        {
+            // Add w*Sf*Up to owner eqn
+            // coeff = sign*wI[faceI]*SfI[faceI]*UI[ownCellID];
+            coeff = sign*wI[faceI]*UI[ownCellID]*SfI[faceI];
+            for (label i = 0; i < (blockSize_ - 1); ++i)
+            {
+                for (label j = 0; j < (blockSize_ - 1); ++j)
+                {
+                    // Copy 3x3 (or 2x2 in 2-D) coeff into the top left of
+                    // the 4x4 (or 3x3 in 2-D) values matrix
+                    values[(i + rowOffset)*blockSize_ + j + colOffset] =
+                        coeff[i*3 + j];
+                }
+            }
+            CHKERRQ
+            (
[…]
