+++ src/solids4FoamModels/numerics/leastSquaresS4fGrad/leastSquaresS4fGrad.C
@@ -28,6 +28,7 @@ License
 #include "symmetryPolyPatch.H"
 #include "symmetryPlanePolyPatch.H"
 #include "solidTractionFvPatchVectorField.H"
+#include "boolIOList.H"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
@@ -70,33 +71,25 @@ Foam::fv::leastSquaresS4fGrad<Type>::calcGrad
     );
     GeometricField<GradType, fvPatchField, volMesh>& lsGrad = tlsGrad.ref();
 
-    // Prepare the useBoundaryFaceValues list for the least squares vector
-    // We will set traction patches to false and displacement patches to true
-    // TESTING: how can I do this in a nice way?
-    // Option: useBoundaryFaceValues field is looked up and must be defined
-    // before grad is called -> avoids misuse, although mildly inconvenient
-    boolList useBoundaryFaceValues(mesh.boundary().size(), true);
-    if (vsf.name() == "D" || vsf.name() == "DD")
+    // Lookup the useBoundaryFaceValues list
+    // This list needs to be created before calling the gradient operation
+    // The list defines which patch values should be used in the least squares
+    // calculation.
+    // There must be a nice way to define this on the fly, but for now this
+    // solution is OK, as it avoids misuse
+    const word useBndFcValName("useBoundaryFaceValues_" + vsf.name());
+    if (!mesh.foundObject<boolIOList>(useBndFcValName))
     {
-        forAll(mesh.boundary(), patchI)
-        {
-            if (isA<solidTractionFvPatchVectorField>(vsf.boundaryField()[patchI]))
-            {
-                useBoundaryFaceValues[patchI] = false;
-            }
-        }
+        FatalErrorInFunction
+            << useBndFcValName << " boolIOList not found! " << nl
+            << "To use the leastSquaresS4fGrad scheme, you must first define an"
+            << " boolIOList called " << useBndFcValName << ", which indicates"
+            << " which patches should be used in the least squares problems for"
+            << " the " << vsf.name() << " field."
+            << abort(FatalError);
     }
-    else
-    {
-        useBoundaryFaceValues = false;
-        forAll(mesh.boundary(), patchI)
-        {
-            if (vsf.boundaryField()[patchI].fixesValue())
-            {
-                useBoundaryFaceValues[patchI] = true;
-            }
-        }
-    }
+    const boolIOList& useBoundaryFaceValues =
+        mesh.lookupObject<boolIOList>("useBoundaryFaceValues_" + vsf.name());
 
     // Get reference to least square vectors
     const leastSquaresS4fVectors& lsv =
