+++ src/solids4FoamModels/solidModels/nonLinGeomTotalLagTotalDispSolid/nonLinGeomTotalLagTotalDispSolid.C
@@ -74,6 +74,16 @@ void nonLinGeomTotalLagTotalDispSolid::predict()
 
     // Calculate the stress using run-time selectable mechanical law
     mechanical().correct(sigma());
+
+    if (solvePressure())
+    {
+        // Predict p using the dp/dt field
+        p() = p().oldTime() + dpdtPtr_.ref()*runTime().deltaT();
+        // p() = p().oldTime() + dpdt*runTime().deltaT()
+        //     + 0.5*sqr(runTime().deltaT())*d2pdt2;
+
+        sigma() = dev(sigma()) - p()*I;
+    }
 }
 
 
@@ -350,8 +360,16 @@ bool nonLinGeomTotalLagTotalDispSolid::evolveSnes()
         (
             D().primitiveFieldRef(),
             foamPetscSnesHelper::solution(),
-            solidModel::twoD() ? 2 : 3, // Block size of x
-            0                           // Location of first component
+            0, // Location of first component
+            solidModel::twoD() ? labelList({0,1}) : labelList({0,1,2})
+        );
+
+        // Map the p field to the SNES solution vector
+        foamPetscSnesHelper::InsertFieldComponents<scalar>
+        (
+            p().primitiveFieldRef(),
+            foamPetscSnesHelper::solution(),
+            blockSize_ -1 // Location of first component
         );
     }
 
@@ -364,10 +382,29 @@ bool nonLinGeomTotalLagTotalDispSolid::evolveSnes()
     (
         foamPetscSnesHelper::solution(),
         D().primitiveFieldRef(),
-        solidModel::twoD() ? 2 : 3, // Block size of x
-        0                           // Location of first component
+        0, // Location of first component
+        solidModel::twoD() ? labelList({0,1}) : labelList({0,1,2})
     );
 
+    D().correctBoundaryConditions();
+
+    if (solvePressure())
+    {
+        // Map the PETSc solution to the p field
+        // p is located in the last ("blockSize - 1") component
+        foamPetscSnesHelper::ExtractFieldComponents<scalar>
+        (
+            foamPetscSnesHelper::solution(),
+            p().primitiveFieldRef(),
+            blockSize_ - 1 // Location of p component
+        );
+
+        p().correctBoundaryConditions();
+
+        // Update dpdt
+        dpdtPtr_.ref() = fvc::ddt(p());
+    }
+    
     // Interpolate cell displacements to vertices
     mechanical().interpolate(D(), gradD(), pointD());
     pointD().correctBoundaryConditions();
@@ -462,7 +499,7 @@ nonLinGeomTotalLagTotalDispSolid::nonLinGeomTotalLagTotalDispSolid
                 "optionsFile", "petscOptions"
             )
         ),
-        mesh(),
+        mesh().nCells(),
         solvePressure()
       ? label(solidModel::twoD() ? 3 : 4)
       : label(solidModel::twoD() ? 2 : 3),
@@ -534,6 +571,7 @@ nonLinGeomTotalLagTotalDispSolid::nonLinGeomTotalLagTotalDispSolid
     impKf_(fvc::interpolate(impK_)),
     rImpK_(1.0/impK_),
     pDiffusivityPtr_(),
+    dpdtPtr_(),
     predictor_(solidModelDict().lookupOrDefault<Switch>("predictor", false)),
     blockSize_
     (
@@ -548,7 +586,6 @@ nonLinGeomTotalLagTotalDispSolid::nonLinGeomTotalLagTotalDispSolid
     fvm::d2dt2(D());
 
     Info<< "solvePressure = " << solvePressure() << endl;
-
     if (solvePressure())
     {
[…]
