### 2025-02-24 97a98bc
README file for sphericalCavity
