+++ tutorials/solids/linearElasticity/sphericalCavity/README.md
@@ -1,0 +1,138 @@
+# Spherical cavity under uniaxial tensions: `sphericalCavity`
+
+---
+
+Prepared by Philip Cardiff and Ivan Batistić
+
+---
+
+## Tutorial Aims
+
+- Demonstrate how to perform a 3D linear-static stress analysis in solids4foam;
+- Demonstrate the process of meshing using Gmsh meshing utility and OpenFOAM `polyDualMesh` utility
+
+---
+
+## Case Overview
+
+This classic 3-D problem consists of a spherical cavity with radius $a = 0.2$ m (Figure 1) in an infinite, isotropic linear elastic solid ($E = 200$ GPa, $\nu = 0.3$). Far from the cavity, the solid is subjected to a tensile stress $\sigma_{zz} = T = 1$ MPa, with all other stress components zero. The analytical expressions for the stress are derived by Southwell and Gough [1], while expression for displacement are derived by Goodier [2]. The problem is characterised by a localised stress concentration near the cavity on the plane perpendicular to the loading, which drops off rapidly away from the cavity. The computational solution domain is taken as one-eighth of a $1 \times 1 \times 1$ m cube aligned with the Cartesian axes, with one corner at the centre of the sphere. The problem is axisymmetric but is analysed here using a 3-D domain. The analytical tractions are applied at the far boundaries of the domain to mitigate the effects of finite geometry. They are defined with `analyticalSphericalCavityTraction` boundary conditions, which require geometry (cavity radius), loading (far-field traction), and material properties (Poisson's ratio):
+
+```c++
+    patch
+    {
+        type            analyticalSphericalCavityTraction;
+        farFieldTractionZ 1e6;
+        cavityRadius    0.2;
+        nu              0.3;
+        value           uniform (0 0 0);
+    }
+```
+
+![-](images/sphericalCavity-geometry.png)
+
+**Figure 1: Spherical cavity case geometry and mesh (4 555 cells).**
+
+---
+
+## Expected Results
+
+The analytical expressions for the stress distributions around the cavity, first derived by Southwell and Gough [1]:
+
+$$
+\sigma_{rr} =
+\frac{T}{14 - 10\nu} \frac{a^3}{R^3}
+\left[ 9 - 15\nu - 12 \frac{a^2}{R^2}  - \frac{r^2}{R^2} \left( 72 - 15\nu - 105 \frac{a^2}{R^2} \right) + 15 \frac{r^4}{R^4} \left( 5 - 7 \frac{a^2}{R^2} \right) \right],
+$$
+
+$$
+\sigma_{\theta\theta} =
+\frac{T}{14 - 10\nu} \frac{a^3}{R^3}
+\left[ 9 - 15\nu - 12 \frac{a^2}{R^2}  - 15 \frac{r^2}{R^2} \left( 1 - 2\nu - \frac{a^2}{R^2} \right) \right],
+$$
+
+$$
+\sigma_{zz} =
+T \left[ 1 - \frac{1}{14 - 10\nu} \frac{a^3}{R^3} \left\{ 38 - 10\nu - 24 \frac{a^2}{R^2}
+- \frac{r^2}{R^2} \left( 117 - 15\nu - 120 \frac{a^2}{R^2} \right)
++ 15 \frac{r^4}{R^4} \left( 5 - 7 \frac{a^2}{R^2} \right) \right\} \right]
+$$
+
+$$
+\sigma_{zr} =
+\frac{T}{14 - 10\nu} \frac{a^3 z r}{R^5}
+\left[ -3(19 - 5\nu) + 60 \frac{a^2}{R^2} + 15 \frac{r^2}{R^2} \left( 5 - 7 \frac{a^2}{R^2} \right)  \right].
+$$
+
[…]
