### 2025-02-24 233ac75
Added sphericalCavity in function object README.md
