### 2025-02-24 f9008f9
[Documentation] sphericalCavity case README.md file (#163)
* README file for sphericalCavity

* Added sphericalCavity in function object README.md

* Removed skew correction
