+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -227,7 +227,7 @@ label newtonCouplingInterface::initialiseJacobian(Mat jac)
         // Set the block size: this corresponds to the row size for a
         // rectangular matrix
         CHKERRQ(MatSetBlockSizes(jac, fluidBlockSize, solidBlockSize));
-        
+
         // Number of blocks: number of faces at the interface
         const label blockn = interfaceMap.zoneBToZoneAFaceMap().size();
 
@@ -280,7 +280,7 @@ label newtonCouplingInterface::initialiseJacobian(Mat jac)
         // Set the block size: this corresponds to the row size for a
         // rectangular matrix
         CHKERRQ(MatSetBlockSizes(jac, solidBlockSize, fluidBlockSize));
-        
+
         // Number of blocks: number of faces at the interface
         const label blockn = interfaceMap.zoneAToZoneBFaceMap().size();
 
@@ -336,15 +336,115 @@ label newtonCouplingInterface::formResidual
     const PetscScalar *x
 )
 {
-    // TODO
-    WarningInFunction
-        << "I need to map the traction from the fluid to the solid!"
-        << endl;
-    // notImplemented("I need to map the traction from the fluid to the solid");
+    // Approach
+    // 1. Update the fluid velocity and pressure fields and calculate the
+    //     traction at the fluid interface
+    // 2. Map the fluid interface traction to the solid interface
+    // 3. Update the solid residual, which now has the correct interface
+    //    traction
+    // 4. Map the solid interface velocity to the fluid interface
+    // 5. Update the fluid residual, which now has the correct interface
+    //    velocity
+
+    // Block sizes
+    const label fluidBlockSize = fluid().twoD() ? 3 : 4;
+    // const label solidBlockSize = solid().twoD() ? 2 : 3;
+
+    // Currently limited to one interface: it should be straight-forward to add
+    // a loop over multiple interface => todo
+    if (fluidSolidInterface::fluidPatchIndices().size() != 1)
+    {
+        FatalError
+            << "Only one interface patch is currently allowed"
+            << abort(FatalError);
+    }
+
+    // 1. Update the fluid velocity and pressure fields and calculate the
+    //     traction at the fluid interface
+    {
+        // Take references
+        volVectorField& U = fluid().U();
+        volScalarField& p = fluid().p();
+
+        // Retrieve the solution
+        // Map the PETSc solution to the U field
+        foamPetscSnesHelper::ExtractFieldComponents<vector>
+        (
+            x,
+            U.primitiveFieldRef(),
+            0, // Location of U
+            fluidBlockSize,
+            fluid().twoD() ? labelList({0,1}) : labelList({0,1,2})
+        );
+
+        U.correctBoundaryConditions();
+
+        // Map the PETSc solution to the p field
+        // p is located in the final component
+        foamPetscSnesHelper::ExtractFieldComponents<scalar>
+        (
+            x,
+            p.primitiveFieldRef(),
+            fluidBlockSize - 1, // Location of p component
+            fluidBlockSize
+        );
+
+        p.correctBoundaryConditions();
+
+        // Fluid interface traction
+        const label fluidPatchID = fluidSolidInterface::fluidPatchIndices()[0];
+        const vectorField fluidNf(fluidMesh().boundary()[fluidPatchID].nf());
+        const vectorField fluidTraction
+        (
+            fluid().patchViscousForce(fluidPatchID)
+          - fluid().patchPressureForce(fluidPatchID)*fluidNf
+        );
+
+        // Lookup the interface map from the fluid faces to the solid faces
+        const interfaceToInterfaceMappings::
+            directMapInterfaceToInterfaceMapping& interfaceMap =
+            refCast
[…]
