### 2025-02-25 b32dcd3
sphericalCavity: add URL to README and make file lint compliant
