+++ tutorials/solids/linearElasticity/sphericalCavity/README.md
@@ -8,14 +8,31 @@ Prepared by Philip Cardiff and Ivan Batistić
 
 ## Tutorial Aims
 
-- Demonstrate how to perform a 3D linear-static stress analysis in solids4foam;
-- Demonstrate the process of meshing using Gmsh meshing utility and OpenFOAM `polyDualMesh` utility
+- Demonstrate how to perform a 3D linear-static stress analysis in solids4foam
+- Demonstrate the process of meshing using the Gmsh meshing utility and
+  OpenFOAM `polyDualMesh` utility
+- Demonstrates using the solids4foam solid solver based on the PETSc SNES
+  nonlinear solver
 
 ---
 
 ## Case Overview
 
-This classic 3-D problem consists of a spherical cavity with radius $a = 0.2$ m (Figure 1) in an infinite, isotropic linear elastic solid ($E = 200$ GPa, $\nu = 0.3$). Far from the cavity, the solid is subjected to a tensile stress $\sigma_{zz} = T = 1$ MPa, with all other stress components zero. The analytical expressions for the stress are derived by Southwell and Gough [1], while expression for displacement are derived by Goodier [2]. The problem is characterised by a localised stress concentration near the cavity on the plane perpendicular to the loading, which drops off rapidly away from the cavity. The computational solution domain is taken as one-eighth of a $1 \times 1 \times 1$ m cube aligned with the Cartesian axes, with one corner at the centre of the sphere. The problem is axisymmetric but is analysed here using a 3-D domain. The analytical tractions are applied at the far boundaries of the domain to mitigate the effects of finite geometry. They are defined with `analyticalSphericalCavityTraction` boundary conditions, which require geometry (cavity radius), loading (far-field traction), and material properties (Poisson's ratio):
+This classic 3-D problem consists of a spherical cavity with radius $a = 0.2$ m
+ (Figure 1) in an infinite, isotropic linear elastic solid ($E = 200$ GPa,
+ $\nu = 0.3$). Far from the cavity, the solid is subjected to a tensile stress
+ $\sigma_{zz} = T = 1$ MPa, with all other stress components zero. The analytical
+ expressions for stress are derived by Southwell and Gough [1], while expressions
+ for displacement are derived by Goodier [2]. The problem is characterised by a
+ localised stress concentration near the cavity on the plane perpendicular to the
+ loading, which drops off rapidly away from the cavity. The computational
+ solution domain is taken as one-eighth of a $1 \times 1 \times 1$ m cube aligned
+ with the Cartesian axes, with one corner at the sphere's centre. The problem is
+ axisymmetric but is analysed here using a 3-D domain. The analytical tractions
+ are applied at the far boundaries of the domain to mitigate the effects of
+ finite geometry. They are defined with `analyticalSphericalCavityTraction`
+ boundary conditions, which require geometry (cavity radius), loading (far-field
+ traction), and material properties (Poisson's ratio):
 
 ```c++
     patch
@@ -36,18 +53,26 @@ This classic 3-D problem consists of a spherical cavity with radius $a = 0.2$ m 
 
 ## Expected Results
 
[…]
