+++ tutorials/fluidSolidInteraction/beamInCrossFlow/README.md
@@ -341,45 +341,45 @@ domain
 
 ```tip
 **Solution 1**: some of the solidModels use a non-moving mesh formulation so, by
-default, some solids may not move at all when shown in ParaView. To show the
-solid domain deformation/motion in ParaView, select the solid case and use the
-“Warp By Vector” filter with the displacement (`D` or `pointD`) field. If there
-is still a gap between the fluid and solid, see Solution 2 below.
+ default, some solids may not move at all when shown in ParaView. To show the
+ solid domain deformation/motion in ParaView, select the solid case and use the
+ “Warp By Vector” filter with the displacement (`D` or `pointD`) field. If there
+ is still a gap between the fluid and solid, see Solution 2 below.
 ```
 
 ```tip
 **Solution 2**: if the FSI loop does not converge then the FSI interface
-constraints (see below) may not be strictly enforced; examine the “fsi residual”
-in the log file and check if the maximum number of FSI iterations is being
-reached. If the FSI method is not converging to the required tolerance, then
-increases the maximum number of FSI correctors and/or decrease the initial
-relaxation factor (in fsiProperties). In addition, you can try a different FSI
-coupling procedure. If the FSI loop is converging but there is still a gap
-between the fluid and solid, see the next points on the following slide.
+ constraints (see below) may not be strictly enforced; examine the “fsi residual”
+ in the log file and check if the maximum number of FSI iterations is being
+ reached. If the FSI method is not converging to the required tolerance, then
+ increases the maximum number of FSI correctors and/or decrease the initial
+ relaxation factor (in fsiProperties). In addition, you can try a different FSI
+ coupling procedure. If the FSI loop is converging but there is still a gap
+ between the fluid and solid, see the next points on the following slide.
 ```
 
 ```tip
 **Solution 3**: if the FSI loop is converging but there is still a gap between
-the fluid and solid, then try increase the FSI solution tolerance (in
-`fsiProperties`); you may also need to increase the maximum number of FSI
-correctors to achieve a tighter tolerance. If this still does not help, see
-Solution 4 below.
+ the fluid and solid, then try increase the FSI solution tolerance (in
+ `fsiProperties`); you may also need to increase the maximum number of FSI
+ correctors to achieve a tighter tolerance. If this still does not help, see
+ Solution 4 below.
 ```
 
 ```tip
 **Solution 4**: try create a conformal fluid-to-solid interface i.e. make the
-fluid interface patch mesh exactly the same as the solid interface patch mesh.
+ fluid interface patch mesh exactly the same as the solid interface patch mesh.
 ```
 
 ```tip
 **Solution 5**: if the FSI method is still struggling to converge, try make the
-problem “easier” by setting the solid to be temporarily stiffer and denser (in
-mechanicalProperties) by e.g. multiple orders of magnitude. If the problem works
-with the artificially stiff/dense solid, then slowly decrease the solid
-properties towards the real ones until you find the point at which the FSI
-methods breaks; then try decreasing the starting FSI relaxation factor and/or
-maximum number of FSI correctors and/or FSI coupling algorithm to determine the
-critical/optimal algorithm variables.
+ problem “easier” by setting the solid to be temporarily stiffer and denser (in
+ mechanicalProperties) by e.g. multiple orders of magnitude. If the problem works
+ with the artificially stiff/dense solid, then slowly decrease the solid
+ properties towards the real ones until you find the point at which the FSI
[…]
