### 2025-03-01 #7407 @philipcardiff
@MakisH : Regarding tutorial 3, we re-ran it using the latest development branch, and now we get the same results as you have shown here. I have updated the image on the website accordingly:
