### 2025-03-01 b38223c
Fix for scalar symmetry plane Jacobian
