+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -97,7 +97,7 @@ bool newtonCouplingInterface::evolve()
     volVectorField& D = solid().D();
 
     const label fluidBlockSize = fluid().twoD() ? 3 : 4;
-    const label solidBlockSize = solid().twoD() ? 2 : 4;
+    const label solidBlockSize = solid().twoD() ? 2 : 3;
 
     WarningInFunction
         << "Todo: map U/p/D to x" << endl;
@@ -175,29 +175,39 @@ label newtonCouplingInterface::initialiseJacobian(Mat jac)
 
     // Set fluid and solid block sizes
     const label fluidBlockSize = fluid().twoD() ? 3 : 4;
-    const label solidBlockSize = solid().twoD() ? 2 : 4;
+    const label solidBlockSize = solid().twoD() ? 2 : 3;
 
     // We will initialise the four sub-matrices:
     //  - Afluid: fluid
     //  - Asolid: solid
-    //  - Asif: solid-in-fluid-coupling
-    //  - Afis: fluid-in-solid coupling
+    //  - Asf: solid-in-fluid-coupling
+    //  - Afs: fluid-in-solid coupling
 
     // Initialise Afluid based on compact stencil fvMesh
     {
+        if (debug)
+        {
+            Info<< "Initialising Afluid" << endl;
+        }
+
         Mat Afluid = subMats[0][0];
         Foam::initialiseJacobian(Afluid, fluid().mesh(), fluidBlockSize);
     }
 
     // Initialise Asolid based on compact stencil fvMesh
     {
+        if (debug)
+        {
+            Info<< "Initialising Asolid" << endl;
+        }
+
         Mat Asolid = subMats[1][1];
         Foam::initialiseJacobian(Asolid, solid().mesh(), solidBlockSize);
     }
 
     // Initially we assume a conformal FSI interface, where each fluid cell
     // shares a face with a solid cell. So we assume the number of blocks in
-    // the Asif (and Afis) matrix is equal to the number of cells at the
+    // the Asf (and Afs) matrix is equal to the number of cells at the
     // interface
 
     // Initially, we only allow one FSI interface
@@ -222,26 +232,50 @@ label newtonCouplingInterface::initialiseJacobian(Mat jac)
 
     // Initialise solid-in-fluid-coupling matrix
     {
-        Mat Asif = subMats[0][1];
-
-        // Set the block size: this corresponds to the row size for a
-        // rectangular matrix
-        CHKERRQ(MatSetBlockSizes(jac, fluidBlockSize, solidBlockSize));
-
-        // Number of blocks: number of faces at the interface
-        const label blockn = interfaceMap.zoneBToZoneAFaceMap().size();
-
-        // Number of on-processor non-zeros blocks per row
-        int* d_nnz = (int*)malloc(blockn*sizeof(int));
-
-        // Number of off-processor non-zeros per row
-        int* o_nnz = (int*)malloc(blockn*sizeof(int));
-
-        // Initialise d_nnz to 1 and o_nnz to zero
-        for (int i = 0; i < blockn; ++i)
+        if (debug)
         {
-            d_nnz[i] = 1;
-            o_nnz[i] = 0;
+            Info<< "Initialising Asf" << endl;
+        }
+
+        Mat Asf = subMats[0][1];
+
+        // CAREFUL: we are setting non-zeros her based on the scalar rows, not
+        // the block rows
+
+        // Set matrix type to AIJ (since BAIJ does not support non-square
+        // blocks)
+        CHKERRQ(MatSetType(Asf, MATMPIAIJ));
+
+        // Total number of scalar rows in the fluid region
+        const label scalarRowN = fluidMesh().nCells()*fluidBlockSize;
+
+        // Allocate per-scalar-row nonzeros, initialised to 0
[…]
