### 2025-03-02 36a0a66
Add solidSystemScaleFactor, which is important for the convergence of the monolithic system
