+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -75,7 +75,17 @@ newtonCouplingInterface::newtonCouplingInterface
         labelListList(),
         fsiProperties().lookupOrDefault<Switch>("stopOnPetscError", true),
         true // Will PETSc be used
-    )
+    ),
+    solidSystemScaleFactor_
+    (
+        fsiProperties().lookupOrDefault<scalar>
+        (
+            "solidSystemScaleFactor",
+            gAverage(1.0/solid().mechanical().shearModulus()().primitiveField())
+        )
+    ),
+    fluidToSolidCoupling_(fsiProperties().lookup("fluidToSolidCoupling")),
+    solidToFluidCoupling_(fsiProperties().lookup("solidToFluidCoupling"))
 {}
 
 
@@ -99,8 +109,8 @@ bool newtonCouplingInterface::evolve()
     const label fluidBlockSize = fluid().twoD() ? 3 : 4;
     const label solidBlockSize = solid().twoD() ? 2 : 3;
 
-    WarningInFunction
-        << "Todo: map U/p/D to x" << endl;
+    // WarningInFunction
+    //     << "Todo: Add linear predictor" << endl;
 
     // Solve the nonlinear system and check the convergence
     foamPetscSnesHelper::solve();
@@ -154,6 +164,18 @@ bool newtonCouplingInterface::evolve()
 
     // Restore the x vector
     VecRestoreArrayRead(foamPetscSnesHelper::solution(), &xx);
+
+    // Interpolate cell displacements to vertices
+    solid().mechanical().interpolate(solid().D(), solid().gradD(), solid().pointD());
+
+    // Increment of displacement
+    solid().DD() = solid().D() - solid().D().oldTime();
+
+    // Increment of point displacement
+    solid().pointDD() = solid().pointD() - solid().pointD().oldTime();
+
+    // Velocity
+    solid().U() = fvc::ddt(solid().D());
 
     return 0;
 }
@@ -386,9 +408,9 @@ label newtonCouplingInterface::initialiseJacobian(Mat jac)
         Info<< "End of newtonCouplingInterface::initialiseJacobian" << endl;
     }
 
-    PetscInt rGlobal, cGlobal;
-    MatGetSize(jac, &rGlobal, &cGlobal);
-    PetscPrintf(PETSC_COMM_WORLD, "\nJacobian Matrix Global Size: %D x %D\n", rGlobal, cGlobal);
+    // PetscInt rGlobal, cGlobal;
+    // MatGetSize(jac, &rGlobal, &cGlobal);
+    // PetscPrintf(PETSC_COMM_WORLD, "\nJacobian Matrix Global Size: %D x %D\n", rGlobal, cGlobal);
 
     return 0;
 }
@@ -412,7 +434,7 @@ label newtonCouplingInterface::formResidual
 
     // Block sizes
     const label fluidBlockSize = fluid().twoD() ? 3 : 4;
-    // const label solidBlockSize = solid().twoD() ? 2 : 3;
+    const label solidBlockSize = solid().twoD() ? 2 : 3;
 
     // Currently limited to one interface: it should be straight-forward to add
     // a loop over multiple interface => todo
@@ -462,55 +484,67 @@ label newtonCouplingInterface::formResidual
 
         p.correctBoundaryConditions();
 
-        // Fluid interface traction
-        const label fluidPatchID = fluidSolidInterface::fluidPatchIndices()[0];
-        const vectorField fluidNf(fluidMesh().boundary()[fluidPatchID].nf());
-        const vectorField fluidTraction
-        (
-            fluid().patchViscousForce(fluidPatchID)
-          - fluid().patchPressureForce(fluidPatchID)*fluidNf
-        );
-
-        // Lookup the interface map from the fluid faces to the solid faces
[…]
