### 2025-03-02 10fd0ac
Use extrapolated displacement at the interface
