### 2025-03-03 a45b424
Updates to paper.md based on JOSS review
