+++ paper/paper.bib
@@ -1,3 +1,160 @@
+@misc{codeAster,
+Author = {Electricit{\'e} de France},
+Howpublished = {Open source on www.code-aster.org},
+Title = {Finite element {$\bf\it code\_aster$}, Analysis of Structures and Thermomechanics for Studies and Research,
+Year = {1989--2017}
+}
+
+@book{Dhondt2004,
+  author    = {Dhondt, Guido},
+  title     = {The Finite Element Method for Three-Dimensional Thermomechanical Applications},
+  publisher = {Wiley},
+  year      = {2004},
+  address   = {Chichester, England},
+  isbn      = {978-0-470-85752-6}
+}
+
+@article{Haider2018,
+  author    = {Haider, Jibran and Lee, Chun Hean and Gil, Antonio J. and Huerta, Antonio and Bonet, Javier},
+  title     = {An upwind cell centred Total Lagrangian finite volume algorithm for nearly incompressible explicit fast solid dynamic applications},
+  journal   = {Computer Methods in Applied Mechanics and Engineering},
+  volume    = {340},
+  pages     = {684--727},
+  year      = {2018},
+  doi       = {10.1016/j.cma.2018.06.033}
+}
+
+@article{Haider2017,
+  author    = {Haider, Jibran and Lee, Chun Hean and Gil, Antonio J. and Bonet, Javier},
+  title     = {A first order hyperbolic framework for large strain computational solid dynamics: An upwind cell centred Total Lagrangian scheme},
+  journal   = {International Journal for Numerical Methods in Engineering},
+  volume    = {109},
+  number    = {3},
+  pages     = {407--456},
+  year      = {2017},
+  doi       = {10.1002/nme.5293}
+}
+
+@article{Mehl2016,
+  author    = {Mehl, Miriam and Uekermann, Benjamin and Bijl, Hester and Blom, David and Gatzhammer, Bernhard and Van Zuijlen, Alexander},
+  title     = {Parallel Coupling Numerics for Partitioned Fluid–Structure Interaction Simulations},
+  journal   = {Computers & Mathematics with Applications},
+  volume    = {71},
+  number    = {4},
+  pages     = {869-891},
+  year      = {2016},
+  doi       = {10.1016/j.camwa.2015.12.025}
+}
+
+@article{Tang2015,
+  author    = {Tang, T. and Hededal, O. and Cardiff, P.},
+  title     = {On finite volume method implementation of poro-elasto-plasticity soil model},
+  journal   = {International Journal for Numerical and Analytical Methods in Geomechanics},
+  volume    = {39},
+  pages     = {1410–1430},
+  year      = {2015},
+  doi       = {10.1002/nag.2361}
+}
+
+@misc{Haider2019,
+  author    = {Haider, Jibran},
+  title     = {{ExplicitSolidDynamics toolkit for OpenFOAM}},
+  year      = {2019},
+  doi       = {10.5281/zenodo.2577033},
+  url       = {https://github.com/jibranhaider/explicitSolidDynamics}
+}
+
+@article{Wagner2022,
+  author    = {Wagner, S. and Münsch, M. and Delgado, A.},
+  title     = {An Integrated OpenFOAM Membrane Fluid-Structure Interaction Solver},
+  journal   = {OpenFOAM® Journal},
+  volume    = {2},
+  pages     = {48–61},
+  year      = {2022},
+  doi       = {10.51560/ofj.v2.45}
+}
+
+@article{Schmitt2022,
+  author    = {Schmitt, P. and Robinson, D.},
+  title     = {A Coupled Actuator Line and Finite Element Analysis Tool},
[…]
