### 2025-03-03 0058a28
Fix missing bracket in paper.bib
