+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @misc{codeAster,
 Author = {Electricit{\'e} de France},
 Howpublished = {Open source on www.code-aster.org},
-Title = {Finite element {$\bf\it code\_aster$}, Analysis of Structures and Thermomechanics for Studies and Research,
+Title = {Finite element {$\bf\it code\_aster$}, Analysis of Structures and Thermomechanics for Studies and Research},
 Year = {1989--2017}
 }
 
