+++ paper/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 
 The solids4foam toolbox is one of the most comprehensive solid mechanics and fluid-solid interaction toolboxes available within OpenFOAM, having evolved from the `solidMechanics` and `extend-bazaar/FluidSolidInteraction` toolboxes of the foam-extend community [@foam-extend].
-Several other OpenFOAM-based toolboxes provide capabilities for solid mechanics and fluid-solid interaction, including `FOAM-FSI` [@Mehl2016], `miniGeotechFoam` [@Tang2015], `explicitSolidDynamics` [@Haider2019, @Haider2017, @Haider2018], as well as more specialised solvers such as the membrane fluid-solid interaction solver [@Wagner2022], a coupled actuator line and finite element analysis tool [@Schmitt2022], and a modular multiphysics framework [@St-Onge2023].
+Several other OpenFOAM-based toolboxes provide capabilities for solid mechanics and fluid-solid interaction, including `FOAM-FSI` [@Mehl2016], `miniGeotechFoam` [@Tang2015], `explicitSolidDynamics` [@Haider2019, @Haider2017, @Haider2018], as well as more specialised solvers such as the membrane fluid-solid interaction solver [@Wagner2022], a coupled actuator line and finite element analysis tool [@Schmitt2022], and a modular multiphysics framework [@StOnge2023].
 However, many of these toolboxes are no longer actively maintained and lack the broad range of solid mechanics and fluid-solid interaction features required for general-purpose simulations. Beyond OpenFOAM-based solutions, preCICE [@Chourdakis2023] provides an alternative approach by coupling OpenFOAM with widely-used finite element solvers such as deal.II [@Arndt2021], CalculiX [@Dhondt2004], FEniCS [@Logg2012], and Code_Aster [@codeAster], enabling flexible multiphysics simulations.
 While solids4foam is among the first general finite volume-based toolboxes for solid mechanics and fluid-solid interaction, established finite element-based codes such as FEniCS [@Logg2012] and deal.II [@Arndt2021] offer comparable functionality but differ in their numerical methodology.
 Furthermore, domain-specific fluid-solid interaction solvers, such as SimVascular [@Zhu2022] and Ambit [@Hirschvogel2024] for cardiovascular simulations or turtlefluid-solid interaction [@Bergersen2020] for general monolithic fluid-solid interaction problems, provide specialised solutions for particular applications.
