### 2025-03-03 b570e1b
Paper updates related to the JOSS review
