+++ paper/paper.md
@@ -57,59 +57,39 @@ The design of `solids4foam` adheres to four guiding principles:
 
 # Features
 
-`solids4foam` employs a modular design, offering generic class interfaces for solid mechanics, fluid dynamics, fluid-solid coupling methods, and solid material models. It also supports all native OpenFOAM modularity, including boundary conditions and function objects.
-
-The `solids4foam-v2.1` release includes the following features:
+The solids4foam toolbox is designed with a modular architecture, allowing for a flexible and extensible approach to solid mechanics and fluid-solid interaction within OpenFOAM. The core framework consists of generic class interfaces for solid mechanics, fluid dynamics, coupling methods, and solid material models. Additionally, it supports all native OpenFOAM modularity, including boundary conditions and function objects. The `solids4foam-v2.1` release includes the following features:
 
 ## Partitioned Fluid-Solid Interaction Coupling Methods
 
-- Fixed under-relaxation [@Tukovic2018a]
-- Aitkens accelerated under-relaxation [@Tukovic2018a]
-- Interface-quasi-Newton coupling [@Degroote2009]
-- Robin-Neumann coupling [@Tukovic2018b]
-- Thermo-fluid-solid interaction coupling
+`solids4foam` provides a range of partitioned coupling methods for fluid-solid interaction, including fixed under-relaxation, Aitken’s accelerated under-relaxation, interface-quasi-Newton coupling [@Degroote2009] based on a Dirichlet-Neumann formulation, as well as an added-mass Robin-Neumann formulation. Details of the implementation can be found in [@Cardiff2018a], [@Tukovic2018a], and [@Tukovic2018b]. Thermo-fluid-solid interaction coupling is also available [@solids4foamTFSI].
 
-## Finite Volume Solid Model Discretizations and Solution Algorithms
+## Finite Volume Solid Model Discretisations and Solution Algorithms
 
-- Segregated [@Cardiff2018a], coupled [@Cardiff2016a], and explicit solution algorithms
-- Linear geometry (small strain) and nonlinear geometry (finite strain) formulations, including total and updated Lagrangian
-- Cell-centered and vertex-centered formulations
-- Continuum and plate formulations
+The toolbox supports multiple discretisation approaches and solution algorithms tailored for solid mechanics. Users can choose between segregated [@Cardiff2018a], coupled [@Cardiff2016a], and explicit solution algorithms, offering flexibility depending on computational constraints and problem requirements. Both linear geometry (small strain) and nonlinear geometry (finite strain) formulations are available, with support for total and updated Lagrangian approaches. `solids4foam` includes both cell-centered and vertex-centered formulations.
 
 ## Solid Material Models
 
-- Linear elasticity (isotropic, orthotropic [@Cardiff2014a]), plasticity ($J_2$ [@Cardiff2016b], Mohr-Coulomb [@Tang2015]), viscoelasticity [@Cardiff2018a], thermo-elasticity [@Cardiff2018a], poroelasticity [@Tang2015]
-- Hyperelasticity (neo-Hookean, Ogden, Mooney-Rivlin [@Oliveira2022; @Oliveira2023], Fung [@Oliveira2022; @Oliveira2023], Yeoh [@Oliveira2022; @Oliveira2023]), hyperelastoplasticity [@Cardiff2016b]
-- Interface to Abaqus material model subroutines (UMATs)
[…]
