+++ paper/paper.bib
@@ -335,15 +335,22 @@ Year = {1989--2017}
 }
 
 @misc{OpenFOAM.org,
-  author = {The OpenFOAM Foundation},
+  author = {{The OpenFOAM Foundation}},
   title = {OpenFOAM.org},
   year = {2024},
   url = {https://www.openfoam.org}
 }
 
 @misc{foam-extend,
-  author = {foam-extend},
+  author = {{foam-extend}},
   title = {foam-extend project},
   year = {2024},
   url = {https://sourceforge.net/projects/foam-extend}
+}
+
+@misc{solids4foamTFSI,
+  author = {{solids4foam}},
+  title = {Thermo-Fluid-Solid Interaction Tutorials},
+  year = {2025},
+  url = {https://www.solids4foam.com/tutorials/more-tutorials/thermo-fluid-solid-interaction/}
 }
