+++ paper/paper.bib
@@ -11,7 +11,8 @@ Year = {1989--2017}
   publisher = {Wiley},
   year      = {2004},
   address   = {Chichester, England},
-  isbn      = {978-0-470-85752-6}
+  isbn      = {978-0-470-85752-6},
+  doi       = {10.1002/0470021217}
 }
 
 @article{Haider2018,
