+++ paper/paper.bib
@@ -344,7 +344,7 @@ Year = {1989--2017}
 
 @misc{foam-extend,
   author = {{foam-extend}},
-  title = {foam-extend project},
+  title = {{foam-extend} project},
   year = {2024},
   url = {https://sourceforge.net/projects/foam-extend}
 }
