### 2025-03-03 #7407 @philipcardiff
Hello @MakisH,

We have addressed your outstanding points below:

> ### Software paper
>
>     * ⚠️ **Suggestion:** In point 1, I would add `within OpenFOAM`, similarly to point 2. ⏳ On the one hand, this is an advantage of solids4foam, on the other hand it separates it from preCICE & co. that target coupling to external codes.*

Done.

>     *  ⚠️ **Suggestion:** I would see the point 4 of "guiding principles" (code quality) as controversial. On the one hand, it is difficult to argue about code quality, especially without concrete measures. On the other hand, following a style guide is only one tiny aspect of what comprises code quality. I would rather drop the point altogether.*

That's a fair point; I have removed it.

>     * [ ] [x]  ⏳ **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>     * ❌ This is missing* [x]

Added now to the summary section.

>     * **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>     * ⚠️ **Suggestion:** The paper is generally easy to read, even though mostly structured around bullet-point lists. I would expect a bit more running text at parts (e.g., the opening of every section).*

We have rewritten the features section in running text rather than bullet points.

>     * [ ] Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
>     * ⚠️ **Suggestion:** In some places, there are references to the methods (e.g., in line 39, the reference to the Degroote 2009 paper), and in other places the references are to later implementations/modifications (e.g., in the surrounding lines, the Tukovic 2018 paper). I am not sure what would be best here, but it feels inconsistent. At the same time, in the same section, I get the impression there is a mix of general mathematical methods with application-specific algorithms.*
>     *  ⚠️ **Suggestion:** In other places (e.g., the thermo-fluid-solid interaction coupling), there are no references. Maybe for all of these uncited features, some URLs (just clickable, not necessarily printable) to the documentation, or some guidance to the configuration would help me (as a reviewer) confirm the claims, and a user to find their way.*

Fair point. We have updated the text now to hopefully make that distinction clear. Just let us know if you think it is still unclear.

>     * ⚠️ **Suggestion:** The citation to the OpenFOAM Foundation is not rendered correctly. Maybe the editors can help here. Also, is `Foam-extend` the official writing of the project's name?* ⏳ **Further points:**

Adding extra curly braces to the bib file entries seems to have resolved both issues.

>     * ⚠️ **Suggestion:** There is a one-line "Tutorials" subsection, but this does not mention any examples. This would be helpful for a reader completely new to the tool to understand what this tool typically applies to.

We have added additional text here, and we point to the tutorial guide in the documentation.

Let us know if any of these points (or others) require further clarification.

Thanks!
