### 2025-03-03 d635795
Refactor to foamPetscSnesHelper: add initialiseJacobian and initialiseSolution virtual functions
