+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelper.C
@@ -71,6 +71,10 @@ PetscErrorCode formJacobianFoamPetscSnesHelper
     void *ctx     // user context
 )
 {
+    // Note
+    // The "-snes_lag_jacobian -2" PETSc option can be used to avoid
+    // re-building the matrix
+
     // Get pointer to solution data
     const PetscScalar *xx;
     CHKERRQ(VecGetArrayRead(x, &xx));
@@ -78,47 +82,9 @@ PetscErrorCode formJacobianFoamPetscSnesHelper
     // Access the OpenFOAM data
     appCtxfoamPetscSnesHelper *user = (appCtxfoamPetscSnesHelper *)ctx;
 
-    // Check if the matrix is a nested matrix
-    PetscBool isNest;
-    PetscObjectTypeCompare((PetscObject)B, MATNEST, &isNest);
-
-    // Lookup the matrix info
-    MatInfo info;
-    if (isNest)
-    {
-        // For a nested matrix, we will only check the top left sub-matrix - if
-        // it is initialised we will assume all sub-matrices are initialised
-        Mat sub;
-        MatNestGetSubMat(B, 0, 0, &sub);
-        MatGetInfo(sub, MAT_LOCAL, &info);
-    }
-    else
-    {
-        MatGetInfo(B, MAT_LOCAL, &info);
-    }
-
-    // Initialise the matrix if it has yet to be allocated; otherwise zero all
-    // entries
-    if (info.nz_used)
-    {
-        // Zero the matrix but do not reallocate the space
-        // The "-snes_lag_jacobian -2" PETSc option can be used to avoid
-        // re-building the matrix
-        // For a nested matrix, this will zero all sub-matrices
-        CHKERRQ(MatZeroEntries(B));
-    }
-    else
-    {
-        // Set the non-zero structure of the matrix B
-        Foam::Info<< "    Initialising the matrix..." << Foam::flush;
-        if (user->solMod_.initialiseJacobian(B) != 0)
-        {
-            Foam::FatalError
-                << "initialiseJacobian(B) returned an error code!"
-                << Foam::abort(Foam::FatalError);
-        }
-        Foam::Info<< "done" << Foam::endl;
-    }
+    // Zero the matrix but do not reallocate the space
+    // For a nested matrix, this will zero all sub-matrices
+    CHKERRQ(MatZeroEntries(B));
 
     // Populate the Jacobian => implemented by the solid model
     if (user->solMod_.formJacobian(B, xx) != 0)
@@ -148,17 +114,31 @@ PetscErrorCode formJacobianFoamPetscSnesHelper
 
 Foam::label Foam::initialiseJacobian
 (
-    Mat jac,
+    Mat& jac,
     const fvMesh& mesh,
-    const label blockSize
+    const label blockSize,
+    const bool createMat
 )
 {
+    // Count number of local blocks and local scalar equations
+    const label blockn = mesh.nCells();
+    const label n = blockn*blockSize;
+
+    // Global system size: total number of scalar equation across all
+    // processors
+    const label N = returnReduce(n, sumOp<label>());
+
+    // // Set the Jacobian matrix size
+    if (createMat)
+    {
+        MatCreate(PETSC_COMM_WORLD, &jac);
+        MatSetFromOptions(jac);
+        MatSetSizes(jac, n, n, N, N);
+        MatSetType(jac, MATMPIAIJ);
+    }
+
     // Set the block size
     CHKERRQ(MatSetBlockSize(jac, blockSize));
-
-    // Number of vector unknowns on this processor
-    //const int blockn = user->solMod_.globalCells().localSize();
-    const int blockn = mesh.nCells();
 
[…]
