+++ src/solids4FoamModels/solidModels/nonLinGeomUpdatedLagSolid/nonLinGeomUpdatedLagSolid.H
@@ -187,14 +187,19 @@ public:
                 return true;
             }
 
-    // Edit
+        // Edit
 
             //- Evolve the solid solver and solve the mathematical model
             virtual bool evolve();
 
             //- Initialise the Jacobian matrix
-            //  This procedure should set the non-zero structure of the matrix
-            virtual label initialiseJacobian(Mat jac);
+            //  This procedure should set the size and non-zero structure of
+            //  the matrix
+            virtual label initialiseJacobian(Mat& jac);
+
+            //- Initialise the solution vector
+            //  This procedure should set the size of the solution vector
+            virtual label initialiseSolution(Vec& x);
 
             //- Calculate the residual (PETSc array) given the solution (PETSc
             //  array) vector
