+++ src/solids4FoamModels/solidModels/solidModel/solidModel.H
@@ -57,6 +57,7 @@ SourceFiles
 #include "checkConvergence.H"
 #include "boolIOList.H"
 
+
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 namespace Foam

[…]
