### 2025-03-04 f99760d
Splited residual calculation
