+++ src/solids4FoamModels/numerics/fvc/fvcSurfaceGaussQuadIntegrate.C
@@ -1,169 +1,0 @@
-/*---------------------------------------------------------------------------*\
-License
-    This file is part of solids4foam.
-
-    solids4foam is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your
-    option) any later version.
-
-    solids4foam is distributed in the hope that it will be useful, but
-    WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-
-\*---------------------------------------------------------------------------*/
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace fvc
-{
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-
-template<class Type>
-tmp<GeometricField<Type, fvPatchField, volMesh>>
-surfaceGaussQuadIntegrate
-(
-    const List<List<symmTensor>>& sigmaGP,
-    const List<List<scalar>>>& gpW
-);
-{
-    // Here I need to perform Gaussian Quadrature integration. For the first
-    // iteration Gauss points and weights can be argument in surfaceGaussQuadIntegrate
-    // function.
-    const fvMesh& mesh = ssf.mesh();
-
-    tmp<GeometricField<Type, fvPatchField, volMesh>> tvf
-    (
-        new GeometricField<Type, fvPatchField, volMesh>
-        (
-            IOobject
-            (
-                "surfaceIntegrate("+ssf.name()+')',
-                ssf.instance(),
-                mesh,
-                IOobject::NO_READ,
-                IOobject::NO_WRITE
-            ),
-            mesh,
-            dimensioned<Type>(ssf.dimensions()/dimVol, Zero),
-            fvPatchFieldBase::extrapolatedCalculatedType()
-        )
-    );
-    GeometricField<Type, fvPatchField, volMesh>& vf = tvf.ref();
-
-    const labelUList& owner = mesh.owner();
-    const labelUList& neighbour = mesh.neighbour();
-
-    //const Field<Type>& issf = vf;
-
-    return tvf;
-}
-
-// template<class Type>
-// void surfaceGaussQuadIntegrate
-// (
-//     Field<Type>& ivf,
-//     const GeometricField<Type, fvsPatchField, surfaceMesh>& ssf
-// )
-// {
-//     const fvMesh& mesh = ssf.mesh();
-
-//     const labelUList& owner = mesh.owner();
-//     const labelUList& neighbour = mesh.neighbour();
-
-//     const Field<Type>& issf = ssf;
-
-//     forAll(owner, facei)
-//     {
-//         ivf[owner[facei]] += issf[facei];
-//         ivf[neighbour[facei]] -= issf[facei];
-//     }
-
-//     forAll(mesh.boundary(), patchi)
-//     {
-//         const labelUList& pFaceCells =
-//             mesh.boundary()[patchi].faceCells();
-
-//         const fvsPatchField<Type>& pssf = ssf.boundaryField()[patchi];
-
-//         forAll(mesh.boundary()[patchi], facei)
-//         {
[…]
