+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -41,6 +41,127 @@ addToRunTimeSelectionTable
 );
 
 
+// * * * * * * * * * * * Private Member Functions* * * * * * * * * * * * * * //
+
+void newtonCouplingInterface::makeMeshMotion() const
+{
+    if (meshMotionPtr_)
+    {
+        FatalErrorInFunction
+            << "Pointer already set!" << abort(FatalError);
+    }
+
+    meshMotionPtr_ =
+        solidModel::New(const_cast<Time&>(runTime()), "meshMotionFluid");
+}
+
+
+solidModel& newtonCouplingInterface::meshMotion()
+{
+    if (!meshMotionPtr_)
+    {
+        makeMeshMotion();
+    }
+
+    return meshMotionPtr_.ref();
+}
+
+
+void newtonCouplingInterface::updateMeshMotionInterfaceDisplacement()
+{
+    // Map solid interface motion to the mesh motion interface
+
+    // Lookup the interface map from the fluid faces to the solid faces
+    const interfaceToInterfaceMappings::
+        directMapInterfaceToInterfaceMapping& interfaceMap =
+        refCast
+        <
+            const interfaceToInterfaceMappings::
+            directMapInterfaceToInterfaceMapping
+        >
+        (
+            interfaceToInterfaceList()[0]
+        );
+
+    // The face map gives the solid face ID for each fluid face
+    const labelList& fluidFaceMap = interfaceMap.zoneBToZoneAFaceMap();
+
+    // Lookup the fluid mesh interface patch
+    const label fluidPatchID = fluidSolidInterface::fluidPatchIndices()[0];
+
+    // Lookup the mesh motion displacement field
+    fvPatchVectorField& motionPatchD =
+        meshMotion().D().boundaryFieldRef()[fluidPatchID];
+    if (!isA<fixedValueFvPatchVectorField>(motionPatchD))
+    {
+        FatalErrorInFunction
+            << "The meshMotionFluid interface patch must be of type "
+            << "'fixedValue'" << abort(FatalError);
+    }
+
+    // Lookup the solid interface patch
+    const label solidPatchID = fluidSolidInterface::solidPatchIndices()[0];
+    const fvPatchVectorField& solidPatchD =
+        solid().D().boundaryField()[solidPatchID];
+
+    // Map the solid interface displacement to the motion interface
+    forAll(motionPatchD, fluidFaceI)
+    {
+        const label solidFaceID = fluidFaceMap[fluidFaceI];
+
+        // Extrapolated patch value (larger stencil)
+        motionPatchD[fluidFaceI] = solidPatchD[solidFaceID];
+    }
+
+    if (interfaceToInterfaceList().size() != 1)
+    {
+        FatalErrorInFunction
+            << "Only one interface allowed" << abort(FatalError);
+    }
+
+    // Take references to zones
+    const standAlonePatch& fluidZone =
+        fluid().globalPatches()[0].globalPatch();
+    const standAlonePatch& solidZone =
+        solid().globalPatches()[0].globalPatch();
+
+    // Get the solid interface patch pointD field
+    const vectorField solidPatchPointD
+    (
+        solid().pointD().boundaryField()[solidPatchID].patchInternalField()
+    );
+
+    // Solid point zone pointD
+    const vectorField solidZonePointD
+    (
+        solid().globalPatches()[0].patchPointToGlobal(solidPatchPointD)
+    );
+
+    // Prepare the fluid mesh interface pointD field
+    vectorField fluidZonePointD(fluidZone.nPoints(), vector::zero);
+
+    // Transfer the point displacements
+    interfaceToInterfaceList()[0].transferPointsZoneToZone
+    (
+        solidZone,
+        fluidZone,
+        solidZonePointD,
+        fluidZonePointD
+    );
+
+    // Map the zone field to the patch
[…]
