### 2025-03-04 a4b19d8
newtonIcoFluid: progress towards ALE
