### 2025-03-06 08dea98
Added ghost point to QRFaceGPCoeffs
