### 2025-03-07 b101ea9
Remove trailing whitespace: addresses #162
Co-authored-by: abzrg <abzrg@users.noreply.github.com>
Co-authored-by: philipcardiff <philipcardiff@gmail.com>
