### 2025-03-11 16ad0a7
resolved bugs in fGradGP calculation
