### 2025-03-11 2e8c0b9
refactored residualMomentum
