+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -23,6 +23,9 @@ License
 #include "fixedValueFvPatchFields.H"
 #include "solidTractionFvPatchVectorField.H"
 #include "newtonIcoFluid.H"
+#include "dynamicMotionSolverFvMesh.H"
+#include "motionSolver.H"
+#include "meshMotionSolidModelFvMotionSolver.H"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
@@ -43,27 +46,1277 @@ addToRunTimeSelectionTable
 
 // * * * * * * * * * * * Private Member Functions* * * * * * * * * * * * * * //
 
-void newtonCouplingInterface::makeMeshMotion() const
+
+foamPetscSnesHelper& newtonCouplingInterface::motion()
 {
-    if (meshMotionPtr_)
+    // meshMotionSolidModelFvMotionSolver is the only currently supported
+    // mesh motion solver
+
+    if (!isA<dynamicMotionSolverFvMesh>(fluidMesh()))
     {
         FatalErrorInFunction
-            << "Pointer already set!" << abort(FatalError);
+            << "meshMotionSolidModelFvMotionSolver is the only currently "
+            << "supported mesh motion solver" << exit(FatalError);
     }
 
-    meshMotionPtr_ =
-        solidModel::New(const_cast<Time&>(runTime()), "meshMotionFluid");
+    if
+    (
+        !isA<meshMotionSolidModelFvMotionSolver>
+        (
+            refCast<dynamicMotionSolverFvMesh>(fluidMesh()).motion()
+        )
+    )
+    {
+        FatalErrorInFunction
+            << "meshMotionSolidModelFvMotionSolver is the only currently "
+            << "supported mesh motion solver" << exit(FatalError);
+    }
+
+    // Cast the fluid mesh to a dynamicMotionSolverFvMesh then access its motion
+    // solver and cast it to a foamPetscSnesHelper
+    // This will only suceed if using a motion solver based on a
+    // foamPetscSnesHelper solver
+    const foamPetscSnesHelper& motion =
+        refCast<const foamPetscSnesHelper>
+        (
+            refCast<const meshMotionSolidModelFvMotionSolver>
+            (
+                refCast<dynamicMotionSolverFvMesh>(fluidMesh()).motion()
+            ).model()
+        );
+
+    // Cast away the const-ness (sorry)
+    return const_cast<foamPetscSnesHelper&>(motion);
 }
 
 
-solidModel& newtonCouplingInterface::meshMotion()
+solidModel& newtonCouplingInterface::motionSolid()
 {
-    if (!meshMotionPtr_)
+    return refCast<solidModel>(motion());
+}
+
+
+label newtonCouplingInterface::initialiseAfm
+(
+    Mat Afm,
+    const fvMesh& fluidMesh,
+    const label fluidBlockSize,
+    const label motionBlockSize,
+    const bool twoD
+) const
+{
+    if (debug)
     {
-        makeMeshMotion();
+        InfoInFunction
+            << "Initialising" << endl;
     }
 
-    return meshMotionPtr_.ref();
+    // The motion appears as the mesh flux in the momentum convection and
+    // pressure equation flux divergence terms.
+    // This means each fluid equation is coupled to the mesh motion in all
+    // cells in the compact stencil
+
+    // CAREFUL: we are setting non-zeros here based on the scalar rows, not
+    // the block rows
+
+    // Set matrix type to AIJ (since BAIJ does not support non-square
+    // blocks)
+    CHKERRQ(MatSetType(Afm, MATMPIAIJ));
+
+    // Total number of scalar rows in Amf (same as in the fluid region)
+    const label scalarRowN = fluidMesh.nCells()*fluidBlockSize;
+
[…]
