+++ src/solids4FoamModels/materialModels/mechanicalModel/mechanicalLaws/nonLinearGeometryLaws/neoHookeanElastic/neoHookeanElastic.H
@@ -130,6 +130,9 @@ public:
         //- Return the bulk modulus field
         virtual tmp<volScalarField> bulkModulus() const;
 
+        //- Return the shear modulus field
+        virtual tmp<volScalarField> shearModulus() const;
+
         //- Calculate the volField stress
         virtual void correct(volSymmTensorField& sigma);
 
