### 2025-03-12 a6390eb
[Feature update] petsc snes higher order (#167)
* Added correct function with Gauss point stress and grad

* calculation of sigma at Gauss points

* Added divGaussPoints skeleton

* update divGaussPoints function

* Splited residual calculation

* Added ghost point to QRFaceGPCoeffs

* Removed info statement

* resolved bugs in fGradGP calculation

* refactored residualMomentum

---------

Co-authored-by: iBatistic <ivan.batistic@gmail.com>
