+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/higherOrderGrad/higherOrderGrad.H
@@ -229,9 +229,6 @@ class higherOrderGrad
         //- Return the face Gauss points
         const List<List<point>>& faceGaussPoints() const;
 
-        //- Return the face Gauss points weights
-        const List<List<scalar>>& faceGaussPointsWeight() const;
-
         //- Return the condition number field
         volScalarField& conditionNumber() const;
 
@@ -278,6 +275,12 @@ public:
 
 
     // Member Functions
+
+            //- Return number of Gauss points per triangle
+            label triQuadraturePtsNb() const {return triQuadraturePtsNb_;};
+
+            //- Return the face Gauss points weights
+            const List<List<scalar>>& faceGaussPointsWeight() const;
 
         // Edit
 
