+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -45,6 +45,82 @@ addToRunTimeSelectionTable(solidModel, linGeomTotalDispSolid, dictionary);
 
 // * * * * * * * * * * *  Private Member Functions * * * * * * * * * * * * * //
 
+tmp<surfaceVectorField> linGeomTotalDispSolid::GaussIntegrate
+(
+    const List<List<symmTensor>>& sigmaGP
+)
+{
+    tmp<surfaceVectorField> tsf
+    (
+        new surfaceVectorField
+        (
+            IOobject
+            (
+                "surfaceGaussIntegrate(traction)",
+                 mesh().time().timeName(),
+                 mesh(),
+                 IOobject::NO_READ,
+                 IOobject::NO_WRITE
+            ),
+            mesh(),
+            dimensionedVector("0", dimPressure, vector::zero)
+        )
+    );
+
+    surfaceVectorField& tf = tsf.ref();
+
+    const List<List<scalar>>& gpW = hoGradPtr_->faceGaussPointsWeight();
+
+    forAll (tf, faceI)
+    {
+        // Sigma at the Gauss quadrature points on the face
+        const List<symmTensor>& gaussQuadStress = sigmaGP[faceI];
+
+        // Gauss quadrature weights on the face
+        const List<scalar>& gaussQuadW = gpW[faceI];
+
+        const vector& faceNormal =
+            mesh().faceAreas()[faceI]/mag(mesh().faceAreas()[faceI]);
+
+        forAll(gaussQuadStress, pI)
+        {
+            // Add traction contribution of this quadrature point
+            tf[faceI] += faceNormal & (gaussQuadStress[pI] * gaussQuadW[pI]);
+        }
+    }
+
+    forAll (tf.boundaryField(), patchI)
+    {
+        vectorField& tfPatch = tf.boundaryFieldRef()[patchI];
+        forAll(tfPatch, faceI)
+        {
+            const label globalFaceID = mesh().boundaryMesh()[patchI].start() + faceI;
+
+            // Sigma at the Gauss quadrature points on the face
+            const List<symmTensor>& gaussQuadStress = sigmaGP[globalFaceID];
+
+            // Gauss quadrature weights on the face
+            const List<scalar>& gaussQuadW = gpW[globalFaceID];
+
+            const vector& faceNormal =
+                mesh().faceAreas()[globalFaceID]/mag(mesh().faceAreas()[globalFaceID]);
+
+            forAll(gaussQuadStress, pI)
+            {
+                // Add traction contribution of this quadrature point
+                tfPatch[faceI] += faceNormal & (gaussQuadStress[pI] * gaussQuadW[pI]);
+            }
+        }
+    }
+
+    // Unit normal vectors at the faces
+    // const vectorField n(mesh().faceAreaNormals());
+    // What if faces are not planar? Should each triangle have corresponding
+    // unit normal?
+
+    return tsf;
+};
+
 
 void linGeomTotalDispSolid::predict()
 {
@@ -465,6 +541,10 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
         dimensionedVector("zero", dimLength/pow(dimTime, 2), vector::zero)
     ),
     predictor_(solidModelDict().lookupOrDefault<Switch>("predictor", false)),
+    hoGradInt_
+    (
+        solidModelDict().lookupOrDefault<Switch>("higherOrderIntegration", false)
+    ),
     ds_
     (
         IOobject
@@ -478,7 +558,8 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
         mesh(),
         dimensionedScalar("ds", (dimForce/dimVolume)/dimVelocity, 1.0)
     ),
-    hoGradPtr_()
+    hoGradPtr_(),
[…]
