### 2025-03-13 57a84d8
linearElastic: speed up new correct function
