+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -514,6 +514,8 @@ label newtonCouplingInterface::formAfm
     const labelList& own = fluidMesh().owner();
     const labelList& nei = fluidMesh().neighbour();
 
+    // The fluid and mesh motion refer to the same mesh so they can use the same
+    // global cells object
     const globalIndex& globalCells =
         refCast<foamPetscSnesHelper>(fluid()).globalCells();
 
@@ -749,8 +751,6 @@ label newtonCouplingInterface::formAfm
 
         // Global block column ID
         const label globalBlockColI = globalCells.toGlobal(neiCellID);
-        Warning
-            << "CHECK WE ARE USING CORRECT GLOBAL CELLS" << endl;
 
         // The scalar row index for the pressure equation
         const label rowI =
@@ -1320,6 +1320,60 @@ label newtonCouplingInterface::formAsf
 }
 
 
+void newtonCouplingInterface::mapInterfaceSolidUToFluidU()
+{
+    // Lookup the interface map from the fluid faces to the solid faces
+    const interfaceToInterfaceMappings::
+        directMapInterfaceToInterfaceMapping& interfaceMap =
+        refCast
+        <
+            const interfaceToInterfaceMappings::
+            directMapInterfaceToInterfaceMapping
+        >
+        (
+            interfaceToInterfaceList()[0]
+        );
+
+    // The face map gives the solid face ID for each fluid face
+    const labelList& fluidFaceMap = interfaceMap.zoneBToZoneAFaceMap();
+
+    // Lookup the fluid interface patch
+    const label fluidPatchID = fluidSolidInterface::fluidPatchIndices()[0];
+    fvPatchVectorField& fluidPatchU =
+        fluid().U().boundaryFieldRef()[fluidPatchID];
+    if (!isA<fixedValueFvPatchVectorField>(fluidPatchU))
+    {
+        FatalErrorInFunction
+            << "The fluid interface patch must be of type 'fixedValue'"
+            << abort(FatalError);
+    }
+
+    // Lookup the solid interface velocity and displacement
+    const label solidPatchID = fluidSolidInterface::solidPatchIndices()[0];
+    const fvPatchVectorField& solidPatchU =
+        solid().U().boundaryField()[solidPatchID];
+
+    // Map the solid interface velocity to the fluid interface and the solid
+    // displacement to the motion interface
+    // const labelList& solidFaceCells =
+    //     solidMesh().boundary()[solidPatchID].faceCells();
+    // const vectorField& solidUI = solid().U();
+    forAll(fluidPatchU, fluidFaceI)
+    {
+        const label solidFaceID = fluidFaceMap[fluidFaceI];
+
+        // Extrapolated patch value (larger stencil)
+        fluidPatchU[fluidFaceI] = solidPatchU[solidFaceID];
+
+        // Adjacent cell value
+        // const label solidCellID = solidFaceCells[solidFaceID];
+        // fluidPatchU[fluidFaceI] = solidUI[solidCellID];
+    }
+
+    fluid().phi() = fvc::interpolate(fluid().U()) & fluidMesh().Sf();
+}
+
+
 void newtonCouplingInterface::updateMeshMotionInterfaceDisplacement()
 {
     // Map solid interface motion to the mesh motion interface
@@ -1483,14 +1537,6 @@ newtonCouplingInterface::newtonCouplingInterface
             << meshMotionSolidModelFvMotionSolver::typeName
             << exit(FatalError);
     }
-
-    // Not required as we manually move the mesh
-    // // Tell the motion solver not to solve its own equations and the motion
-    // // will be calculated here
-    // refCast<const meshMotionSolidModelFvMotionSolver>
-    // (
-    //     refCast<dynamicMotionSolverFvMesh>(fluidMesh()).motion()
[…]
