### 2025-03-16 #7407 @MakisH
**On the software paper:**

All previous points on the software paper are now fully resolved, thanks! :rocket:

Juggling a bit with many hats and potential conflict of interest here, but I think this helps your paper: in the features, you could mention somewhere that solids4foam already supports preCICE (and you have a tutorial for that). I think this is an important point, as it shows how "competitive" tools can also work together. And it adds additional flexibility points to solids4foam: even if you don't want to do the coupling in it, you can use it as a solid solver in a preCICE-coupled simulation (which might even include more physics than "just" a fluid and a solid). A fitting place might be the "Fluid Models" section: "even in case these are not enough, you can still couple it to standard OpenFOAM flow solvers or something else".

On the nitpicking side: "forks" might be a bit of a tense term in the OpenFOAM community, assuming that forking means starting something a new path (who forked from whom, who is the original?). One more neutral term might be "variants" (but happy to find an even better one). But I think it is already clear how this is meant.

> We have now added more explicit instructions for installing the `development` branch and have added the `development` to the **compatible OpenFOAM versions** table.

@philipcardiff The title is currently "solids4foam-v2.1: ...". Given these changes, and to make future citation easier, does it make sense to either:

- change the title "solids4foam v2: ...", or
- make a new v2.2 release before publishing (this might even be instructed by the journal) and update the title accordingly?

Especially the support for OpenFOAM v2412 already sounds very important to me.

But you should probably agree on a publication strategy for new versions first. Publishing a new reference paper for every new minor release might not be a good idea.
