### 2025-03-16 solids4foam/solids4foam#169 @MakisH
3dTube tutorial: Images not rendered on the website
https://www.solids4foam.com/tutorials/customWorkflowTutorial.html

![Image](https://github.com/user-attachments/assets/9a75f40a-a956-4f19-8933-60e0b207dd60)

The images of figure 3 and 4 (only) seem to be rendering on GitHub, but not on the website. The URLs seem to be correct in the `README.md`. Do they need to be declared anywhere else?
