### 2025-03-16 solids4foam/solids4foam#170 @MakisH
3dTube tutorial: undocumented IQN config differences
In the [3dTube tutorial](https://www.solids4foam.com/tutorials/tutorial4.html#figure-10-number-of-fsi-iterations-per-time-step-with-deltat--25e-5-s), it would be good to comment on the settings used for the IQN, since there are differences in the iteration numbers which are difficult to trace.

I think an important difference is that preCICE is configured to reuse 15 past time steps:

```xml
      <acceleration:IQN-ILS>
        <data name="Displacement" mesh="Solid-Mesh" />
        <data name="Force" mesh="Solid-Mesh" />
        <preconditioner type="residual-sum" />
        <filter type="QR2" limit="1e-2" />
        <initial-relaxation value="0.5" />
        <max-used-iterations value="100" />
        <time-windows-reused value="15" />
      </acceleration:IQN-ILS>
```

while solids4foam reuses 2 (if I understand the mapping correctly):

```c++
IQNILSCoeffs
{
    // Solid interface patch list
    solidPatch         inner-wall;

    // Fluid interface patch list
    fluidPatch         wall;

    // Under-relaxation factor for passing the solid interface
    // displacement/velocity to the fluid interface
    // For fixedRelaxation, this value is used for all iterations
    // For Aitken, this value is used for the first iteration each time-step,
    // then it adaptively changes
    // For IQNILS, this value is used for the first two iterations each
    // time-step, then the IQNILS procedure is used
    relaxationFactor   0.05;

    // Tolerance for the FSI loop within each time-step
    outerCorrTolerance 1e-6;

    // Maximum number of outer FSI loop correctors within each time-step
    nOuterCorr         30;

    // Enable/disable coupling from the start
    coupled             yes;

    // Optional: if coupled is set to "no", then the couplingStartTime also you
    // to enable coupling at specified time
    // Defaults to 0.0
    // couplingStartTime    0.1;

    // Optional (only for IQNILS): re-use modes from previous time-steps.
    // The number of modes to be re-used is specified
    // Defaults to 0 modes
    couplingReuse       2;

    // Optional: if any point on the fluid interface moves greater than the
    // interfaceDeformationLimit then the entire fluid mesh is updated; if not
    // then only the interface points are moved and the remaining fluid mesh is
    // not. Note than interfaceDeformationLimit is a relative motion: it is the
    // point displacement normalised with respect to the local cell size (delta
    // coeffs are used)
    // Defaults to 0.0 i.e. the fluid mesh is always updated
    interfaceDeformationLimit 0;

    // Frequency of time-step at which the fluid-solid interface interpolation
    // weights are updated
    // 0 means there will be no interpolator update during simulation, whereas
    // 10 means the interpolator will be updated every 10 time-steps
    // Defaults to 0
    interpolatorUpdateFrequency 0;

    // Write fsi residuals to file
    writeResidualsToFile yes;
}
```

An overview of these differences would be useful.
