### 2025-03-16 #7407 @MakisH
I had another read in the [3dTube tutorial](https://www.solids4foam.com/tutorials/tutorial4.html) and I opened two minor documentation-related issues:

https://github.com/solids4foam/solids4foam/issues/170

https://github.com/solids4foam/solids4foam/issues/169

My previous comments on this have also been addressed. Given that my latest comments on the paper were also minor, I think this is essentially ready from my side.
