### 2025-03-17 #7407 @mohd-afeef-badri
Dear @philipcardiff and co-authors,

Thank you for addressing all the points I previously raised. I  followed the new [tutorial](https://www.solids4foam.com/tutorials/customWorkflowTutorial.html) you provided and found it well-structured and easy to follow.

@kyleniemeyer, from my perspective, I recommend this paper for publication. I have no further questions or comments.
