### 2025-03-17 7f8cc3e
foamPetscSnesHelper: fix for reading options and symmetry plane convection Jacobian
