### 2025-03-18 3fb12de
fluidSolidInteraction/3dTube/README.md: add explanation of differences in preCICE and solids4foam IQN-ILS
