### 2025-03-18 solids4foam/solids4foam#170 @philipcardiff
Thanks Makis,

I have added the following text to the READMe.md to provide a basic explanation for the differences:
```
 The reason that the solids4foam and preCICE IQN-ILS implementations show
 different numbers of iterations may be attributed to differences in the coupling
 configurations and implementation; for example, the preCICE IQN-ILS is
 configured in this case to use the results from the previous 15 time steps,
 whereas solids4foam here uses results from two time steps; in addition,
 implementations of the residual convergence checks are different.
```

Let me know if you think other specific details should be mentioned.
