### 2025-03-18 d1646e1
3dTube: replace PDFs with PNGs due to rendering issues in some browsers
