### 2025-03-18 #7407 @philipcardiff
> I had another read in the [3dTube tutorial](https://www.solids4foam.com/tutorials/tutorial4.html) and I opened two minor documentation-related issues:
>
> [solids4foam/solids4foam#170](https://github.com/solids4foam/solids4foam/issues/170)
>
> [solids4foam/solids4foam#169](https://github.com/solids4foam/solids4foam/issues/169)
>
> My previous comments on this have also been addressed. Given that my latest comments on the paper were also minor, I think this is essentially ready from my side.

Thanks, @MakisH, I have addressed these points now (see my replies at the issues above).

I agree with your suggestions regarding the paper and will also incorporate these.
