### 2025-03-18 9e201fe
paper.md: mention preCICE coupling
