### 2025-03-18 ed79794
paper.md: replace fork with variant, and remove tmp file
