### 2025-03-18 fcdf1d0
paper.md: remove version in paper title
