### 2025-03-18 solids4foam/solids4foam#170 @MakisH
This should be clear enough already, thanks! What would also be useful would be links or generally references to the two files where these settings are set.
