### 2025-03-18 f668faa
3dTube/README.md: give the location of relevant IQN-ILS settings
