+++ tutorials/fluidSolidInteraction/3dTube/README.md
@@ -313,9 +313,10 @@ The predictions from all approaches agree closely. Examining the number of FSI
  The reason that the solids4foam and preCICE IQN-ILS implementations show
  different numbers of iterations may be attributed to differences in the coupling
  configurations and implementation; for example, the preCICE IQN-ILS is
- configured in this case to use the results from the previous 15 time steps,
- whereas solids4foam here uses results from two time steps; in addition,
- implementations of the residual convergence checks are different.
+ configured in this case to use the results from the previous 15 time steps
+ (`time-windows-reused` in `precice-config.xml`), whereas solids4foam here uses
+ results from two time steps (`couplingReuse` in `constant/fsiProperties`); in
+ addition, implementations of the residual convergence checks are different.
 
 ### Small Time Step
 
