### 2025-03-18 solids4foam/solids4foam#170 @philipcardiff
OK, I have add the setting name and file location in parentheses:
```
 The reason that the solids4foam and preCICE IQN-ILS implementations show
 different numbers of iterations may be attributed to differences in the coupling
 configurations and implementation; for example, the preCICE IQN-ILS is
 configured in this case to use the results from the previous 15 time steps
 (`time-windows-reused` in `precice-config.xml`), whereas solids4foam here uses
 results from two time steps (`couplingReuse` in `constant/fsiProperties`); in
 addition, implementations of the residual convergence checks are different.
```

More generally, I have started to look through the specific Python implementation of the Degrotte group at https://github.com/pyfsi/coconut and noticed some differences between their implementation and the one in solids4foam (which came from Extend-Bazaar and ultimately from Degroote himself before that, I think); I plan to look into this more and ensure the solid4foam implementation is consistent with the coconut implementation. If I find anything, I will update this and other tutorials as appropriate.
