### 2025-03-18 #7407 @kyleniemeyer
@philipcardiff I agree that the paper title does not need the version number; when we've had that in the past, it was to distinguish a new major version. So, while you could use `solids4foam-v3` in a future paper title, I don't think it's needed here.
