### 2025-03-18 solids4foam/solids4foam#170 @MakisH
Excellent, thank you! 😄
