### 2025-03-22 8c069cc
analytical solution for square plate
