+++ src/solids4FoamModels/functionObjects/squarePlateAnalyticalSolution/squarePlateAnalyticalSolution.C
@@ -1,0 +1,330 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+\*----------------------------------------------------------------------------*/
+
+#include "squarePlateAnalyticalSolution.H"
+#include "addToRunTimeSelectionTable.H"
+#include "volFields.H"
+#include "pointFields.H"
+#include "areaFields.H"
+#include "squarePlateStressDisplacement.H"
+
+// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
+
+namespace Foam
+{
+    defineTypeNameAndDebug(squarePlateAnalyticalSolution, 0);
+
+    addToRunTimeSelectionTable
+    (
+        functionObject,
+        squarePlateAnalyticalSolution,
+        dictionary
+    );
+}
+
+
+// * * * * * * * * * * * * * Private Member Functions  * * * * * * * * * * * //
+
+bool Foam::squarePlateAnalyticalSolution::writeData()
+{
+    // Lookup the solid mesh
+    const fvMesh* meshPtr = NULL;
+    if (time_.foundObject<fvMesh>("solid"))
+    {
+        meshPtr = &(time_.lookupObject<fvMesh>("solid"));
+    }
+    else
+    {
+        meshPtr = &(time_.lookupObject<fvMesh>("region0"));
+    }
+    const fvMesh& mesh = *meshPtr;
+
+    // Lookup the point mesh
+    const pointMesh& pMesh = mesh.lookupObject<pointMesh>("pointMesh");
+
+    // Cell-centres coordinates
+    const volVectorField& C = mesh.C();
+    const vectorField& CI = C.internalField();
+
+    // Point coordinates
+    const pointField& points = mesh.points();
+
+    // Pointer to analytical plate displacement function
+    scalar (*plateDisplacement)
+    (
+        const vector&,
+        const scalar,
+        const scalar,
+        const scalar,
+        const scalar,
+        const scalar,
+        const scalar
+    ) = nullptr;
+
+    if (boundaryConditions_ == "allEdgesClamped")
+    {
+        plateDisplacement = &allEdgesClampedMaxDisplacement;
+
+        // For all edges clamped getting analytical deflection field is quite
+        // complicated, so we will print only central deflection
+        vector zero = vector::zero;
+        const scalar maxD = plateDisplacement(zero, p_, E_, nu_, a_, b_, h_);
+
+        Info<<"Analytical solution for maximal plate deflection at centre "
+            << "point is: " << maxD << endl;
+
+        return true;
+    }
+    else if (boundaryConditions_ == "allEdgesSupported")
+    {
+        plateDisplacement = &allEdgesSupportedDisplacement;
+    }
+    else
+    {
+        FatalErrorInFunction
+            << "Unknown boundaryConditions: " << boundaryConditions_ << nl
+            << "Valid options are: allEdgesClamped or allEdgesSupported"
+            << exit(FatalError);
+    }
+
+    // Cell analytical fields
+    if (cellDisplacement_)
+    {
+        // Analytical displacement field
[…]
