+++ src/solids4FoamModels/functionObjects/README.md
@@ -61,7 +61,7 @@ $$
 u_y = \frac{P}{6EI} \left(3\nu y^2(L-x)+(4+5\nu)\frac{D^2x}{4}+(3L-x)x^2  \right),
 $$
 
-​ where $\nu$ is Poisson's ratio, $E$ is Young modulus, $I$ is the second moment
+ where $\nu$ is Poisson's ratio, $E$ is Young modulus, $I$ is the second moment
 of are of the cross-section, $P$ is applied load and $L$ is length of the beam.
 
 ```note
@@ -150,7 +150,7 @@ $$
   \qquad \sigma_z = \nu \sigma_y.
 $$
 
-​ where $$E$$ is Young's modulus, $$\nu$$ Poisson's ratio and $$\Delta$$
+ where $$E$$ is Young's modulus, $$\nu$$ Poisson's ratio and $$\Delta$$
 prescribed displacement of upper block top surface.
 
 ```note
@@ -229,7 +229,7 @@ $$
 -\frac{a^2+b^2}{r}\right)\cos (\theta),
 $$
 
-​ where $a$ is beam inner radius, $b$ is beam outer radius and $P$ is applied
+ where $a$ is beam inner radius, $b$ is beam outer radius and $P$ is applied
 shear force.
 
 - **Example of usage**
@@ -297,7 +297,7 @@ $$
 T = \displaystyle{\frac{\Delta T}{\ln \frac{b}{a}} \ln \frac{b}{r}},
 $$
 
-​ where $a$ is pipe inner radius, $b$ is pipe outer radius, $\nu$ is Poisson's
+ where $a$ is pipe inner radius, $b$ is pipe outer radius, $\nu$ is Poisson's
 ratio, $E$ is Young modulus, $\alpha$ is coefficient of linear thermal expansion
 and $\Delta T$ is temperature difference between inner and outer pipe surface.
 
@@ -372,7 +372,7 @@ $$
 + \frac{2a^2}{r^2} \right)\sin(2\theta),
 $$
 
-​ same in cartesian coordinates:
+ same in cartesian coordinates:
 
 $$
 \sigma_{xx} = T \left( 1-\frac{a^2}{r^2}\left(\frac{3}{2}\cos(2\theta)
@@ -389,7 +389,7 @@ $$
 +\sin(4\theta) \right) + \frac{3}{2}\frac{a^4}{r^4}\sin(4\theta) \right).
 $$
 
-​ Displacement field in cartesian coordinates:
+ Displacement field in cartesian coordinates:
 
 $$
 u_x = \frac{Ta}{8\mu}\left( \frac{r}{a}(\kappa+1)\cos\theta+\frac{2a}{r}
@@ -403,7 +403,7 @@ u_y = \frac{Ta}{8\mu}\left( \frac{r}{a}(\kappa-3)\sin\theta+\frac{2a}{r}
 -\frac{2a^3}{r^3}\sin(3\theta)  \right),
 $$
 
-​ where $$a$$ is hole radius, $$T$$ is far field traction in $$x$$ direction,
+ where $$a$$ is hole radius, $$T$$ is far field traction in $$x$$ direction,
 $$\nu$$ is Poisson's ratio, $$\mu$$ is shear modulus and $$\kappa$$ parameter is
 equal to $$3-4\nu$$.
 
@@ -928,7 +928,7 @@ equal to $$3-4\nu$$.
           type    solidPointDisplacementAlongLine;
           startPoint   (0 0 0);
           endPoint  (2 0 0);
-
+  
           // Optional
           minDist  1e-5;
           region  "solid";
@@ -1370,6 +1370,86 @@ This function object is currently only implemented for serial run!
 
[…]
