+++ src/solids4FoamModels/solidModels/diffusionSolid/diffusionSolid.H
@@ -1,0 +1,196 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+Class
+    diffusionSolid
+
+Description
+    A simple solid model represented by the diffusion equation. This is
+    equivalent to assuming the stress is given as sigma = impK*grad(D).
+    This solid model was created for checking the Jacobian is correct in
+    Newton-Raphson procedures, as its Jacobian is simple.
+
+Author
+    Philip Cardiff, UCD.  All rights reserved.
+
+SourceFiles
+    diffusionSolid.C
+
+\*---------------------------------------------------------------------------*/
+
+#ifndef diffusionSolid_H
+#define diffusionSolid_H
+
+#include "solidModel.H"
+#include "volFields.H"
+#include "surfaceFields.H"
+#include "pointFields.H"
+#include "uniformDimensionedFields.H"
+#include "foamPetscSnesHelper.H"
+
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+namespace Foam
+{
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+namespace solidModels
+{
+
+/*---------------------------------------------------------------------------*\
+                        Class diffusionSolid Declaration
+\*---------------------------------------------------------------------------*/
+
+class diffusionSolid
+:
+    public solidModel,
+    public foamPetscSnesHelper
+{
+    // Private data
+
+        //- Implicit stiffness; coefficient of the Laplacian term
+        //  The value of this term only affects convergence and not the answer
+        //  In the case of Hookean elasticity, (2*mu + lambda) is the optimum
+        const volScalarField impK_;
+
+        //- Implicit stiffness surface field
+        surfaceScalarField impKf_;
+
+        //- Reciprocal of impK
+        //  We divide by impK every time tractionBoundarySnGrad is called so we
+        //  will store the reciprocal for effficiency
+        const volScalarField rImpK_;
+
+        //- Number of unknowns per cell, used by PETSc SNES solver
+        const label blockSize_;
+
+
+    // Private Member Functions
+
+        //- Enforce traction boundary conditions on the provided traction field
+        void enforceTractionBoundaries
+        (
+            surfaceVectorField& traction,
+            const volVectorField& D,      // displacement field
+            const surfaceVectorField& n   // unit normals field
+        ) const;
+
+        //- Disallow default bitwise copy construct
+        diffusionSolid(const diffusionSolid&);
+
+        //- Disallow default bitwise assignment
+        void operator=(const diffusionSolid&);
+
+
+protected:
+
+    // Protected member functions
+
+        //- Return nonlinear geometry enumerator
+        virtual nonLinearGeometry::nonLinearType nonLinGeom() const
+        {
+            return nonLinearGeometry::LINEAR_GEOMETRY;
+        }
+
+
+public:
+
+    //- Runtime type information
+    TypeName("diffusionSolid");
+
+    // Constructors
+
+        //- Construct from components
+        diffusionSolid
[…]
