+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -197,6 +197,7 @@ label newtonCouplingInterface::initialiseAfm
 
     // Raise an error if mallocs are required during matrix assembly
     CHKERRQ(MatSetOption(Afm, MAT_NEW_NONZERO_ALLOCATION_ERR, PETSC_TRUE));
+    //CHKERRQ(MatSetOption(Afm, MAT_NEW_NONZERO_ALLOCATION_ERR, PETSC_FALSE));
 
     return 0;
 }
@@ -215,7 +216,7 @@ label newtonCouplingInterface::initialiseAfs
     if (debug)
     {
         InfoInFunction
-            << "Initialising" << endl;
+            << "Initialising Afs" << endl;
     }
 
     // Initially we assume a conformal FSI interface, where each fluid cell
@@ -258,7 +259,7 @@ label newtonCouplingInterface::initialiseAfs
     std::vector<int> o_nnz(scalarRowN, 0);
 
     // Set non-zeros for each interface fluid cells
-    const labelList& fluidFaceMap = interfaceMap.zoneAToZoneBFaceMap();
+    const labelList& fluidFaceMap = interfaceMap.zoneBToZoneAFaceMap();
     const label fluidPatchID = fluidSolidInterface::fluidPatchIndices()[0];
     const labelList& fluidFaceCells =
         fluidMesh.boundary()[fluidPatchID].faceCells();
@@ -266,20 +267,23 @@ label newtonCouplingInterface::initialiseAfs
     {
         const label fluidCellID = fluidFaceCells[fluidFaceI];
 
-        // Calculate the row index for this cells first scalar equation
-        label rowI = fluidCellID*fluidBlockSize;
+        // Calculate the row index for this cell's first scalar equation
+        const label rowI = fluidCellID*fluidBlockSize;
 
         // Set the number of non-zeros to be the number of solid equations
         // e.g., The x-momentum equation could have a coefficient for the
         // solid x/y/z displacement
-        d_nnz[rowI++] = solidBlockSize;
-        d_nnz[rowI++] = solidBlockSize;
-        d_nnz[rowI++] = solidBlockSize;
 
+        // Momentum equation
+        d_nnz[rowI] = 1;
+        d_nnz[rowI + 1] = 1;
         if (!twoD)
         {
-            d_nnz[rowI++] = solidBlockSize;
+            d_nnz[rowI + 2] = 1;
         }
+
+        // Pressure equation
+        d_nnz[rowI + fluidBlockSize - 1] = solidBlockSize;
     }
 
     // Parallel: we need to decide how to deal with this, e.g. do we allow
@@ -300,6 +304,7 @@ label newtonCouplingInterface::initialiseAfs
 
     // Raise an error if mallocs are required during matrix assembly
     CHKERRQ(MatSetOption(Afs, MAT_NEW_NONZERO_ALLOCATION_ERR, PETSC_TRUE));
+    //CHKERRQ(MatSetOption(Afs, MAT_NEW_NONZERO_ALLOCATION_ERR, PETSC_FALSE));
 
     return 0;
 }
@@ -360,7 +365,7 @@ label newtonCouplingInterface::initialiseAms
     std::vector<int> o_nnz(scalarRowN, 0);
 
     // Set non-zeros for each interface fluid cells
-    const labelList& fluidFaceMap = interfaceMap.zoneAToZoneBFaceMap();
+    const labelList& fluidFaceMap = interfaceMap.zoneBToZoneAFaceMap();
     const label fluidPatchID = fluidSolidInterface::fluidPatchIndices()[0];
     const labelList& fluidFaceCells =
         fluidMesh.boundary()[fluidPatchID].faceCells();
[…]
