### 2025-03-23 5108f18
Add fluid-solid monolithic system with outer mesh motion loop
