+++ src/solids4FoamModels/fluidSolidInterfaces/newtonCouplingInterface/newtonCouplingInterface.C
@@ -96,6 +96,120 @@ solidModel& newtonCouplingInterface::motionSolid()
 }
 
 
+void newtonCouplingInterface::createSubMatsAndMat
+(
+    Mat& jac,
+    Mat*& subMatsPtr,
+    const labelPairList& nBlocksAndBlockSize,
+    const labelPairHashSet& nullSubMats
+) const
+{
+    // Set the number of regions
+    const label nRegions = nBlocksAndBlockSize.size();
+
+    // Create arrays (vectors) of IS objects for rows and columns.
+    // These will indicate where in the matrix the different regions are
+    // located
+    std::vector<IS> isRow(nRegions), isCol(nRegions);
+    label globalOffset = 0;
+    for (label r = 0; r < nRegions; ++r)
+    {
+        const label nBlocks = nBlocksAndBlockSize[r].first();
+        const label blockSize = nBlocksAndBlockSize[r].second();
+        const label regionSize = nBlocks*blockSize;
+
+        // Create an IS that covers the indices for region r
+        ISCreateStride
+        (
+            PETSC_COMM_WORLD, regionSize, globalOffset, 1, &isRow[r]
+        );
+        ISCreateStride
+        (
+            PETSC_COMM_WORLD, regionSize, globalOffset, 1, &isCol[r]
+        );
+        globalOffset += regionSize;
+    }
+
+    // Create an array of submatrices.
+    // Each submatrix represents the coupling between region i and
+    // region j.
+    // For example, subMat[i*nRegions + i] might be the matrix for
+    // region i, while subMat[i*nRegions + j] (i != j) are the coupling
+    // matrices.
+    subMatsPtr = new Mat[nRegions*nRegions];
+    for (label i = 0; i < nRegions; ++i)
+    {
+        for (label j = 0; j < nRegions; ++j)
+        {
+            // Create the submatrix for regions i and j.
+            Mat subA = nullptr;
+
+            // Skip this submatrix as it is specified as null
+            labelPair subMatID(i, j);
+            if (nullSubMats.found(subMatID))
+            {
+                continue;
+            }
+
+            MatCreate(PETSC_COMM_WORLD, &subA);
+
+            // Number of rows
+            const label nRowBlocks = nBlocksAndBlockSize[i].first();
+            const label rowBlockSize = nBlocksAndBlockSize[i].second();
+            const label nRowsLocal = nRowBlocks*rowBlockSize;
+            const label nRowsGlobal = returnReduce(nRowsLocal, sumOp<label>());
+
+            // Number of columns
+            const label nColBlocks = nBlocksAndBlockSize[j].first();
+            const label colBlockSize = nBlocksAndBlockSize[j].second();
+            const label nColsLocal = nColBlocks*colBlockSize;
+            const label nColsGlobal = returnReduce(nColsLocal, sumOp<label>());
+
+            if (debug)
+            {
+                Info<< "subMat(" << i << "," << j << ") " << nl
+                    << "nRowsLocal = " << nRowsLocal << nl
+                    << "nColsLocal = " << nColsLocal << endl;
+            }
+
+            MatSetSizes
+            (
+                subA,
+                nRowsLocal,
+                nColsLocal,
+                nRowsGlobal,
+                nColsGlobal
+            );
+            MatSetFromOptions(subA);
+            MatSetType(subA, MATMPIAIJ);
+
+            // Store subA in row-major order at index [i*nRegions + j]
+            subMatsPtr[i*nRegions + j] = subA;
+        }
+    }
+
+    // Create the nest matrix using the index sets and submatrices
+    jac = Mat();
+    MatCreateNest
+    (
+        PETSC_COMM_WORLD,
+        nRegions,
[…]
