### 2025-03-23 2a82bac
Added linear predictor for solution in newtonCouplingInterface
