### 2025-03-24 2749566
functionObjects/README.md: lint fixes
