### 2025-03-24 29d0f87
Compilation fix OF9: disable squarePlateAnalyticalSolution
