### 2025-03-24 f43b659
Compilation fix for FE41: squarePlateAnalyticalSolution
