### 2025-03-24 d72ec1b
workflows/removeTrailingWhitespace: fix for pull request
