### 2025-03-24 d530fb0
workflows/removeTrailingWhitespace: just give error; don't fix
