### 2025-03-24 b5688e5
[Feature] squarePlateAnalyticalSolution (#171)
* analytical solution for square plate

* update README

* functionObjects/README.md: lint fixes

* Reorder stressTriaxiality in functionObjects/README.md to ensure alphabetical ordering

* Update authors

* Compilation fix OF9: disable squarePlateAnalyticalSolution

* Compilation fix for OF2012: perturbMeshPoints

* Compilation fix for FE41: squarePlateAnalyticalSolution

* workflows/removeTrailingWhitespace: explicit specify the branch when pushing

* workflows/removeTrailingWhitespace: fix for pull request

* workflows/removeTrailingWhitespace:
[…]
