+++ .github/workflows/removeTrailingWhitespace.yml
@@ -1,5 +1,5 @@
 # Removing trailing whitespace from all files
-name: Remove trailing whitespace
+name: Check for trailing whitespace
 
 on:
   push:
@@ -11,7 +11,7 @@ on:
   workflow_dispatch:
 
 jobs:
-  trim-whitespace:
+  check-whitespace:
     runs-on: ubuntu-latest
 
     steps:
@@ -19,18 +19,13 @@ jobs:
       with:
         fetch-depth: 0
 
-    - name: Remove trailing whitespace
+    - name: Check for trailing whitespace
       run: |
-        git config --global user.name 'github-actions[bot]'
-        git config --global user.email 'github-actions[bot]@users.noreply.github.com'
-        git grep -I -l '[[:blank:]]\+$' | xargs -r sed -i 's/[[:blank:]]\+$//'
-
-    - name: Commit and push changes if needed
-      run: |
-        if [[ $(git status --porcelain) ]]; then
-          git add -A
-          git commit -m "chore: remove trailing whitespace [automated]"
-          git push
+        # Find files with trailing whitespace
+        if git grep -I -n '[[:blank:]]\+$'
+        then
+          echo "❌ Trailing whitespace found in the files above. Please fix them manually."
+          exit 1
         else
-          echo "No trailing whitespace found."
+          echo "✅ No trailing whitespace found."
         fi
