### 2025-03-24 c50da1d
newtonCouplingInterface: add adaptive time step based on newton iterations
