+++ CONTRIBUTORS.md
@@ -34,3 +34,4 @@ missing, please let <philip.cardiff@ucd.ie> know.
 - Mark Olesen
 - Sandro Manservisi
 - Greg Burgreen
+- Ali Bozorgzadeh
