+++ src/solids4FoamModels/solidModels/solidModel/foamPetscSnesHelper/foamPetscSnesHelperTemplates.C
@@ -31,7 +31,7 @@ namespace Foam
 template <class Type>
 label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
 (
-    const fvMatrix<Type>& fvM,
+    fvMatrix<Type>& fvM,
     Mat jac,
     const label rowOffset,
     const label colOffset,
@@ -57,8 +57,13 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
         {
             // Obtain a pointer to the underlying scalar data in
             // diag[blockRowI] (assumes contiguous storage)
+#ifdef OPENFOAM_COM
             const scalar* curDiagPtr =
                 reinterpret_cast<const scalar*>(diag[blockRowI].begin());
+#else
+            const scalar* curDiagPtr =
+                &diag[blockRowI].component(0);
+#endif
 
             // Construct the diag block coefficient
             for (label cmptI = 0; cmptI < nScalarEqns; ++cmptI)
@@ -154,8 +159,18 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
             // Take a copy of the faceCells (local IDs) and convert them to
             // global IDs
             labelList globalFaceCells(mesh.boundary()[patchI].faceCells());
+#ifdef OPENFOAM_COM
             foamPetscSnesHelper::globalCells().inplaceToGlobal(globalFaceCells);
-
+#else
+            forAll(globalFaceCells, fI)
+            {
+                globalFaceCells[fI] =
+                    foamPetscSnesHelper::globalCells().toGlobal
+                    (
+                        globalFaceCells[fI]
+                    );
+            }
+#endif
             // Send global IDs to the neighbour proc
             const processorFvPatch& procPatch =
                 refCast<const processorFvPatch>(mesh.boundary()[patchI]);
@@ -209,11 +224,16 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
                 {
                     // Obtain a pointer to the underlying scalar data in
                     // intCoeffs[faceI] (assumes contiguous storage)
+#ifdef OPENFOAM_COM
                     const scalar* curIntCoeffsPtr =
                         reinterpret_cast<const scalar*>
                         (
                             intCoeffs[faceI].begin()
                         );
+#else
+                    const scalar* curIntCoeffsPtr =
+                        &intCoeffs[faceI].component(0);
+#endif
 
                     for (label cmptI = 0; cmptI < nScalarEqns; ++cmptI)
                     {
@@ -243,11 +263,16 @@ label foamPetscSnesHelper::InsertFvMatrixIntoPETScMatrix
                 {
                     // Obtain a pointer to the underlying scalar data in
                     // neiCoeffs[faceI] (assumes contiguous storage)
+#ifdef OPENFOAM_COM
                     const scalar* curNeiCoeffsPtr =
                         reinterpret_cast<const scalar*>
                         (
                             neiCoeffs[faceI].begin()
                         );
+#else
+                    const scalar* curNeiCoeffsPtr =
+                        &neiCoeffs[faceI].component(0);
+#endif
 
                     for (label cmptI = 0; cmptI < nScalarEqns; ++cmptI)
                     {
@@ -303,7 +328,11 @@ void foamPetscSnesHelper::ExtractFieldComponents
     // Obtain a pointer to the underlying scalar array in vf
     // We assume vf is stored as contiguous data, i.e.
     // (vx1, vy1, vz1, vx2, vy2, vz2, ..., vxN, vyN, vzN)
+#ifdef OPENFOAM_COM
[…]
