### 2025-03-31 67d4547
blockCoupledSolids4FoamTools: compilation fix for FE50
