### 2025-03-31 3908902
FE50 compilation fixes
