+++ src/solids4FoamModels/solidModels/linGeomTotalDispSolid/linGeomTotalDispSolid.C
@@ -86,8 +86,13 @@ void linGeomTotalDispSolid::enforceTractionBoundaries
 
             const vectorField& nPatch = n.boundaryField()[patchI];
 
+#ifdef OPENFOAM_NOT_EXTEND
             traction.boundaryFieldRef()[patchI] =
                 tracPatch.traction() - nPatch*tracPatch.pressure();
+#else
+            traction.boundaryField()[patchI] =
+                tracPatch.traction() - nPatch*tracPatch.pressure();
+#endif
         }
         else if
         (
@@ -105,8 +110,13 @@ void linGeomTotalDispSolid::enforceTractionBoundaries
             const vectorField& nPatch = n.boundaryField()[patchI];
 
             // Set shear traction to zero
+#ifdef OPENFOAM_NOT_EXTEND
             traction.boundaryFieldRef()[patchI] =
                 sqr(nPatch) & traction.boundaryField()[patchI];
+#else
+            traction.boundaryField()[patchI] =
+                sqr(nPatch) & traction.boundaryField()[patchI];
+#endif
         }
     }
 }
@@ -204,14 +214,23 @@ bool linGeomTotalDispSolid::evolveImplicitSegregated()
                     (
                         magSqr
                         (
+#ifdef OPENFOAM_NOT_EXTEND
                             D().primitiveField()
                           - D().prevIter().primitiveField()
+#else
+                            D().internalField()
+                          - D().prevIter().internalField()
+#endif
                         )
                     )
                 );
 
             // Norm of the solution
+#ifdef OPENFOAM_NOT_EXTEND
             xNorm = sqrt(gSum(magSqr(D().primitiveField())));
+#else
+            xNorm = sqrt(gSum(magSqr(D().internalField())));
+#endif
 
             // Store the initial residual
             if (iCorr == 0)
@@ -293,7 +312,11 @@ bool linGeomTotalDispSolid::evolveSnes()
         // Map the D field to the SNES solution vector
         foamPetscSnesHelper::InsertFieldComponents<vector>
         (
+#ifdef OPENFOAM_NOT_EXTEND
             D().primitiveFieldRef(),
+#else
+            D().internalField(),
+#endif
             foamPetscSnesHelper::solution(),
             solidModel::twoD() ? 2 : 3, // Block size of x
             0                           // Location of first component
@@ -308,7 +331,11 @@ bool linGeomTotalDispSolid::evolveSnes()
     foamPetscSnesHelper::ExtractFieldComponents<vector>
     (
         foamPetscSnesHelper::solution(),
+#ifdef OPENFOAM_NOT_EXTEND
         D().primitiveFieldRef(),
+#else
+        D().internalField(),
+#endif
         solidModel::twoD() ? 2 : 3, // Block size of x
         0                           // Location of first component
     );
@@ -384,7 +411,11 @@ bool linGeomTotalDispSolid::evolveExplicit()
         {
             if (mesh.geometricD()[dirI] < 0)
             {
+#ifdef OPENFOAM_NOT_EXTEND
                 D.primitiveFieldRef().replace(dirI, 0.0);
+#else
+                D.internalField().replace(dirI, 0.0);
+#endif
             }
         }
     }
@@ -509,7 +540,11 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
     // Check the gradScheme
     const word gradDScheme
     (
+#ifdef OPENFOAM_NOT_EXTEND
         mesh().gradScheme("grad(" + D().name() +')')
+#else
+        mesh().schemesDict().gradScheme("grad(" + D().name() +')')
+#endif
     );
 
     if
@@ -553,7 +588,11 @@ linGeomTotalDispSolid::linGeomTotalDispSolid
[…]
