+++ src/solids4FoamModels/solidModels/coupledNonLinGeomPressureDisplacementSolid/coupledNonLinGeomPressureDisplacementSolid.H
@@ -1,237 +1,0 @@
-/*---------------------------------------------------------------------------*\
-License
-    This file is part of solids4foam.
-
-    solids4foam is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by the
-    Free Software Foundation, either version 3 of the License, or (at your
-    option) any later version.
-
-    solids4foam is distributed in the hope that it will be useful, but
-    WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
-
-Class
-    linGeomPressureVelocitySolid
-
-Description
-    Mathematical model where nonlinear geometry is assumed i.e. large strains
-    are assumed, and the total displacement and hydrostatic pressure are the
-    primary unknowns. The displacement-pressure coupling is treated implicitly
-    in a block-coupled manner. The transpose Laplacian of the momentum
-    equations is assembled by decomposing the linear version in the tangential
-    and normal components.
-
-    Rhie-Chow stabilisation is applied directly to the pressure and displcement
-    fields.
-
-    The stress is calculated by the run-time selectable mechanical law.
-
-Author
-    Zeljko Tukovic, FSB Zagreb.  All rights reserved.
-    Philip Cardiff, UCD.  All rights reserved.
-
-SourceFiles
-    coupledNonLinGeomPressureDisplacementSolid.C
-
-\*---------------------------------------------------------------------------*/
-
-#ifndef coupledNonLinGeomPressureDisplacementSolid_H
-#define coupledNonLinGeomPressureDisplacementSolid_H
-
-#include "solidModel.H"
-#include "volFields.H"
-#include "surfaceFields.H"
-#include "pointFields.H"
-#include "volVectorNFields.H"
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace solidModels
-{
-
-/*---------------------------------------------------------------------------*\
-                        Class linGeomTotalDispSolid Declaration
-\*---------------------------------------------------------------------------*/
-
-class coupledNonLinGeomPressureDisplacementSolid
-:
-    public solidModel
-{
-    // Private data
-
-        //- Cauchy (true) stress surface field
-        surfaceSymmTensorField sigmaf_;
-
-        //- Surface displacement gradient
-        surfaceTensorField gradDf_;
-
-        //- Total deformation gradient
-        volTensorField F_;
-
-        //- Total deformation gradient surface field
-        surfaceTensorField Ff_;
-
-        //- Inverse of the total deformation gradient
-        volTensorField Finv_;
-
-        //- Inverse of the total deformation gradient
-        surfaceTensorField Finvf_;
-
-        //- Jacobian of total deformation gradient
-        volScalarField J_;
-
-        //- Surface Jacobian of total deformation gradient
-        surfaceScalarField Jf_;
-
-        //- Implicit stiffness; coefficient of the Laplacian term
-        //  The value of this term only affects convergence and not the answer
-        //  In the case of Hookean elasticity, (2*mu + lambda) is the optimum
-        const volScalarField impK_;
-
-        //- Implicit stiffness surface field
-        surfaceScalarField impKf_;
-
-        //- Reciprocal of impK
-        //  We divide by impK every time tractionBoundarySnGrad is called so we
-        //  will store the reciprocal for effficiency
[…]
