+++ src/solids4FoamModels/numerics/meshDual/meshDual.C
@@ -19,11 +19,14 @@ License
 
 #include "meshDual.H"
 #include "meshDualiser.H"
-#include "polyTopoChange.H"
 #include "pointMesh.H"
 #include "pointFields.H"
 #include "meshTools.H"
-
+#ifdef OPENFOAM_NOT_EXTEND
+    #include "polyTopoChange.H"
+#else
+    #include "directTopoChange.H"
+#endif
 
 // * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
 
@@ -457,10 +460,17 @@ Foam::meshDual::meshDual(const fvMesh& mesh, const dictionary& dict)
             Foam::IOobject::READ_IF_PRESENT,
             Foam::IOobject::NO_WRITE
         ),
+#ifdef OPENFOAM_NOT_EXTEND
         pointField(mesh.points()),
         faceList(mesh.faces()),
         labelList(mesh.faceOwner()),
         labelList(mesh.faceNeighbour()),
+#else
+        xferCopy(mesh.points()),
+        xferCopy(mesh.faces()),
+        xferCopy(mesh.faceOwner()),
+        xferCopy(mesh.faceNeighbour()),
+#endif
         true
     ),
     mesh_(mesh),

[…]
