### 2025-03-31 8ae6d03
FE50 compilation fixes and tidying
