### 2025-04-01 871017c
Update paper.md and paper.bib
