+++ paper/paper.bib
@@ -190,7 +190,6 @@ Year = {1989--2017}
 	volume = {136},
 	doi = {10.1016/j.jmbbm.2022.105498},
 	url = {https://www.scopus.com/inward/record.uri?eid=2-s2.0-85139832727&doi=10.1016%2fj.jmbbm.2022.105498&partnerID=40&md5=15ed032f7053f474031f945afae34a1f},
-	type = {Article},
 	publication_stage = {Final},
 	source = {Scopus},
 	note = {Cited by: 8; All Open Access, Green Open Access}
@@ -204,7 +203,6 @@ Year = {1989--2017}
 	volume = {163},
 	doi = {10.1016/j.compbiomed.2023.107178},
 	url = {https://www.scopus.com/inward/record.uri?eid=2-s2.0-85163170668&doi=10.1016%2fj.compbiomed.2023.107178&partnerID=40&md5=1785104da461fde5fe43cb85fecdec8c},
-	type = {Article},
 	publication_stage = {Final},
 	source = {Scopus},
 	note = {Cited by: 0; All Open Access, Green Open Access}
@@ -219,7 +217,6 @@ Year = {1989--2017}
 	pages = {673 – 693},
 	doi = {10.1016/j.apm.2021.09.009},
 	url = {https://www.scopus.com/inward/record.uri?eid=2-s2.0-85116356495&doi=10.1016%2fj.apm.2021.09.009&partnerID=40&md5=a61dff60271c00963eac526d1c905b82},
-	type = {Article},
 	publication_stage = {Final},
 	source = {Scopus},
 	note = {Cited by: 3}
@@ -235,7 +232,6 @@ Year = {1989--2017}
 	pages = {4171 – 4191},
 	doi = {10.1002/nme.7302},
 	url = {https://www.scopus.com/inward/record.uri?eid=2-s2.0-85161647124&doi=10.1002%2fnme.7302&partnerID=40&md5=a5179c3afadb764134dfcddcc4ea1d5c},
-	type = {Article},
 	publication_stage = {Final},
 	source = {Scopus},
 	note = {Cited by: 1; All Open Access, Hybrid Gold Open Access}
