+++ tutorials/solids/beamsPlatesShells/squarePlate/README.md
@@ -1,3 +1,7 @@
+---
+sort: 1
+---
+
 # Square plate with transverse pressure: `squarePlate`
 
 ---
@@ -8,111 +12,145 @@ Prepared by Philip Cardiff and Ivan Batistić
 
 ## Tutorial Aims
 
-- Demonstrate the analysis of a simple plate-bending problem using Kirchoff-Love plate model.
+- Demonstrate analysing a simple plate-bending problem using a finite area
+  implementation of a Kirchoff-Love plate model.
 
 ---
 
-## Some caveats on plate formulation
+```warning
+This case currently only runs with foam extend as it uses the foam extend
+finite area discretisation framework.
+```
 
-Two types of beam/shell theory are widely used:
+## Plate Theory Assumptions
 
-- **Euler–Bernoulli** beam theory, corresponding to **Kirchhoff–Love** shell theory
-  - the planes normal to the midline are assumed to remain plane and normal (no shear stress); this is also called engineering
-    beam theory 	
+Two types of beam/plate/shell theory are widely used:
+
+- **Euler–Bernoulli** beam theory, corresponding to **Kirchhoff–Love** shell
+  theory
+
+  - The planes normal to the midline are assumed to remain plane and normal (no
+    shear stress); this is also called engineering beam theory.
 
 - **Timoshenko beam** theory, corresponding to **Mindlin–Reissner** shell theory
-  - the planes normal to the midline are assumed to remain plane, but not necessarily normal (shear stress may be non-zero); this
-    is also called shear beam theory.
+  - The planes normal to the midline are assumed to remain plane but not
+    necessarily normal (shear stress may be non-zero); this is also called shear
+    beam theory.
 
-Kirchhoff–Love shell theory is a subset of Mindlin–Reissner shell theory i.e. Mindlin–Reissner is applicable is every case that Kirchhoff–Love shell theory is applicable, but Kirchhoff–Love theory is not application in every case that Mindlin–Reissner theory is applicable.
+Kirchhoff–Love shell theory is a subset of Mindlin–Reissner shell theory, i.e.
+Mindlin–Reissner is applicable in every case that Kirchhoff–Love shell theory is
+applicable, but Kirchhoff–Love theory is not applicable in every case that
+Mindlin–Reissner theory is applicable. The following table demonstrates the
+valid length $$L$$ to thickness $$h$$ ranges for Kirchhoff and Mindlin
+approaches relative to a 3-D continuum approach.
 
 |                                   | Kirchhoff | Mindlin | 3-D continuum |
 | --------------------------------- | :-------: | :-----: | :-----------: |
-| **thin**: $$L/h > 10$$            |     ✓     |    ✓    |       ✓       |
-| **moderately thick**: $$L/h > 5$$ |     x     |    ✓    |       ✓       |
-| **thick**: $$L/h < 5$$            |     x     |    x    |       ✓       |
+| **Thin**: $$L/h > 10$$            |     ✓     |    ✓    |       ✓       |
+| **Moderately thick**: $$L/h > 5$$ |     x     |    ✓    |       ✓       |
+| **Thick**: $$L/h < 5$$            |     x     |    x    |       ✓       |
 
-**Kirchhoff-Love plate formulation**   
+Kirchhoff-Love Plate Formulation
 
-For thin plates $$(L/h > 10)$$, Kirchhoff–Love shell theory allows conservation of momentum is be reformulated into a fourth-order Laplacian equation, where the unknown scalar, $$w$$, is the transverse displacement (displacement normal to the plate):
+For thin plates $$(L/h > 10)$$, Kirchhoff–Love shell theory allows the
+conservation of momentum to be reformulated into a fourth-order partial
+differential equation known as the
[…]
