### 2025-04-02 fd2f88c
Development (#173)
* Minor updates to applications/scripts/README.md

* Update CONTRIBUTORS.md

* Updates to compile with OpenFOAM-v2012 to OpenFOAM-v2306 (#60)

* COMP: compile with OpenFOAM-v2306 and other recent versions

STYLE: introduce OPENFOAM_NOT_EXTEND, OPENFOAM_COM, OPENFOAM_ORG defs

- allows clearer distinction between flavours

ENH: detect openfoam.com based on -DOPENFOAM= flag and/or vYYMM version

- easier testing against development branches

* STYLE: update OpenFOAM 'flavour' defines

- align more with swak4foam, improve readability etc
  * OPENFOAM_COM         (replaces: OPEN
[…]
