+++ applications/utilities/perturbMeshPoints/perturbMeshPoints.C
@@ -19,17 +19,30 @@ Application
     perturbMeshPoints
 
 Description
-    Add a random perturbation to each mesh points. Boundary points are not
-    changed except for empty patches.
+    Add a random perturbation to each mesh points.The perturbation of a point
+    is calculated as a scale factor times the local minimum edge length.
 
-    This utility is useful for creating distorteg grids for testing
-    discretisations.
+    By default, boundary points are slide along the patch. Patches which should
+    not move can be defined via the fixedPatches entry.
 
-    The inputs are defined in $FOAM_CASE/systemm/perturbMeshPointsDict, and
+    In OpenFOAM.com, points on feature edges are not moved, where feature edges
+    are defined by the minimum cosine angle (minCos). This feature is not
+    currently implemented with OpenFOAM.org and foam-extend.
+
+    The inputs are defined in $FOAM_CASE/system/perturbMeshPointsDict, and
     consist of a seed (for the random number generator) and a scaling factor
     to scale the perturbations. The scaling factor is a vector to allow
     different scalings in different directions; for example, for 2-D, the Z
     component should be set to 0.0.
+
+    If there are negative or small cell volumes after moving the points, the
+    local motion is reduced by the factor beta (defaults to 0.8) and the motion
+    is performed again. The maximum number of corrections iterations is set with
+    maxIter (defaults to 1000). A small volume is defined as the factor
+    minCellVol (defaults to 0.1) times the original cel volume.
+
+    This utility is useful for creating distorted grids for testing
+    discretisations.
 
 Author
     Philip Cardiff, UCD.
@@ -40,10 +53,191 @@ Author
 #include "argList.H"
 #include "Random.H"
 #include "twoDPointCorrector.H"
+#include "unitConversion.H"
+#ifdef OPENFOAM_NOT_EXTEND
+    #include "primitiveMeshTools.H"
+#endif
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 using namespace Foam;
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+#ifdef OPENFOAM_COM
+
+// Copied from Foam::polyDualMesh in OpenFOAM-v2312
+void calcFeatures
+(
+    const polyMesh& mesh,
+    const scalar featureCos,
+    labelList& featureEdges,
+    labelList& featurePoints
+)
+{
+    // Create big primitivePatch for all outside.
+    primitivePatch allBoundary
+    (
+        SubList<face>
+        (
+            mesh.faces(),
+            mesh.nBoundaryFaces(),
+            mesh.nInternalFaces()
+        ),
+        mesh.points()
+    );
+
+    // For ease of use store patch number per face in allBoundary.
+    labelList allRegion(allBoundary.size());
+
+    const polyBoundaryMesh& patches = mesh.boundaryMesh();
+
+    forAll(patches, patchi)
+    {
+        const polyPatch& pp = patches[patchi];
+
+        forAll(pp, i)
+        {
+            allRegion[i + pp.start() - mesh.nInternalFaces()] = patchi;
+        }
+    }
+
+
+    // Calculate patch/feature edges
+    // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    const labelListList& edgeFaces = allBoundary.edgeFaces();
+    const vectorField& faceNormals = allBoundary.faceNormals();
+    const labelList& meshPoints = allBoundary.meshPoints();
+
+    bitSet isFeatureEdge(edgeFaces.size(), false);
+
+    forAll(edgeFaces, edgeI)
+    {
+        const labelList& eFaces = edgeFaces[edgeI];
+
+        if (eFaces.size() != 2)
+        {
[…]
