+++ src/solids4FoamModels/solidModels/fvPatchFields/normalDisplacementZeroShear/normalDisplacementZeroShearFvPatchVectorField.C
@@ -1,0 +1,271 @@
+/*---------------------------------------------------------------------------*\
+License
+    This file is part of solids4foam.
+
+    solids4foam is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or (at your
+    option) any later version.
+
+    solids4foam is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with solids4foam.  If not, see <http://www.gnu.org/licenses/>.
+
+\*---------------------------------------------------------------------------*/
+
+#include "normalDisplacementZeroShearFvPatchVectorField.H"
+#include "addToRunTimeSelectionTable.H"
+#include "transformField.H"
+#include "volFields.H"
+#include "lookupSolidModel.H"
+
+// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
+
+namespace Foam
+{
+
+// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
+
+normalDisplacementZeroShearFvPatchVectorField::
+normalDisplacementZeroShearFvPatchVectorField
+(
+    const fvPatch& p,
+    const DimensionedField<vector, volMesh>& iF
+)
+:
+    solidDirectionMixedFvPatchVectorField(p, iF),
+    normalDisp_(p.size(), 0.0),
+    dispSeries_(),
+    forceZeroShearGrad_(false)
+{}
+
+
+normalDisplacementZeroShearFvPatchVectorField::
+normalDisplacementZeroShearFvPatchVectorField
+(
+    const normalDisplacementZeroShearFvPatchVectorField& pvf,
+    const fvPatch& p,
+    const DimensionedField<vector, volMesh>& iF,
+    const fvPatchFieldMapper& mapper
+)
+:
+    solidDirectionMixedFvPatchVectorField(pvf, p, iF, mapper),
+#ifdef OPENFOAM_ORG
+    normalDisp_(mapper(pvf.normalDisp_)),
+#else
+    normalDisp_(pvf.normalDisp_, mapper),
+#endif
+    dispSeries_(pvf.dispSeries_),
+    forceZeroShearGrad_(pvf.forceZeroShearGrad_)
+{}
+
+
+normalDisplacementZeroShearFvPatchVectorField::
+normalDisplacementZeroShearFvPatchVectorField
+(
+    const fvPatch& p,
+    const DimensionedField<vector, volMesh>& iF,
+    const dictionary& dict
+)
+:
+    solidDirectionMixedFvPatchVectorField(p, iF),
+    normalDisp_(p.size(), 0.0),
+    dispSeries_(),
+    forceZeroShearGrad_
+    (
+        dict.lookupOrDefault<Switch>("forceZeroShearGrad", false)
+    )
+{
+    // Check if displacement is time-varying
+    if (dict.found("displacementSeries"))
+    {
+        Info<< "    displacement is time-varying" << endl;
+        dispSeries_ =
+            interpolationTable<scalar>(dict.subDict("displacementSeries"));
+
+        refValue() =
+            patch().nf()*dispSeries_(this->db().time().timeOutputValue());
+    }
+    else
+    {
+        normalDisp_ = scalarField("normalDisplacement", dict, p.size());
+
+        refValue() = patch().nf()*normalDisp_;
+    }
+
+    this->refGrad() = vector::zero;
+
[…]
