+++ src/solids4FoamModels/materialModels/mechanicalModel/mechanicalLaws/mechanicalLaw/mechanicalLaw.C
@@ -338,7 +338,11 @@ void Foam::mechanicalLaw::makeF()
                 mesh().time().timeName(),
                 mesh(),
                 IOobject::READ_IF_PRESENT,
-                IOobject::NO_WRITE
+                // ZT: Must write in order to allow restart
+                // for incremental TL solvers.
+                // Required if correct(sigma) is used in the solver
+                IOobject::AUTO_WRITE
+                // IOobject::NO_WRITE
             ),
             mesh(),
             dimensionedTensor("I", dimless, I)
