### 2025-04-02 #7407 @kyleniemeyer
@philipcardiff it looks like a reference/ref ID may be duplicated in the `bib` file, can you fix that?
