### 2025-04-02 33b09c0
paper.bib: remove duplicate reference
