+++ paper/paper.bib
@@ -47,16 +47,6 @@ Year = {1989--2017}
   doi       = {10.1016/j.camwa.2015.12.025}
 }
 
-@article{Tang2015,
-  author    = {Tang, T. and Hededal, O. and Cardiff, P.},
-  title     = {On finite volume method implementation of poro-elasto-plasticity soil model},
-  journal   = {International Journal for Numerical and Analytical Methods in Geomechanics},
-  volume    = {39},
-  pages     = {1410–1430},
-  year      = {2015},
-  doi       = {10.1002/nag.2361}
-}
-
 @misc{Haider2019,
   author    = {Haider, Jibran},
   title     = {{ExplicitSolidDynamics toolkit for OpenFOAM}},
