### 2025-04-02 #7407 @philipcardiff
> [@philipcardiff](https://github.com/philipcardiff) it looks like a reference/ref ID may be duplicated in the `bib` file, can you fix that?

Done: https://github.com/solids4foam/solids4foam/commit/33b09c03a71f2fa36e3d96af3be9a105249ccead
